### 2017-06-03 b0cfd5e
Create Kameleon object in separate function and pass it as pointer to tracing function
