### 2017-06-03 18c8679
Add function to save a downsampled fieldline state to binary file
