### 2017-06-03 d98ec4d
Fix bug that caused a crash when no state was active and user changed what variable to color fieldlines by from the GUI.
