### 2017-06-03 cddec57
Make search of all files of certain extension in location non recursive
