### 2017-06-04 325813c
Protect concurrent job manager with mutex
