### 2017-06-04 b6cf747
Comment out experimental BMP decoding
