### 2017-06-04 cfcb74f
Add skeleton for h265 decoder
