+++ modules/solarbrowsing/util/h265decoder.h
@@ -1,0 +1,47 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#ifndef __OPENSPACE_MODULE_SOLARBROWSING___H265DECODER___H__
+#define __OPENSPACE_MODULE_SOLARBROWSING___H265DECODER___H__
+
+#include <modules/solarbrowsing/util/decodejob.h>
+
+namespace openspace {
+
+class H265Decoder {
+public:
+    H265Decoder(const unsigned int bufferSize, const std::string& path);
+    ~H265Decoder();
+
+    bool hasMoreFrames();
+    void setBufferSize(const unsigned int size);
+    void popFrame();
+private:
+    void initialize();
+    unsigned int _bufferSize = 0;
+};
+
+} // namespace openspace
+
+#endif // __OPENSPACE_MODULE_SOLARBROWSING___H265DECODER___H__
