### 2017-06-04 4d0a682
Make compliant to updated master branch
