+++ modules/fieldlinessequence/rendering/renderablefieldlinessequence.h
@@ -27,13 +27,18 @@
 
 #include <openspace/rendering/renderable.h>
 #include <openspace/properties/propertyowner.h>
-#include <openspace/properties/scalarproperty.h>
+#include <openspace/properties/optionproperty.h>
+#include <openspace/properties/scalar/boolproperty.h>
+#include <openspace/properties/scalar/floatproperty.h>
+#include <openspace/properties/scalar/intproperty.h>
 #include <openspace/properties/stringproperty.h>
 #include <openspace/properties/vector/vec2property.h>
 #include <openspace/properties/vector/vec4property.h>
 #include <openspace/rendering/transferfunction.h>
 
 #include <modules/fieldlinessequence/util/fieldlinesstate.h>
+#include <string>
+#include <vector>
 
 namespace ghoul {
     namespace opengl {
