+++ modules/solarbrowsing/ext/libde265/libde265/image.cc
@@ -1,0 +1,806 @@
+/*
+ * H.265 video codec.
+ * Copyright (c) 2013-2014 struktur AG, Dirk Farin <farin@struktur.de>
+ *
+ * This file is part of libde265.
+ *
+ * libde265 is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * libde265 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with libde265.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "image.h"
+#include "decctx.h"
+#include "encoder/encoder-context.h"
+
+#include <stdlib.h>
+#include <string.h>
+#include <assert.h>
+
+#include <limits>
+
+
+#ifdef HAVE_MALLOC_H
+#include <malloc.h>
+#endif
+
+#ifdef HAVE_SSE4_1
+#define MEMORY_PADDING  8
+#else
+#define MEMORY_PADDING  0
+#endif
+
+#define STANDARD_ALIGNMENT 16
+
+#ifdef HAVE___MINGW_ALIGNED_MALLOC
+#define ALLOC_ALIGNED(alignment, size)         __mingw_aligned_malloc((size), (alignment))
+#define FREE_ALIGNED(mem)                      __mingw_aligned_free((mem))
+#elif _WIN32
+#define ALLOC_ALIGNED(alignment, size)         _aligned_malloc((size), (alignment))
+#define FREE_ALIGNED(mem)                      _aligned_free((mem))
+#elif defined(HAVE_POSIX_MEMALIGN)
+static inline void *ALLOC_ALIGNED(size_t alignment, size_t size) {
+    void *mem = NULL;
+    if (posix_memalign(&mem, alignment, size) != 0) {
+        return NULL;
+    }
+    return mem;
+};
+#define FREE_ALIGNED(mem)                      free((mem))
+#else
+#define ALLOC_ALIGNED(alignment, size)      malloc(size)//memalign((alignment), (size))
+#define FREE_ALIGNED(mem)                   free((mem))
+#endif
+
+#define ALLOC_ALIGNED_16(size)              ALLOC_ALIGNED(16, size)
+
+static const int alignment = 16;
+
+LIBDE265_API void* de265_alloc_image_plane(struct de265_image* img, int cIdx,
+                                           void* inputdata, int inputstride, void *userdata)
+{
+  int alignment = STANDARD_ALIGNMENT;
+  int stride = (img->get_width(cIdx) + alignment-1) / alignment * alignment;
+  int height = img->get_height(cIdx);
+
+  uint8_t* p = (uint8_t *)ALLOC_ALIGNED_16(stride * height + MEMORY_PADDING);
+
+  if (p==NULL) { return NULL; }
+
+  img->set_image_plane(cIdx, p, stride, userdata);
+
+  // copy input data if provided
+
+  if (inputdata != NULL) {
+    if (inputstride == stride) {
+      memcpy(p, inputdata, stride*height);
+    }
+    else {
+      for (int y=0;y<height;y++) {
+        memcpy(p+y*stride, ((char*)inputdata) + inputstride*y, inputstride);
+      }
+    }
+  }
+
+  return p;
+}
+
+
[…]
