### 2017-06-04 3904ca3
Implement basic first version of video decoder logic
