### 2017-06-05 3716003
Add experimental solar video class
