+++ modules/solarbrowsing/rendering/renderablesolarvideo.h
@@ -1,0 +1,135 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#ifndef __OPENSPACE_MODULE_BASE___RENDERABLESOLARVIDEO___H__
+#define __OPENSPACE_MODULE_BASE___RENDERABLESOLARVIDEO___H__
+
+// TODO(mnoven): A-Z
+#include <modules/base/rendering/renderableplane.h>
+#include <openspace/properties/scalar/doubleproperty.h>
+#include <openspace/properties/scalar/intproperty.h>
+#include <openspace/properties/scalar/uintproperty.h>
+#include <openspace/properties/optionproperty.h>
+#include <openspace/properties/scalar/boolproperty.h>
+#include <openspace/properties/stringproperty.h>
+#include <openspace/engine/downloadmanager.h> // Make pointer & forward declare?
+#include <modules/solarbrowsing/util/spacecraftimagerymanager.h>
+#include <openspace/rendering/transferfunction.h>
+#include <memory>
+#include <modules/solarbrowsing/util/simplej2kcodec.h>
+#include <unordered_set>
+
+#include <modules/solarbrowsing/rendering/spacecraftcameraplane.h>
+
+#include <modules/solarbrowsing/util/streambuffer.h>
+#include <modules/solarbrowsing/util/pixelbufferobject.h>
+#include <modules/solarbrowsing/util/decodejob.h>
+
+namespace ghoul { namespace opengl { class Texture; }}
+
+namespace openspace {
+
+class H265Decoder;
+
+class RenderableSolarVideo : public Renderable {
+
+public:
+    RenderableSolarVideo(const ghoul::Dictionary& dictionary);
+
+    void render(const RenderData& data);
[…]
