+++ modules/solarbrowsing/rendering/renderablesolarvideo.cpp
@@ -1,0 +1,448 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#include <modules/solarbrowsing/rendering/renderablesolarvideo.h>
+
+#include <ghoul/logging/logmanager.h>
+#include <openspace/engine/openspaceengine.h>
+#include <openspace/scene/scenegraphnode.h>
+#include <openspace/rendering/renderengine.h>
+
+// TODO(mnoven) CLEAN REDUNDANT STUFF
+#include <ghoul/filesystem/filesystem>
+#include <ghoul/io/texture/texturereader.h>
+#include <ghoul/opengl/programobject.h>
+#include <ghoul/opengl/texture.h>
+#include <ghoul/opengl/textureunit.h>
+
+#include <ghoul/filesystem/filesystem.h>
+#include <modules/solarbrowsing/util/h265decoder.h>
+
+#include <openspace/util/time.h>
+#include <openspace/util/timemanager.h>
+#include <openspace/scene/scene.h>
+
+#include <modules/solarbrowsing/util/videoplayer.h>
+
+#include <chrono>
+#include <math.h>
+
+using namespace ghoul::opengl;
+using namespace std::chrono;
+
+typedef std::chrono::high_resolution_clock Clock;
+
+namespace {
+    static const std::string _loggerCat = "RenderableSolarVideo";
+    const double EPSILON = std::numeric_limits<double>::epsilon();
+}
+
+namespace openspace {
+
+RenderableSolarVideo::RenderableSolarVideo(const ghoul::Dictionary& dictionary)
+    : Renderable(dictionary)
[…]
