### 2017-06-05 5fd0d94
Remove some old code from spacecraftplane
