### 2017-06-05 34c736d
Don't pass solarimagedata object to projection component
