### 2017-06-05 776ce17
Begin work on video player class
