+++ modules/solarbrowsing/util/h265decoder.h
@@ -30,14 +30,25 @@
 #include "de265.h"
 #include <stdio.h>
 
-#define BUFFER_SIZE 40000
+#include <chrono>
+
+typedef std::chrono::high_resolution_clock Clock;
+
+#define CHUNK_SIZE 500000
 
 namespace openspace {
 
+// class VideoJob : public StreamJob<SolarImageData> {
+//     virtual void execute() final {
+//     }
+//     virtual std::shared_ptr<SolarImageData> product() final {
+//         return std::move(_solarImageData);
+//     }
+// }
+
 class H265Decoder {
 public:
-    H265Decoder(const std::string& path)
-    {
+    H265Decoder(const std::string& path) {
       initialize(path);
     };
 
@@ -48,32 +59,69 @@ public:
 
     bool hasMoreFrames() {
         return _isRunning;
-    };
-    void setBufferSize(const unsigned int size);
-    void popFrame() {
+    }
+
+    const uint8_t* popFrame() {
+        //auto t1 = Clock::now();
+        //bool hasMoreWork = decode();
+        //auto t2 = Clock::now();
+        //std:: cerr << "Bottleneck function ? "
+        //<< std::chrono::duration_cast<std::chrono::milliseconds>(t2 - t1).count()
+        //<< " ms" << std::endl;
         const de265_image* img = de265_get_next_picture(_ctx);
         if (img) {
-            std::cout << "And we got an image!!!";
+            int width  = de265_get_image_width(img,0);
+            int height = de265_get_image_height(img,0);
+            //std::cerr << " We got an image!!!" << width << " " << height;
+            int stride; // bytes per row
+            const uint8_t* yptr = de265_get_image_plane(img, 0, &stride);
+            // Empty buffer, push more data
+            return yptr;
         }
-        // Do some decoding
-        bool hasMoreWork = decode();
-        // Empty buffer, push more data
-        if (!hasMoreWork && _isRunning) {
-            giveDataToDecoder();
+        return nullptr;
+    }
+
+    bool decode() {
+        int more = 0;
+        if (_isRunning) {
+            de265_error err = de265_decode(_ctx, &more);
+            // ERROR
+            if (err != DE265_OK) {
+                if (err == DE265_ERROR_CHECKSUM_MISMATCH) {
+                    std::cerr << "Some error mismatch";
+                    _isRunning = false;
+                    more = 0;
+                }
+                // Must collect some more
+                else if (err == DE265_ERROR_IMAGE_BUFFER_FULL) {
+                    std::cerr << "Warning, buffer is full needs collect";
+                    more = 1;
+                }
+
+                else if (err == DE265_ERROR_WAITING_FOR_INPUT_DATA) {
+                    std::cerr << "Warning, decoder is empty"; 
+                    more = 0;
+                }
+            }
+            // if (!more && _isRunning) {
+            //     giveDataToDecoder();
+            // }
         }
-        return img;
-    };
-private:
+        return more;
+    }
+
     void giveDataToDecoder() {
         if (!_isRunning) {
-            return;
+            std::cerr << "Is not running " << std::endl;
+             return;
         }
+        std::cerr << "Pushing some data to decoder " << std::endl;
 
         de265_error err;
-        uint8_t buf[BUFFER_SIZE];
-        int n = fread(buf, 1, BUFFER_SIZE, _fh);
+        uint8_t buf[CHUNK_SIZE];
+        int n = fread(buf, 1, CHUNK_SIZE, _fh);
         if (n) {
-            err = de265_push_data(_ctx, buf, n, 0   , (void*)2);
+            err = de265_push_data(_ctx, buf, n, 0, (void*)2);
             if (err != DE265_OK) {
[…]
