### 2017-06-05 c85f404
[WebGUI] Ensure external process helper executable works as intended
This adds the manifest to the executable and tells CEF to use the
executable instead of OpenSpace.exe. Also, limit Windows version support
to ensure required Windows APIs may be used.
