### 2017-06-05 #281 @whedon
[REVIEW]: OpenSpace: An open-source astrovisualization framework
**Submitting author:** @alexanderbock (<a href="http://orcid.org/0000-0002-2849-6146">Alexander Bock</a>)
**Repository:** <a href="https://github.com/OpenSpace/OpenSpace" target ="_blank">https://github.com/OpenSpace/OpenSpace</a>
**Version:** v0.8.0
