### 2017-06-05 #281 @danielskatz
@MatthewTurk - this is ready for you to review.  You should review the reviewer instructions on http://joss.theoj.org/about. Let me know if you have any problems.
