### 2017-06-05 75f2c8e
[WebGUI] Move CEF logic to new CefHost class
