+++ modules/webgui/include/cef_host.h
@@ -22,21 +22,50 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
  ****************************************************************************************/
 
-#include "webguimodule.h"
+#ifndef OPENSPACE_MODULE_WEBGUI__CEF_HOST_H
+#define OPENSPACE_MODULE_WEBGUI__CEF_HOST_H
 
-namespace {
-    std::string _loggerCat = "WebGUI";
-}
+#include <ghoul/filesystem/filesystem.h>
+#include <include/wrapper/cef_helpers.h>
+#include "include/browser_client.h"
+#include "include/gui_render_handler.h"
+#include "include/event_handler.h"
 
 namespace openspace {
 
-WebGUIModule::WebGUIModule()
-    : OpenSpaceModule("WebGUI") {
-    cefHost = std::make_unique<CefHost>();
-}
+#ifdef __APPLE__
+static const std::string SUBPROCESS_PATH = "./webgui_cef_helper";
+#endif
+#ifdef WIN32
+static const std::string SUBPROCESS_PATH = "C:\\Users\\Klas\\Development\\OpenSpace\\bin\\openspace\\Debug\\webgui_cef_helper.exe";
+#endif
 
-void WebGUIModule::internalInitialize() {
-    cefHost->internalInitialize();
-}
+class CefHost {
+public:
+    CefHost();
+    ~CefHost();
 
-}
+    void load(const std::string&);
+    bool loadLocalPath(std::string);
+    void internalInitialize();
+
+protected:
+    void initialize();
+    void initializeCallbacks();
+private:
+    void render();
+    void deinitialize();
+    void attachDebugSettings(CefSettings&);
+
+    std::shared_ptr<EventHandler> eventHandler;
+
+    CefRefPtr<GUIRenderHandler> renderHandler;
+    CefRefPtr<BrowserClient> client;
+    CefRefPtr<CefBrowser> browser;
+
+    void reshape(WindowWrapper&);
+};
+
+} // namespace openspace
+
+#endif //OPENSPACE_MODULE_WEBGUI__CEF_HOST_H
