+++ modules/webgui/src/cef_host.cpp
@@ -1,0 +1,152 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#include "include/cef_host.h"
+
+namespace {
+std::string _loggerCat = "WebGUI";
+}
+
+namespace openspace {
+
+CefHost::CefHost() {
+    LDEBUG("Initializing CEF...");
+    CefMainArgs args;
+    CefSettings settings;
+
+    CefString(&settings.browser_subprocess_path).FromASCII((char*) SUBPROCESS_PATH.c_str());
+    attachDebugSettings(settings);
+
+    CefInitialize(args, settings, nullptr, NULL);
+    LDEBUG("Initializing CEF... done!");
+
+    renderHandler = new GUIRenderHandler();
+    client = new BrowserClient(renderHandler);
+
+    CefBrowserSettings browserSettings;
+    CefWindowInfo windowInfo;
+    bool renderTransparent = true;
+    windowInfo.SetAsWindowless(nullptr, renderTransparent);
+    std::string url = "";
+    browser = CefBrowserHost::CreateBrowserSync(windowInfo, client.get(), url, browserSettings, NULL);
+    eventHandler = std::make_shared<EventHandler>(EventHandler(browser));
+}
+
+CefHost::~CefHost() {
+    // TODO(klas): This is causing a crash during shutdown
+    browser->GetHost()->CloseBrowser(true);
+
+    CefShutdown();
+}
+
+void CefHost::internalInitialize() {
+    OsEng.registerModuleCallback(
+            OpenSpaceEngine::CallbackOption::Initialize,
+            [this](){
+                LDEBUG("Initializing CefHost members");
+                initialize();
+                renderHandler->initialize();
+                eventHandler->initialize();
+                initializeCallbacks();
+            }
+    );
+    OsEng.registerModuleCallback(
[…]
