### 2017-06-05 c28b782
Improved HDR and initial implementation of the Ozone Layer (still needs some work).
