### 2017-06-05 5623f18
Add support to read from files with new json format
