+++ modules/fieldlinessequence/util/fieldlinessequencemanager.cpp
@@ -408,8 +408,6 @@ bool FieldlinesSequenceManager::getFieldlinesState(
     json jfile;// = json::parse(ifs);
     ifs >> jfile;
 
-    size_t numLines = jfile.size();
-
     // if the JSON file contains the key "0. meta" it is in fieldline state format
     if (jfile.find("0. meta") != jfile.end()) {
         // check my format
@@ -455,8 +453,133 @@ bool FieldlinesSequenceManager::getFieldlinesState(
             outFieldlinesState._extraVariables.push_back(it.value());
         }
 
-    } else {
-        // in the json file I have gotten from CCMC, fieldline data is structured like this:
+    } else if (jfile.find("0") != jfile.end()) {
+        // File is in a format provided by CCMC.. currently 2 different available
+
+        // tmp json variable to determina which CCMC structure it is.
+        json jsonTmp = jfile["0"];
+        const std::string strTrace = "trace";
+        if (jsonTmp.find(strTrace) != jsonTmp.end()) {
+            // New json format provided for SUN-EARTH event at AMNH June 27th 2017.. for BATSRUS magnetosphere
+            // is structured something like this:
+            //
+            // {
+            //     "0":{
+            //         "ux": -52.186,
+            //         "uy": 25.266,
+            //         "uz": 14.403,
+            //         "transition": 0.118,
+            //         "x": 13.200,
+            //         "y": 1.4780,
+            //         "z": 0.0,
+            //         "resolution": 0.5,
+            //         "topology": "solar_wind",
+            //         "trace": {
+            //             "index": [0,1,...2081,2082],
+            //             "columns": ["x","y","z","s","temperature","rho","j_para"],
+            //             "data": [[8.694,127.853,115.304,0.0,0.047,9.249,-5e-10],...,[8.698,127.253,114.768,0.800,0.0,9.244,-5e-10]]
+            //         },
+            //     },
+            //     "1":{
+            //         "ux": -52.1869049072,
+            //         "uy": 25.2660446167,
+            //         "uz": 14.4033794403,
+            //         "transition": 0.1182744259,
+            //         "x": 13.2007786501,
+            //         "y": 1.4780422477,
+            //         "z": 0.0,
+            //         "resolution": 0.5,
+            //         "topology": "solar_wind",
+            //         "trace": {
+            //             "index": [0,1,...2081,2082],
+            //             "columns": ["x","y","z","s","temperature","rho","j_para"],
+            //             "data": [[8.698,127.845,115.306,0.0,0.027,9.249,-5e-10],...,[8.698,127.253,114.768,0.8,0.004,9.244,-5e-10]]
+            //         },
+            //     }
+            // }
+
+            // Determine how many variables there are in each line. (Assumes each line is structured the same)
+            const std::string strColumns = "columns";
+            std::vector<std::string> variableNames = jsonTmp[strTrace][strColumns];
+            //if (jsonTmp.find(strTrace) == jsonTmp.end()) {
+            //}
+
+            // TODO: Don't hardcode this!!
+            const float scalingFactor = R_E_TO_METER;
+
+            const size_t numDataVariables = variableNames.size();
+            if (numDataVariables < 3) {
+                LERROR("Json file: " << pathToJsonFile << " doesn't contain necessary fields." <<
+                        "Expects a field \"columns\" which contains at least 3 variables representing position");
+                return false;
+            }
+            auto xIter = std::find(variableNames.begin(), variableNames.end(), "x");
+            auto yIter = std::find(variableNames.begin(), variableNames.end(), "y");
[…]
