### 2017-06-06 caf1336
Implementation for dynamic mastcam coloring of models
