+++ modules/globebrowsing/models/renderableroversurface.cpp
@@ -32,6 +32,7 @@
 #include <openspace/scene/scene.h>
 
 #include <ghoul/io/texture/texturereader.h>
+#include <glm/gtx/quaternion.hpp>
 
 #include <fstream>
 #include <gdal_priv.h>
@@ -58,12 +59,15 @@ namespace globebrowsing {
 		, _generalProperties({
 				BoolProperty("enable", "Enabled", true),
 				BoolProperty("enablePath", "Enable path", true),
-				BoolProperty("lockSubsite", "Lock subsite", false)
+				BoolProperty("lockSubsite", "Lock subsite", false),
+				BoolProperty("useMastCam", "Show mastcam coloring", false)
 		})
 		, _debugModelRotation("modelrotation", "Model Rotation", glm::vec3(0.0f), glm::vec3(0.0f), glm::vec3(360.0f))
 		, _modelSwitch()
 		, _prevLevel(3)
 		, _isFirst(true)
+		, _isFirstLow(true)
+		, _isFirstHigh(true)
 {
 	if (!dictionary.getValue(keyRoverLocationPath, _roverLocationPath))
 		throw ghoul::RuntimeError(std::string(keyRoverLocationPath) + " must be specified!");
@@ -96,6 +100,7 @@ namespace globebrowsing {
 
 	addProperty(_generalProperties.enablePath);
 	addProperty(_generalProperties.lockSubsite);
+	addProperty(_generalProperties.useMastCam);
 	addProperty(_debugModelRotation);
 
 	_cachingModelProvider = std::make_shared<CachingSurfaceModelProvider>(this);
@@ -148,9 +153,14 @@ void RenderableRoverSurface::render(const RenderData& data) {
 	ghoul::Dictionary modelDic;
 	std::unique_ptr<ModelProvider>_modelProvider;
 	int level;
+	
 	switch (_modelSwitch.getLevel(data)) {
 		case LodModelSwitch::Mode::Low :	
 			//Low
+			if (_isFirstLow) LERROR("GOIING LOW");
+			_isFirstLow = false;
+			_isFirstHigh = true;
+			_isFirst = true;
 			modelDic.setValue("Type", "MultiModelProvider");
 			_modelProvider = std::move(ModelProvider::createFromDictionary(modelDic));
 			ss = _modelProvider->calculate(subSitesVector, data);
@@ -161,6 +171,9 @@ void RenderableRoverSurface::render(const RenderData& data) {
 			//Close
 			if (_isFirst) LERROR("GOING CLOSE");
 			_isFirst = false;
+			_isFirstLow = true;
+			_isFirstHigh = true;
+			
 			modelDic.setValue("Type", "SingleModelProvider");
 			_modelProvider = std::move(ModelProvider::createFromDictionary(modelDic));
 			ss = _modelProvider->calculate(subSitesVector, data);
@@ -168,6 +181,10 @@ void RenderableRoverSurface::render(const RenderData& data) {
 			break;
 
 		case LodModelSwitch::Mode::High :
+			if (_isFirstHigh) LERROR("GOING HIGH");
+			_isFirstHigh = false;
+			_isFirstLow = true;
+			_isFirst = true;
 			//High up
 			level = 1;
 			break;
@@ -342,20 +359,72 @@ void RenderableRoverSurface::render(const RenderData& data) {
 
 		_programObject->setUniform("modelViewTransform", glm::mat4(modelViewTransform));
 		_programObject->setUniform("projectionTransform", data.camera.projectionMatrix());
[…]
