### 2017-06-06 7426f23
Reading of mastcam textures
