### 2017-06-06 a470d79
Added fading and code cleanup
