+++ modules/globebrowsing/models/cachingsurfacemodelprovider.h
@@ -90,8 +90,6 @@ public:
 	std::vector<std::shared_ptr<SubsiteModels>> getModels(const std::vector<std::shared_ptr<Subsite>> Subsites, const int level);
 
 	void update(Renderable* parent);
-	
-	void setLevel(const int level);
 
 private:
 	std::shared_ptr<AsyncSurfaceModelProvider> _asyncSurfaceModelProvider;
@@ -99,17 +97,10 @@ private:
 	std::vector<int> getLevelsAbove(const std::vector<int> availableLevels, const int requestedLevel);
 
 	void initModelsFromLoadedData(Renderable* parent);
-	void clearQueuesAndJobs();
-
-	uint64_t hashKey(const std::string site, const std::string drive, const int level);
-	
-	int _prevLevel = -1;
 
 	Renderable* _parent;
 
 	Vertex* _vertexBufferData = nullptr;
-
-	int _previousLevel = 0;
 };
 
 } // namespace globebrowsing
