+++ modules/globebrowsing/models/cachingsurfacemodelprovider.cpp
@@ -42,21 +42,21 @@ CachingSurfaceModelProvider::CachingSurfaceModelProvider(Renderable* parent)
 	_modelCache = std::make_shared<ModelCache>(static_cast<size_t>(cacheSize));
 }
 
-std::vector<std::shared_ptr<SubsiteModels>> CachingSurfaceModelProvider::getModels(const std::vector<std::shared_ptr<Subsite>> Subsites,
+std::vector<std::shared_ptr<SubsiteModels>> CachingSurfaceModelProvider::getModels(const std::vector<std::shared_ptr<Subsite>> subsites,
 	const int level) {
-
-	// Check if the chunk is already in the cache. If it is in the cache, loop through the corresponding vector
-	// and look if the model is already loaded. If the model is already loaded, return the vector. If the model is
-	// not already loaded, enqueue the model.
-	std::vector<std::shared_ptr<SubsiteModels>> vectorOfSubsiteModels;
+	// Reset the fading direction automatically makes all unwanted models to fade out
+	auto itemListResetDirection = _modelCache->list();
+	for (auto entry : itemListResetDirection) {
+		entry.second->_dir = -1;
+	}
 
 	// Temporary solution to not load LOD 0 models
 	if(level > 0) {
-		for (auto subsite : Subsites) {			
+		for (auto subsite : subsites) {			
 			bool requestedExistsInCache = true;
 			ProviderSubsiteKey key = { level, subsite->site, subsite->drive };
 			if (_modelCache->exist(key)) {
-				vectorOfSubsiteModels.push_back(_modelCache->get(key));
+				_modelCache->get(key)->_dir = 1;
 			}
 			else {
 				for(const int& tempLevel : subsite->availableLevels){
@@ -64,7 +64,6 @@ std::vector<std::shared_ptr<SubsiteModels>> CachingSurfaceModelProvider::getMode
 					if(tempLevel == level)
 						_asyncSurfaceModelProvider->enqueueModelIO(subsite, level);
 				}
-			
 				// Check for available LODs above the requested.
 				std::vector<int> levelsAbove = getLevelsAbove(subsite->availableLevels, level);
 				for (int levelAbove = levelsAbove.size() + 1; levelAbove-- > 1; ) {
@@ -72,7 +71,7 @@ std::vector<std::shared_ptr<SubsiteModels>> CachingSurfaceModelProvider::getMode
 					// If the cache holds the correct model but with lower resultion than requested,
 					// return highest available resolution.
 					if (_modelCache->exist(keyLowerLevel)) {
-						vectorOfSubsiteModels.push_back(_modelCache->get(keyLowerLevel));
+						_modelCache->get(keyLowerLevel)->_dir = 1;
 						// Breaks the loop that goes through levelsAbove
 						break;
 					}
@@ -94,23 +93,22 @@ std::vector<std::shared_ptr<SubsiteModels>> CachingSurfaceModelProvider::getMode
 			}
 		}
 	}
+	
+	std::vector<std::shared_ptr<SubsiteModels>> vectorOfSubsiteModels;
+	// Save all models that are still visible and fade at the same time
+	auto itemListSetDirection = _modelCache->list();
+	for (auto entry : itemListSetDirection) {
+		if (entry.second->alpha > 0.0) {
+			vectorOfSubsiteModels.push_back(entry.second);
+		}
+		entry.second->fade();
+	}
 
 	return vectorOfSubsiteModels;
 }
 
 void CachingSurfaceModelProvider::update(Renderable* parent) {
 	initModelsFromLoadedData(parent);
-}
-
[…]
