+++ modules/globebrowsing/models/subsitemodels.h
@@ -30,7 +30,7 @@
 #include <modules/globebrowsing/geometry/geodetic2.h>
 #include <modules/globebrowsing/models/model.h>
 #include <modules/globebrowsing/tasks/imgreader.h>
-#include <ghoul/opengl/texturearray.h>
+//#include <ghoul/opengl/texturearray.h>
 
 #include <memory>
 
@@ -45,7 +45,7 @@ struct SubsiteModels {
 	std::vector <std::shared_ptr<ghoul::opengl::Texture>> textures;
 	std::vector <std::shared_ptr<ghoul::opengl::Texture>> coloredTextures;
 	std::shared_ptr<openspace::modelgeometry::AsyncMultiModelGeometry> model;
-	std::shared_ptr<ghoul::opengl::TextureArray> textureArray;
+	//std::shared_ptr<ghoul::opengl::TextureArray> textureArray;
 
 	glm::dvec3 cartesianPosition;
 
