### 2017-06-06 55743fc
Reverted change causing bug
