### 2017-06-06 0e9f125
Code cleanup
