+++ modules/globebrowsing/models/subsitemodels.h
@@ -69,16 +69,16 @@ struct SubsiteModels {
 	std::string drive;
 	int level;
 	float alpha = 0.0f;
-	int _dir = -1;
+	int fadeDirection = -1;
 
-	inline bool operator==(const SubsiteModels& other) const {
-		return (site == other.site) && (drive == other.drive) && (level == other.level);
+	void setFadeDirection(const int direction) {
+		fadeDirection = direction;
 	}
 
 	void fade() {
-		if (_dir == 1 && alpha < 1.0f)
+		if (fadeDirection == 1 && alpha < 1.0f)
 			alpha = alpha + 0.01f;
-		else if (_dir == -1 && alpha > 0.0f)
+		else if (fadeDirection == -1 && alpha > 0.0f)
 			alpha = alpha - 0.01f;
 	}
 };
