### 2017-06-06 f8f42f6
Added subsitemodels.cpp and updated CMakeLists, also added functions and private variables
