### 2017-06-06 d0492b2
fix misinterpretation on roll, fixed step size in LMA gradient the projectionmatrix caused, fixed infinite spin bug on LMA failure
