+++ modules/touch/src/TouchInteraction.cpp
@@ -56,11 +56,11 @@ namespace {
 using namespace TUIO;
 namespace openspace {
 
-	TouchInteraction::TouchInteraction()
-		: properties::PropertyOwner("TouchInteraction"),
-		_origin("origin", "Origin", ""),
-		_unitTest("Click to take a unit test", "Take a unit test saving the LM data into file", false),
-		_onlyPan("Toggle Panning Mode", "Toggle pan interaction on direct-manipulation three finger case (FOR FEEDBACK)", false), // temp
+TouchInteraction::TouchInteraction()
+	: properties::PropertyOwner("TouchInteraction"),
+	_origin("origin", "Origin", ""),
+	_unitTest("Click to take a unit test", "Take a unit test saving the LM data into file", false),
+	_onlyPan("Toggle Panning Mode", "Toggle pan interaction on direct-manipulation three finger case (FOR FEEDBACK)", false), // temp
 	_touchActive("TouchEvents", "True if we have a touch event", false, properties::Property::Visibility::Hidden),
 	_reset("Default Values", "Reset all properties to default", false),
 	_maxTapTime("Max Tap Time", "Max tap delay (in ms) for double tap", 300, 10, 1000),
@@ -73,10 +73,11 @@ namespace openspace {
 	_panSpeedThreshold("Activate pan spinning", "Threshold to activate pan spinning in direct-manipulation", 0.0005f, 0.0, 0.01),
 	_spinSensitivity("Sensitivity of Spinning", "Sensitivity of spinning in direct-manipulation", 1.0f, 0, 2),
 	_inputStillThreshold("Input still", "Threshold for interpreting input as still", 0.0005f, 0, 0.001),
+	_centroidStillThreshold("Centroid stationary", "Threshold for stationary centroid", 0.0018f, 0, 0.01), // used to void wrongly interpreted roll interactions
 	_interpretPan("Pan delta distance", "Delta distance between fingers allowed for interpreting pan interaction", 0.015f, 0, 0.1),
 	_slerpTime("Time to slerp", "Time to slerp in seconds to new orientation with new node picking", 1, 0, 5),
 	_guiButton("GUI Button", "GUI button size in pixels.", glm::ivec2(32, 64), glm::ivec2(8, 16), glm::ivec2(128, 256)),
-	_friction("Friction", "Friction for different interactions (orbit, zoom, roll, pan)", glm::vec4(0.01, 0.02, 0.02, 0.02), glm::vec4(0.0), glm::vec4(0.2)), 
+	_friction("Friction", "Friction for different interactions (orbit, zoom, roll, pan)", glm::vec4(0.01, 0.025, 0.02, 0.02), glm::vec4(0.0), glm::vec4(0.2)), 
 	
 	_vel{ glm::dvec2(0.0), 0.0, 0.0, glm::dvec2(0.0) },
 	_sensitivity{glm::dvec2(0.0808181818181818, 0.0454545454545455), 4.0, 2.75, glm::dvec2(0.0808181818181818, 0.0454545454545455) },
@@ -100,6 +101,7 @@ namespace openspace {
 	addProperty(_panSpeedThreshold);
 	addProperty(_spinSensitivity);
 	addProperty(_inputStillThreshold);
+	addProperty(_centroidStillThreshold);
 	addProperty(_interpretPan);
 	addProperty(_slerpTime);
 	addProperty(_guiButton);
@@ -132,7 +134,7 @@ void TouchInteraction::updateStateFromInput(const std::vector<TuioCursor>& list,
 	
 	if (!guiMode(list)) {
[…]
