### 2017-06-06 bbea96f
Final updates to video player before going back to imagery
