### 2017-06-06 5abd1bd
Add functionality to get number of jobs to stream buffer
