### 2017-06-06 262d520
Update manager with metadata
