### 2017-06-06 d043d3f
Work on rotating magnetogram texture coordinate to align with HMI
