### 2017-06-06 681a7ec
Enable culling on image plane
