### 2017-06-06 53264aa
Update buffer logic
