+++ modules/solarbrowsing/rendering/renderablesolarimagery.cpp
@@ -42,8 +42,12 @@
 #include <openspace/util/timemanager.h>
 #include <openspace/scene/scene.h>
 
+#include <openspace/util/spicemanager.h>
+
 #include <chrono>
 #include <math.h>
+
+#define BUFFER_MAX_SIZE 100
 
 using namespace ghoul::opengl;
 using namespace std::chrono;
@@ -68,7 +72,7 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     , _gammaValue("gammaValue", "Gamma", 0.9, 0.1, 10.0)
     , _asyncUploadPBO("asyncUploadPBO", "Upload to PBO Async", true)
     , _activeInstruments("activeInstrument", "Active Instrument", properties::OptionProperty::DisplayType::Radio)
-    , _bufferSize("bufferSize", "Buffer Size", 5, 1, 300)
+    , _bufferSize("bufferSize", "Buffer Size", 5, 1, 150)
     , _displayTimers("displayTimers", "Display Timers", false)
     , _lazyBuffering("lazyBuffering", "Lazy Buffering", true)
     , _minRealTimeUpdateInterval("minRealTimeUpdateInterval", "Min Update Interval", 85, 0, 300)
@@ -89,10 +93,10 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         throw ghoul::RuntimeError("Nodename has to be specified");
     }
 
-    glm::dvec2 magicPlaneOffset;
-    if (dictionary.getValue("MagicOffsetFromCenter", magicPlaneOffset)) {
-        _magicPlaneOffset = magicPlaneOffset;
-    }
+    // glm::dvec2 magicPlaneOffset;
+    // if (dictionary.getValue("MagicOffsetFromCenter", magicPlaneOffset)) {
+    //     _magicPlaneOffset = magicPlaneOffset;
+    // }
 
     float tmpStartResolutionLevel;
     if (!dictionary.getValue("StartResolutionLevel", tmpStartResolutionLevel)) {
@@ -100,9 +104,9 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     }
     _resolutionLevel = static_cast<int>(tmpStartResolutionLevel);
 
-    if (!dictionary.getValue("MagicPlaneFactor", _magicPlaneFactor)) {
-        throw ghoul::RuntimeError("Plane must at the moment have a magic factor");
-    }
+    // if (!dictionary.getValue("MagicPlaneFactor", _magicPlaneFactor)) {
+    //     throw ghoul::RuntimeError("Plane must at the moment have a magic factor");
+    // }
 
     std::string rootPath;
     if (!dictionary.getValue("RootPath", rootPath)) {
@@ -186,27 +190,28 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         // Upload asap
         updateTextureGPU(/*asyncUpload=*/false);
         if (_useBuffering) {
-
             //double oktime = OsEng.timeManager().time().deltaTime();
             //TimeManager& lel = OsEng.timeManager();
-
-            fillBuffer(OsEng.timeManager().time().deltaTime());
+            _streamBuffer.clear();
+            //fillBuffer(OsEng.timeManager().time().deltaTime());
         } /*else {
             uploadImageDataToPBO();
         }*/
     });
 
     _resolutionLevel.onChange([this]() {
+        _updatingCurrentLevelOfResolution = true;
         _pboIsDirty = false;
         _imageSize = _fullResolution / (pow(2, _resolutionLevel));
[…]
