### 2017-06-06 d5f7d7b
point size for the subsites in the path are now based on distance to camera
