+++ modules/volume/rendering/renderabletimevaryingvolume.cpp
@@ -32,6 +32,8 @@
 #include <openspace/rendering/raycastermanager.h>
 #include <openspace/documentation/documentation.h>
 #include <openspace/documentation/verifier.h>
+#include <openspace/util/timemanager.h>
+#include <openspace/util/time.h>
 
 #include <ghoul/glm.h>
 #include <ghoul/opengl/ghoul_gl.h>
@@ -87,6 +89,12 @@ RenderableTimeVaryingVolume::RenderableTimeVaryingVolume(const ghoul::Dictionary
     _lowerValueBound = dictionary.value<float>(KeyLowerValueBound);
     _upperValueBound = dictionary.value<float>(KeyUpperValueBound);
     _gridType = VolumeGridType::Cartesian;
+    _transferFunction = std::make_shared<TransferFunction>(_transferFunctionPath);
+
+    ghoul::Dictionary clipPlanesDictionary;
+    dictionary.getValue(KeyClipPlanes, clipPlanesDictionary);
+    _clipPlanes = std::make_shared<volume::VolumeClipPlanes>(clipPlanesDictionary);
+    _clipPlanes->setName("clipPlanes");
 
     if (dictionary.hasValue<std::string>(KeyGridType)) {
         _gridType = volume::parseGridType(dictionary.value<std::string>(KeyGridType));
@@ -141,7 +149,8 @@ bool RenderableTimeVaryingVolume::initialize() {
         t.texture->uploadTexture();
     }
 
-    
+    _clipPlanes->initialize();
+    _transferFunction->update();
     _raycaster = std::make_unique<volume::BasicVolumeRaycaster>(nullptr, _transferFunction, _clipPlanes);
     _raycaster->initialize();
     OsEng.renderEngine().raycasterManager().attachRaycaster(*_raycaster.get());
@@ -153,6 +162,11 @@ bool RenderableTimeVaryingVolume::initialize() {
         }
     };
     onEnabledChange(onChange);
+
+    addProperty(_stepSize);
+    addProperty(_transferFunctionPath);
+    addProperty(_sourceDirectory);
+    addPropertySubOwner(_clipPlanes.get());
 
     return true;
 }
@@ -170,10 +184,12 @@ void RenderableTimeVaryingVolume::loadTimestepMetadata(const std::string& path) 
     t.baseName = ghoul::filesystem::File(path).baseName();
     t.dimensions = dictionary.value<glm::vec3>(KeyDimensions);
     t.lowerDomainBound = dictionary.value<glm::vec3>(KeyLowerDomainBound);
-    t.upperDomainBound = dictionary.value<glm::vec3>(KeyLowerDomainBound);
+    t.upperDomainBound = dictionary.value<glm::vec3>(KeyUpperDomainBound);
     t.minValue = dictionary.value<float>(KeyMinValue);
     t.maxValue = dictionary.value<float>(KeyMaxValue);
-    t.time = dictionary.value<double>(KeyTime);
+
+    std::string timeString = dictionary.value<std::string>(KeyTime);
+    t.time = Time::convertTime(timeString);
     t.inRam = false;
     t.onGpu = false;
     
@@ -181,14 +197,30 @@ void RenderableTimeVaryingVolume::loadTimestepMetadata(const std::string& path) 
 }
 
 RenderableTimeVaryingVolume::Timestep* RenderableTimeVaryingVolume::currentTimestep() {
-    return &(_volumeTimesteps.begin()->second);
+    double currentTime = OsEng.timeManager().time().j2000Seconds();
+    auto currentTimestepIt = _volumeTimesteps.upper_bound(currentTime);
+    if (currentTimestepIt == _volumeTimesteps.end() && _volumeTimesteps.size() > 0) {
[…]
