+++ modules/globebrowsing/rendering/gpu/gpulayer.cpp
@@ -47,7 +47,7 @@ void GPULayer::setValue(ghoul::opengl::ProgramObject* programObject, const Layer
             gpuChunkTilePile.setValue(programObject, chunkTilePile);
             break;
         case layergroupid::TypeID::SolidColor:
-            gpuColor.setValue(programObject, layer.color.value());
+            gpuColor.setValue(programObject, layer.adjustmentProperties().color.value());
             break;
         default:
             break;
