### 2017-06-07 b89ac85
Protect concurrent job manager with mutex
