### 2017-06-07 b1c3249
Explain calculateTileLevelDifference
