### 2017-06-07 5575f17
Feature/gdalcrashfix (#331)
Saving GDAL dataset values upon initialization of dataset to avoid calling GDAL functions afterwards since it can be done concurrently. Not done for the data reading since some (not all) GDAL dataset drivers support concurrent reading. The WMS driver does not support concurrent reading, which is why we only use one worker thread per AsyncTileDataReader. Perhaps something for the future if GDAL will ensure thread saefty.

Protecting GdalRawTileDataReader with mutex lock when initializing.

Make sure read only functions in RawTileDataReader are const.

Protecting Conc
[…]
