### 2017-06-07 606faf1
Fix typo in stream buffer and some debug outputs in solar imagery
