### 2017-06-07 75b5ef7
Increased the speed of fadeing
