### 2017-06-07 1b6f85c
Refactor CefHost into BrowserInstance
The purpose of CefHost is now to initialize CEF and trigger its work while
rendering. BrowserInstance take care of the rest - browsing, events,
rendering etc.
