### 2017-06-07 ea44e7a
Fix billboard rendering for satellite trails Add necessary scene graph node for satellites into lodglobes/earth Update base files of satellite TLEs
