+++ data/scene/satellites/tle/gps-ops.txt
@@ -1,93 +1,93 @@
 GPS BIIR-2  (PRN 13)    
-1 24876U 97035A   17109.73375644 -.00000012  00000-0  00000-0 0  9992
-2 24876  55.5643 229.4105 0036557 106.5416 253.8548  2.00562036144850
+1 24876U 97035A   17157.59786847  .00000019  00000-0  00000-0 0  9996
+2 24876  55.5513 227.5274 0034589 106.0906 254.3641  2.00561883145810
 GPS BIIR-3  (PRN 11)    
-1 25933U 99055A   17109.65740979  .00000008  00000-0  00000-0 0  9993
-2 25933  51.5943  80.2017 0164482  93.5980 140.6974  2.00559092128473
+1 25933U 99055A   17157.81823278 -.00000081  00000-0  00000-0 0  9990
+2 25933  51.6254  78.0928 0166095  94.4525 354.0300  2.00558378129440
 GPS BIIR-4  (PRN 20)    
-1 26360U 00025A   17109.81822694  .00000034  00000-0  00000-0 0  9999
-2 26360  53.1062 157.3245 0041561  90.8273   1.8392  2.00564454124177
+1 26360U 00025A   17157.82515987  .00000060  00000-0  00000-0 0  9996
+2 26360  53.1170 155.3222 0040993  91.9775 104.3666  2.00561532125131
 GPS BIIR-5  (PRN 28)    
-1 26407U 00040A   17109.72770267 -.00000058  00000-0  00000-0 0  9993
-2 26407  56.6488 347.4270 0199991 270.0939  38.2665  2.00567294122845
+1 26407U 00040A   17157.93926722 -.00000076  00000-0  00000-0 0  9998
+2 26407  56.6372 345.5470 0199158 270.1220 289.7792  2.00567008123812
 GPS BIIR-6  (PRN 14)    
-1 26605U 00071A   17109.91594260 -.00000014  00000-0  00000-0 0  9997
-2 26605  55.1353 227.1576 0091621 248.1584 104.4436  2.00567152120445
+1 26605U 00071A   17157.87543271  .00000026  00000-0  00000-0 0  9993
+2 26605  55.1227 225.2520 0093903 248.3083 174.0281  2.00565489121414
 GPS BIIR-7  (PRN 18)    
-1 26690U 01004A   17109.78566808  .00000030  00000-0  00000-0 0  9997
-2 26690  53.0287 160.2705 0181728 256.2994 113.2448  2.00553627118851
+1 26690U 01004A   17157.89507387  .00000063  00000-0  00000-0 0  9995
+2 26690  53.0389 158.2602 0183394 256.9116 288.9071  2.00570451119814
 GPS BIIR-8  (PRN 16)    
-1 27663U 03005A   17109.59259250 -.00000059  00000-0  00000-0 0  9992
-2 27663  56.7184 347.1682 0092496  23.4683 306.5317  2.00563261104201
+1 27663U 03005A   17157.94951778 -.00000076  00000-0  00000-0 0  9997
+2 27663  56.7061 345.2850 0092504  24.7076 300.8478  2.00560319105174
 GPS BIIR-9  (PRN 21)    
-1 27704U 03010A   17109.45896894  .00000044  00000-0  00000-0 0  9993
-2 27704  53.8252 100.6579 0242330 263.6172 289.5204  2.00560332102998
+1 27704U 03010A   17157.88465515 -.00000050  00000-0  00000-0 0  9993
+2 27704  53.8528  98.6489 0242249 264.5394 334.3334  2.00563816103977
 GPS BIIR-10 (PRN 22)    
[…]
