### 2017-06-07 d22d76f
Add WebRenderHandler superclass
This makes it possible to plug in different kinds of render handlers to
the BrowserInstance class. This in turn should make it possible for us to
create a screen space browser.
