### 2017-06-07 c152e39
restructured the code and clean up
