+++ modules/globebrowsing/models/renderableexplorationpath.cpp
@@ -46,8 +46,6 @@ RenderableExplorationPath::RenderableExplorationPath()
 	, _siteShader(nullptr)
 	, _fading(0.0)
 	, _fading2(0.0)
-	, _vertexBufferID(0)
-	, _vaioID(0)
 	, _isEnabled(properties::BoolProperty("enabled", "enabled", false))
 {}
 
@@ -85,45 +83,26 @@ bool RenderableExplorationPath::initialize(RenderableGlobe* globe, const std::ve
 
 	calculatePathModelCoordinates();
 	if (_allGeodetics.size() == 0) return false;
-
+	
+	// 
 	// Initialize and upload to graphics card
-	glGenVertexArrays(1, &_vaioID);
+	glGenVertexArrays(1, &_vaPathID);
 	ghoul_assert(_vaioID != 0, "Could not generate vertex arrays");
 
-	glGenBuffers(1, &_vertexBufferID);
+	glGenBuffers(1, &_vbPathID);
 	ghoul_assert(_vertexBufferID != 0, "Could not create vertex buffer");
 
-	glBindVertexArray(_vaioID);
-	glBindBuffer(GL_ARRAY_BUFFER, _vertexBufferID);
-	glBufferData(GL_ARRAY_BUFFER,
-		_stationPointsModelCoordinates.size() * sizeof(_stationPointsModelCoordinates[0]),
-		&_stationPointsModelCoordinates[0],
-		GL_STATIC_DRAW);
+	bufferData(_vaPathID, _vbPathID, _stationPointsModelCoordinates);
 
-	glEnableVertexAttribArray(0);
-	glVertexAttribPointer(0, 4, GL_FLOAT, GL_FALSE, sizeof(_stationPointsModelCoordinates[0]), 0);
-
-	glBindVertexArray(0);
-
-	// Staions with models
+	// Stations with models
 	// Initialize and upload to graphics card
-	glGenVertexArrays(1, &_vaioID2);
+	glGenVertexArrays(1, &_vaModelsID);
 	ghoul_assert(_vaioID2 != 0, "Could not generate vertex arrays");
 
-	glGenBuffers(1, &_vertexBufferID2);
+	glGenBuffers(1, &_vbModelsID);
 	ghoul_assert(_vertexBufferID2 != 0, "Could not create vertex buffer");
-
-	glBindVertexArray(_vaioID2);
-	glBindBuffer(GL_ARRAY_BUFFER, _vertexBufferID2);
-	glBufferData(GL_ARRAY_BUFFER,
-		_stationPointsModelCoordinatesWithModel.size() * sizeof(_stationPointsModelCoordinatesWithModel[0]),
-		&_stationPointsModelCoordinatesWithModel[0],
-		GL_STATIC_DRAW);
-
-	glEnableVertexAttribArray(0);
-	glVertexAttribPointer(0, 4, GL_FLOAT, GL_FALSE, sizeof(_stationPointsModelCoordinatesWithModel[0]), 0);
-
-	glBindVertexArray(0);
+	
+	bufferData(_vaModelsID, _vbModelsID, _stationPointsModelCoordinatesWithModel);
 
 	return true;
 }
@@ -139,11 +118,11 @@ bool RenderableExplorationPath::deinitialize() {
 		_siteShader = nullptr;
 	}
 
-	glDeleteVertexArrays(1, &_vaioID);
-	glDeleteBuffers(1, &_vertexBufferID);
+	glDeleteVertexArrays(1, &_vaPathID);
+	glDeleteBuffers(1, &_vbPathID);
 
-	glDeleteVertexArrays(1, &_vaioID2);
[…]
