### 2017-06-07 58c6f17
fix to duplicate refs
