+++ modules/globebrowsing/models/renderableexplorationpath.h
@@ -52,8 +52,7 @@ public:
 	void setLevel(const int level);
 
 private:
-	void calculatePathModelCoordinates();
-	void recalculateCartesianCoordinates();
+	std::vector<glm::vec4> calculateModelCoordinates(std::vector<Geodetic2> geodetics, const int offset);
 	void bufferData(GLuint vaID, GLuint vbID, const std::vector<glm::vec4> coordinates);
 	void reBufferData(GLuint vaID, GLuint vbID, const std::vector<glm::vec4> coordinates);
 
