### 2017-06-07 acf25d8
Move gui render callback lambda to Module
This should not be in the browser instance, the browser instance should be
ignorant for how it gets drawn.
