### 2017-06-07 940b752
Improve BrowserInstance::reshape
