### 2017-06-07 edd2467
Add bunch of minor changes/additions
