+++ modules/fieldlinessequence/util/fieldlinessequencemanager.cpp
@@ -153,7 +153,8 @@ bool FieldlinesSequenceManager::getSeedPointsFromFile(
 bool FieldlinesSequenceManager::getAllFilePathsOfType(
         const std::string& pathToDirectory,
         const std::string& fileExtension,
-        std::vector<std::string>& outFilePaths) {
+        std::vector<std::string>& outFilePaths,
+        bool recursiveSearch /*= false*/) {
 
     std::string absFolderPath;
     absFolderPath = absPath(pathToDirectory);
@@ -176,7 +177,7 @@ bool FieldlinesSequenceManager::getAllFilePathsOfType(
 
     // Get absolute path to
     ghoul::filesystem::Directory dir(absFolderPath, RawPath::Yes);
-    outFilePaths = dir.read(FileSystem::Recursive::No, Sort::Yes);
+    outFilePaths = dir.read(static_cast<ghoul::Boolean>(recursiveSearch), Sort::Yes);
 
     outFilePaths.erase(std::remove_if(
             outFilePaths.begin(), outFilePaths.end(), [extLength, extension](std::string s) {
@@ -332,43 +333,43 @@ bool FieldlinesSequenceManager::getFieldlinesStateFromBinary(
             }
     }
 
-    ifs.read( reinterpret_cast<char*>(&outState._triggerTime), sizeof(double));
+    // Define tmp variables to store meta data in
     FieldlinesState::Model model;
-    ifs.read( reinterpret_cast<char*>(&model), sizeof(int));
-    outState.setModel(model);
-    ifs.read( reinterpret_cast<char*>(&outState._isMorphable), sizeof(bool));
     size_t numLines;
-    ifs.read( reinterpret_cast<char*>(&numLines), sizeof(size_t));
     size_t numPoints;
-    ifs.read( reinterpret_cast<char*>(&numPoints), sizeof(size_t));
     size_t numExtras;
-    ifs.read( reinterpret_cast<char*>(&numExtras), sizeof(size_t));
     size_t byteSizeAllNames;
-    ifs.read( reinterpret_cast<char*>(&byteSizeAllNames), sizeof(size_t));
 
+    // Read single value variables
+    ifs.read( reinterpret_cast<char*>(&outState._triggerTime), sizeof(double));
+    ifs.read( reinterpret_cast<char*>(&model),                 sizeof(int));
+    ifs.read( reinterpret_cast<char*>(&outState._isMorphable), sizeof(bool));
+    ifs.read( reinterpret_cast<char*>(&numLines),              sizeof(size_t));
+    ifs.read( reinterpret_cast<char*>(&numPoints),             sizeof(size_t));
+    ifs.read( reinterpret_cast<char*>(&numExtras),             sizeof(size_t));
+    ifs.read( reinterpret_cast<char*>(&byteSizeAllNames),      sizeof(size_t));
+
+    outState.setModel(model);
+    // RESERVE/RESIZE vectors
     // TODO: figure out how to do this with reserve instead of resize! resize is about 1,4 times slower
-    // outState._lineStart.reserve(numLines);
     outState._lineStart.resize(numLines);
-    ifs.read( reinterpret_cast<char*>(outState._lineStart.data()), sizeof(GLint)*numLines);
-
-    // outState._lineCount.reserve(numLines);
     outState._lineCount.resize(numLines);
-    ifs.read( reinterpret_cast<char*>(outState._lineCount.data()), sizeof(GLsizei)*numLines);
-
-    // outState._vertexPositions.reserve(numPoints);
     outState._vertexPositions.resize(numPoints);
+    outState._extraVariables.resize(numExtras);
+    outState._extraVariableNames.reserve(numExtras);
+
[…]
