### 2017-06-07 0725ef3
Decode using striped memory buffers with Kakadu. At least 3x speed improvement.
