+++ modules/solarbrowsing/util/decodejob.h
@@ -25,6 +25,7 @@
 #define __OPENSPACE_MODULE_SOLARBROWSING___DECODEJOB___H__
 
 #include <modules/solarbrowsing/util/streambuffer.h>
+#include <modules/solarbrowsing/util/kakaduwrapper.h>
 
 namespace openspace {
 
@@ -61,16 +62,19 @@ public:
 
         // TODO RESOLUTION BUGG!!!!
         const unsigned int totalImageSize
-              = _decodeData.im->fullResolution * _decodeData.im->fullResolution;
+              = (_decodeData.im->fullResolution / (_decodeData.resolutionLevel + 1)) * _decodeData.im->fullResolution /  (_decodeData.resolutionLevel + 1);
         SolarImageData imd {new unsigned char[totalImageSize],
                             _decodeData.im,
                             _decodeData.timeObserved};
         SimpleJ2kCodec j2c(_decodeData.verboseMode);
 
+        KakaduWrapper w(_decodeData.verboseMode);
+        w.DecodeIntoBuffer(imd.im->filename, imd.data, _decodeData.resolutionLevel);
+        //j2c.DecodeIntoBuffer(imd.im->filename, imd.data, _decodeData.resolutionLevel);
+
       /*  if (imd.im->preprocessedFilename != "") {
             j2c.DecodeBMPIntoBuffer(imd.im->preprocessedFilename, imd.data);
         } else {*/
-        j2c.DecodeIntoBuffer(imd.im->filename, imd.data, _decodeData.resolutionLevel);
         //}
         _solarImageData = std::make_shared<SolarImageData>(std::move(imd));
     }
