+++ modules/solarbrowsing/rendering/renderablesolarimagery.cpp
@@ -190,11 +190,11 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         // Upload asap
         updateTextureGPU(/*asyncUpload=*/false);
         if (_useBuffering) {
-            //double oktime = OsEng.timeManager().time().deltaTime();
+            double oktime = OsEng.timeManager().time().deltaTime();
             //TimeManager& lel = OsEng.timeManager();
-            _currentActiveImageTime
-                      = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
-                              .timeObserved;
+            // _currentActiveImageTime
+            //           = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
+            //                   .timeObserved;
             _streamBuffer.clear();
             //fillBuffer(OsEng.timeManager().time().deltaTime());
         } /*else {
@@ -209,10 +209,9 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         _pbo->setSize(_imageSize * _imageSize * sizeof(IMG_PRECISION));
         updateTextureGPU(/*asyncUpload=*/false, /*resChanged=*/true);
         if (_useBuffering) {
-
-            _currentActiveImageTime
-                      = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
-                              .timeObserved;
+            // _currentActiveImageTime
+            //           = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
+            //                   .timeObserved;
             _streamBuffer.clear();
             //fillBuffer(OsEng.timeManager().time().deltaTime());
         } /*else {
@@ -330,9 +329,16 @@ void RenderableSolarImagery::uploadImageDataToPBO() {
         _pboIsDirty = true;
     } else {
         // WARNING - this can be an old job - but looks smoother - bug or feature?
-        std::shared_ptr<SolarImageData> _solarImageData = _streamBuffer.popFinishedJob();
-        if (_solarImageData) {
+        const double& osTime = OsEng.timeManager().time().j2000Seconds();
+        std::shared_ptr<SolarImageData> _solarImageData = nullptr;
+        while (_streamBuffer._concurrentJobManager.numFinishedJobs() > 0) {
+            _solarImageData = _streamBuffer.popFinishedJob();
+            if (osTime <= _solarImageData->timeObserved) {
+                break;
+            }
+        }
 
+        if (_solarImageData) {
             _currentActiveImageTime = _solarImageData->timeObserved;
             _currentScale = _solarImageData->im->scale;
             _currentCenterPixel = _solarImageData->im->centerPixel;
@@ -358,7 +364,8 @@ void RenderableSolarImagery::uploadImageDataToPBO() {
             }
         } else {
             if (_verboseMode) {
-                LWARNING("Nothing to update, buffer is not ready");
+                LWARNING("Nothing to update, buffer is not ready, missing frames " << _frameSkipCount);
+                _frameSkipCount++;
             }
         }
     }
@@ -400,8 +407,10 @@ void RenderableSolarImagery::updateTextureGPU(bool asyncUpload, bool resChanged)
 
 void RenderableSolarImagery::decode(unsigned char* buffer, const std::string& filename)
 {
-    SimpleJ2kCodec j2c(_verboseMode);
-    j2c.DecodeIntoBuffer(filename, buffer, _resolutionLevel);
[…]
