### 2017-06-08 9a7eba1
Remove warnings Update Ghoul repository Fix spelling mistake in openspace.cfg
