+++ src/engine/downloadmanager.cpp
@@ -169,19 +169,21 @@ std::shared_ptr<DownloadManager::FileFuture> DownloadManager::downloadFile(
 #ifdef WIN32
     FILE* fp;
     errno_t error = fopen_s(&fp, file.path().c_str(), "wb");
-    ghoul_assert(
-        error == 0,
-        "Could not open/create file:" + file.path() + ". Errno: " + std::to_string(errno)
-    );
+    if (error != 0) {
+        LERROR(
+            "Could not open/create file:" + file.path() +
+            ". Errno: " + std::to_string(errno)
+        );
+    }
 #else
     FILE* fp = fopen(file.path().c_str(), "wb"); // write binary
 #endif // WIN32
-    ghoul_assert(
-        fp != nullptr,
-        "Could not open/create file:" + file.path() + ". Errno: " + std::to_string(errno)
-    );
-
-    //LDEBUG("Start downloading file: '" << url << "' into file '" << file.path() << "'");
+    if (!fp) {
+        LERROR(
+            "Could not open/create file:" + file.path() +
+            ". Errno: " + std::to_string(errno)
+        );
+    }
     
     auto downloadFunction = [url,
                              failOnError,
