+++ src/engine/wrapper/sgctwindowwrapper.cpp
@@ -158,6 +158,7 @@ int SGCTWindowWrapper::currentNumberOfAaSamples() const {
 bool SGCTWindowWrapper::isRegularRendering() const {
     sgct::SGCTWindow* w = sgct::Engine::instance()->getCurrentWindowPtr();
     std::size_t nViewports = w->getNumberOfViewports();
+    (void)nViewports; // Unused in Release mode
     ghoul_assert(nViewports > 0, "At least one viewport must exist at this time");
     sgct_core::Viewport* vp = w->getViewport(0);
     sgct_core::NonLinearProjection* nlp = vp->getNonLinearProjectionPtr();
