+++ openspace.cfg
@@ -79,7 +79,7 @@ return {
     -- OnScreenTextScaling = "framebuffer",
     -- PerSceneCache = true,
     -- DisableRenderingOnMaster = true,
-    KeyDisableSceneOnMaster = true,
+    -- DisableSceneOnMaster = true,
     DownloadRequestURL = "http://data.openspaceproject.com/request.cgi",
     RenderingMethod = "Framebuffer",
     OpenGLDebugContext = {
