### 2017-06-08 65efef2
Fix order of deletion in concurrent job manager and clean up
