+++ modules/globebrowsing/rendering/layer/layerrendersettings.cpp
@@ -37,11 +37,12 @@ namespace {
 
 LayerRenderSettings::LayerRenderSettings()
     : properties::PropertyOwner("Settings")
-    , opacity(properties::FloatProperty("Opacity", "opacity", 1.f, 0.f, 1.f))    
-    , gamma(properties::FloatProperty("Gamma", "gamma", 1, 0, 5))
-    , multiplier(properties::FloatProperty("Multiplier", "multiplier", 1.f, 0.f, 20.f))
-    , offset(properties::FloatProperty("Offset", "offset", 0.f, -10000.f, 10000.f))
-    , valueBlending(properties::FloatProperty("Value blending", "valueBlending",
+    , setDefault("setDefault", "Set Default")
+    , opacity(properties::FloatProperty("opacity", "Opacity", 1.f, 0.f, 1.f))    
+    , gamma(properties::FloatProperty("gamma", "Gamma", 1, 0, 5))
+    , multiplier(properties::FloatProperty("multiplier", "Multiplier", 1.f, 0.f, 20.f))
+    , offset(properties::FloatProperty("offset", "Offset", 0.f, -10000.f, 10000.f))
+    , valueBlending(properties::FloatProperty("valueBlending", "Value Blending",
                                                1.f, 0.f, 1.f))
     , useValueBlending(false)
 {
@@ -50,6 +51,9 @@ LayerRenderSettings::LayerRenderSettings()
     addProperty(gamma);
     addProperty(multiplier);
     addProperty(offset);
+    addProperty(setDefault);
+
+    setDefault.onChange([this](){ setDefaultValues(); });
 }
 
 void LayerRenderSettings::setValuesFromDictionary(
@@ -98,6 +102,14 @@ glm::vec4 LayerRenderSettings::performLayerSettings(glm::vec4 currentValue) cons
         performLayerSettings(currentValue.a));
 
     return newValue;
+}
+
+void LayerRenderSettings::setDefaultValues() {
+    opacity.setValue(1.0f);
+    gamma.setValue(1.0f);
+    multiplier.setValue(1.0f);
+    offset.setValue(0.0f);
+    valueBlending.setValue(1.0f);
 }
 
 } // namespace globebrowsing
