+++ modules/globebrowsing/other/lruthreadpool.h
@@ -65,6 +65,7 @@ template<typename KeyType>
 class LRUThreadPool {
 public:
     LRUThreadPool(size_t numThreads, size_t queueSize);
+    LRUThreadPool(const LRUThreadPool& toCopy);
     ~LRUThreadPool();
 
     void enqueue(std::function<void()> f, KeyType key);
