### 2017-06-08 e75bb9e
Can create by level tile provider with empty dictionary.
