### 2017-06-08 80fabcb
Allow the taskrunner to be configured with command line args
