### 2017-06-08 024a5ac
Configure logging properly for task runner
