+++ apps/TaskRunner/main.cpp
@@ -32,6 +32,8 @@
 #include <ghoul/io/texture/texturereaderfreeimage.h>
 #include <ghoul/filesystem/filesystem.h>
 #include <ghoul/filesystem/directory.h>
+#include <ghoul/logging/logmanager.h>
+#include <ghoul/logging/consolelog.h>
 #include <ghoul/ghoul.h>
 
 #include <openspace/scripting/scriptengine.h>
@@ -97,6 +99,22 @@ int main(int argc, char** argv) {
     using namespace openspace;
 
     ghoul::initialize();
+
+    ghoul::logging::LogManager::initialize(
+        ghoul::logging::LogLevel::Debug,
+        ghoul::logging::LogManager::ImmediateFlush::Yes
+    );
+    LogMgr.addLog(std::make_unique< ghoul::logging::ConsoleLog>());
+
+    LDEBUG("Initialize FileSystem");
+
+#ifdef __APPLE__
+    ghoul::filesystem::File app(argv[0]);
+    std::string dirName = app.directoryName();
+    LINFO("Setting starting directory to '" << dirName << "'");
+    FileSys.setCurrentDirectory(dirName);
+#endif
+
     initTextureReaders();
 
     FactoryManager::initialize();
