### 2017-06-08 f736434
Minor improvements and fixes in volume rendering
