### 2017-06-08 23fad9e
Update to use local wrapper, also add property to enable/disable depth
