+++ modules/globebrowsing/models/renderableroversurface.cpp
@@ -60,7 +60,8 @@ namespace globebrowsing {
 				BoolProperty("enable", "Enabled", true),
 				BoolProperty("enablePath", "Enable path", true),
 				BoolProperty("lockSubsite", "Lock subsite", false),
-				BoolProperty("useMastCam", "Show mastcam coloring", false)
+				BoolProperty("useMastCam", "Show mastcam coloring", false),
+				BoolProperty("enableDepth", "Enable depth", true)
 		})
 		, _debugModelRotation("modelrotation", "Model Rotation", glm::vec3(0.0f), glm::vec3(0.0f), glm::vec3(360.0f))
 		, _modelSwitch()
@@ -101,6 +102,7 @@ namespace globebrowsing {
 	addProperty(_generalProperties.enablePath);
 	addProperty(_generalProperties.lockSubsite);
 	addProperty(_generalProperties.useMastCam);
+	addProperty(_generalProperties.enableDepth);
 	addProperty(_debugModelRotation);
 
 	_cachingModelProvider = std::make_shared<CachingSurfaceModelProvider>(this);
@@ -213,58 +215,6 @@ void RenderableRoverSurface::render(const RenderData& data) {
 
 	_programObject->activate();
 	for (auto subsiteModels : vectorOfsubsiteModels) {
-
-		/*float dir = 1;
-		float alpha = subsiteModels->_alpha;
-		int subsitePrevLevel = subsiteModels->level;
-		if (level != _prevLevel) {
-			// We went from one level to the next
-			// we need to fade out previously rendered models
-			if (level != subsitePrevLevel && alpha >= 0.0) {
-				// Current subsite should fade out
-				dir *= -1;
-			}
-			else if (alpha >= 1.0) {
-				// Current subsite is the correct level and has been faded in correctly
-				dir = 0;
-			}
-
-			if (level == 3) {
-				_prevSubsite = subsiteModels;
-			}
-		}
-		else if (_prevSubsite != nullptr 
-			&& _prevSubsite->drive != subsiteModels->drive 
-			&& level == 3) {
-			// Walking between models
-			// And this is the new model that should be faded in
-			// if it's not already done
-			dir = subsiteModels->_alpha >= 1.0 ? 0 : 1;
-			if (dir == 0.0) {
-				_prevSubsite = subsiteModels;
-			}
-		}
-		else if (_prevSubsite != nullptr 
-			&& vectorOfsubsiteModels.size() > 1
-			&& _prevSubsite->drive == subsiteModels->drive 
-			&& level == 3) {
-			// Walking between models
-			// And this is the previous model that should be faded out
-			// if it's not already done
-			dir = subsiteModels->_alpha <= 0.0 ? 0 : -1;
-		}
-		else {
-			dir = subsiteModels->_alpha >= 1.0 ? 0 : 1;
-		}*/
-
-		//alpha = alpha + dir * 0.005;
-		//subsiteModels->_alpha = alpha;
-
-		/*if (_prevSubsiteModels != nullptr && _prevSubsiteModels != subsiteModels
-				&& level == 3) {
-			_prevSubsiteModels = subsiteModels;
-		}*/
-
 		glm::dmat4 globeTransform = _globe->modelTransform();
 
 		glm::dvec3 positionWorldSpace = globeTransform * glm::dvec4(subsiteModels->cartesianPosition, 1.0);
@@ -273,10 +223,6 @@ void RenderableRoverSurface::render(const RenderData& data) {
[…]
