### 2017-06-08 ac9ec2a
Include reading of subsite rotationmatrix, used mainly for colorized textures
