+++ modules/globebrowsing/models/roverpathfilereader.h
@@ -41,10 +41,11 @@ public:
 
 	static std::vector<std::shared_ptr<Subsite>> extractAllSubsites(const ghoul::Dictionary dictionary);
 	static std::vector<std::shared_ptr<Subsite>> extractSubsitesWithModels(const ghoul::Dictionary dictionary);
-
+	
 private:
 	static std::string RoverPathFileReader::convertString(const std::string sitenr, const std::string type);
 	static std::shared_ptr<TextureInformation> RoverPathFileReader::extractTextureInfo(const std::string absoluteFilePath);
+	static glm::dmat4 extractRotationMatrix(std::string filename);
 };
 } // namespace globebrowsing
 } // namespace openspace
