### 2017-06-08 897f3f9
Update to use new texturearray and subsite rotation matrix
