### 2017-06-08 c0fd234
Update CMakeLists to include texture array
