### 2017-06-08 1169625
Update GDAL version to 2.2.0
