+++ modules/globebrowsing/ext/gdal/include/gdalexif.h
@@ -1,5 +1,5 @@
 /******************************************************************************
- * $Id: gdalexif.h 33175 2016-01-27 00:37:49Z goatbar $
+ * $Id: gdalexif.h 35883 2016-10-24 05:49:15Z goatbar $
  *
  * Project:  JPEG JFIF Driver
  * Purpose:  Implement GDAL JPEG Support based on IJG libjpeg.
@@ -26,6 +26,8 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  ****************************************************************************/
+
+//! @cond Doxygen_Suppress
 
 #ifdef RENAME_INTERNAL_LIBTIFF_SYMBOLS
 #include "../frmts/gtiff/libtiff/gdal_libtiff_symbol_rename.h"
@@ -73,119 +75,117 @@ static const struct tagname {
   GUInt16       tag;
   const char*   name;
 } tagnames [] = {
-
-//    { 0x100,	"EXIF_Image_Width"},
-//    { 0x101,	"EXIF_Image_Length"},
-    { 0x102,	"EXIF_BitsPerSample"},
-    { 0x103,	"EXIF_Compression"},
-    { 0x106,	"EXIF_PhotometricInterpretation"},
-    { 0x10A,	"EXIF_Fill_Order"},
-    { 0x10D,	"EXIF_Document_Name"},
-    { 0x10E,	"EXIF_ImageDescription"},
-    { 0x10F,	"EXIF_Make"},
-    { 0x110,	"EXIF_Model"},
-    { 0x111,	"EXIF_StripOffsets"},
-    { 0x112,	"EXIF_Orientation"},
-    { 0x115,	"EXIF_SamplesPerPixel"},
-    { 0x116,	"EXIF_RowsPerStrip"},
-    { 0x117,	"EXIF_StripByteCounts"},
-    { 0x11A,	"EXIF_XResolution"},
-    { 0x11B,	"EXIF_YResolution"},
-    { 0x11C,	"EXIF_PlanarConfiguration"},
-    { 0x128,	"EXIF_ResolutionUnit"},
-    { 0x12D,	"EXIF_TransferFunction"},
-    { 0x131,	"EXIF_Software"},
-    { 0x132,	"EXIF_DateTime"},
-    { 0x13B,	"EXIF_Artist"},
-    { 0x13E,	"EXIF_WhitePoint"},
-    { 0x13F,	"EXIF_PrimaryChromaticities"},
-    { 0x156,	"EXIF_Transfer_Range"},
-    { 0x200,	"EXIF_JPEG_Proc"},
-    { 0x201,	"EXIF_JPEGInterchangeFormat"},
-    { 0x202,	"EXIF_JPEGInterchangeFormatLength"},
-    { 0x211,	"EXIF_YCbCrCoefficients"},
-    { 0x212,	"EXIF_YCbCrSubSampling"},
-    { 0x213,	"EXIF_YCbCrPositioning"},
-    { 0x214,	"EXIF_ReferenceBlackWhite"},
-    { 0x828D,	"EXIF_CFA_Repeat_Pattern_Dim"},
-    { 0x828E,	"EXIF_CFA_Pattern"},
-    { 0x828F,	"EXIF_Battery_Level"},
-    { 0x8298,	"EXIF_Copyright"},
-    { 0x829A,	"EXIF_ExposureTime"},
-    { 0x829D,	"EXIF_FNumber"},
-    { 0x83BB,	"EXIF_IPTC/NAA"},
-//	{ 0x8769,	"EXIF_Offset"},
[…]
