### 2017-06-08 f487107
Added missing semicolon to vertex shader.
