+++ openspace.cfg
@@ -34,6 +34,7 @@ return {
     -- Scene = "${SCENE}/newhorizons.scene",
     Scene = "${SCENE}/atmosphereearth.scene",
     -- Scene = "${SCENE}/osirisrex.scene",
+    -- Scene = "${SCENE}/satellites.scene",
 
     Task = "${TASKS}/default.task",
 
