### 2017-06-08 e7d04ff
Fix typo in NPM building instructions
