+++ apps/OpenSpace/CMakeLists.txt
@@ -242,7 +242,7 @@ if (OPENSPACE_MODULE_WEBGUI AND CEF_ROOT)
     add_custom_command(
             TARGET OpenSpace POST_BUILD
             COMMAND ${NPM_COMMAND} install
-            COMMAND ${NPM_COMMAND} run b uild
+            COMMAND ${NPM_COMMAND} run build
             COMMAND ${CMAKE_COMMAND} -E copy_directory ${WEBGUI_MODULE_PATH}/web/dist "$<TARGET_FILE_DIR:${CEF_TARGET}>/gui"
             COMMAND ${CMAKE_COMMAND} -E copy_directory ${WEBGUI_MODULE_PATH}/web/dist/static "$<TARGET_FILE_DIR:${CEF_TARGET}>/gui/static"
             WORKING_DIRECTORY ${WEBGUI_MODULE_PATH}/web
