### 2017-06-08 1b3325b
[WebGUI] Add support for rendering web as a screen space object
This allows the user to add an arbitrary number of web browser windows to
the screen.
