### 2017-06-08 a89f4a3
Fix ScreenSpaceBrowser mirrored rendering issue
