### 2017-06-08 aa1b160
Implement new version of buffer
