+++ modules/solarbrowsing/rendering/renderablesolarimagery.cpp
@@ -56,7 +56,7 @@ typedef std::chrono::high_resolution_clock Clock;
 
 namespace {
     static const std::string _loggerCat = "RenderableSolarImagery";
-    double HALF_SUN_RADIUS = (1391600000.0 * 0.50); // Half sun radius divided by magic factor
+    //double HALF_SUN_RADIUS = (1391600000.0 * 0.50); // Half sun radius divided by magic factor
     const double EPSILON = std::numeric_limits<double>::epsilon();
 }
 
@@ -190,16 +190,24 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         // Upload asap
         updateTextureGPU(/*asyncUpload=*/false);
         if (_useBuffering) {
-            double oktime = OsEng.timeManager().time().deltaTime();
+            //double oktime = OsEng.timeManager().time().deltaTime();
             //TimeManager& lel = OsEng.timeManager();
             // _currentActiveImageTime
             //           = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
             //                   .timeObserved;
             _streamBuffer.clear();
+            _frameSkipCount = 0;
+            _bufferCountOffset = 1;
             //fillBuffer(OsEng.timeManager().time().deltaTime());
         } /*else {
             uploadImageDataToPBO();
         }*/
+    });
+
+    _minRealTimeUpdateInterval.onChange([this]() {
+        _streamBuffer.clear();
+        _frameSkipCount = 0;
+        _bufferCountOffset = 1;
     });
 
     _resolutionLevel.onChange([this]() {
@@ -213,6 +221,8 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
             //           = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
             //                   .timeObserved;
             _streamBuffer.clear();
+            _frameSkipCount = 0;
+            _bufferCountOffset = 1;
             //fillBuffer(OsEng.timeManager().time().deltaTime());
         } /*else {
             uploadImageDataToPBO();
@@ -232,9 +242,9 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     // }
 
 
-    _currentActiveImageTime
-          = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
-                  .timeObserved;
+    // _currentActiveImageTime
+    //       = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
+    //               .timeObserved;
 
    // _imageMetadataMap2[_currentActiveInstrument].displayStateTimes();
 
@@ -282,28 +292,28 @@ DecodeData RenderableSolarImagery::getDecodeDataFromOsTime(const int& osTime) {
     //DecodeData decodeData {_imageSize * _imageSize, stateFilename, _resolutionLevel, _verboseMode, timeObserved};
 
     DecodeData decodeData {std::move(im), static_cast<unsigned int>(_resolutionLevel), timeObserved, _verboseMode};
-    return std::move(decodeData);
+    return decodeData;
 }
 
-void RenderableSolarImagery::fillBuffer(const double& dt) {
-    _streamBuffer.clear();
-    if (_verboseMode) {
-        LDEBUG("Refilling Buffer. dt: " << dt);
-    }
-    const double& osTime = OsEng.timeManager().time().j2000Seconds();
-    const double& startTime = getDecodeDataFromOsTime(osTime).timeObserved;
+// void RenderableSolarImagery::fillBuffer(const double& dt) {
+//     _streamBuffer.clear();
+//     if (_verboseMode) {
[…]
