### 2017-06-09 2db3092
Add script to add GIBS datasets.
