+++ include/openspace/properties/propertyowner.h
@@ -183,8 +183,6 @@ public:
      * \param owner The owner that should be assigned to this PropertyOwner
      */
     void addPropertySubOwner(PropertyOwner* owner);
-  
-    bool addPropertySubOwnerUnsorted(PropertyOwner* owner);
 
     /// \see PropertyOwner::addPropertySubOwner(PropertyOwner*)
     void addPropertySubOwner(PropertyOwner& owner);
