### 2017-06-09 215020b
Start working with layer adjustment
