### 2017-06-09 fd0cef0
Initialize with a list of all the models that can be loaded
