### 2017-06-09 5e1894d
Update globebrowsing module with conditional group, temporary until meshgeneration is in its own module
