### 2017-06-09 d1b98ff
Fix memory leak
