### 2017-06-09 7775435
Do not sort properties and propertyowners automatically (closes #269)
