### 2017-06-09 ec8fc78
Add OpenGL Debug supression for Intel card GL_LINE_WIDTH warnings
