+++ openspace.cfg
@@ -66,7 +66,7 @@ return {
         CapabilitiesVerbosity = "Full"
     },
     Launcher = {
-        LogLevel="None"
+        LogLevel = "None"
     },
     LuaDocumentation = "${DOCUMENTATION}/LuaScripting.html",
     PropertyDocumentation = "${DOCUMENTATION}/Properties.html",
@@ -87,6 +87,7 @@ return {
        FilterIdentifier = {
            { Type = "Other", Source = "API", Identifier = 131185 },
            { Type = "Performance", Source = "API", Identifier = 131186 } --Buffer performance warning: "copied/moved from VIDEO memory to HOST memory"
+           { Type = "Deprecated", Source = "API", Identifier = 7} -- API_ID_LINE_WIDTH deprecated behavior warning has been generated
        },
 --      FilterSeverity = { }
 
