+++ openspace.cfg
@@ -86,7 +86,7 @@ return {
        Activate = true,
        FilterIdentifier = {
            { Type = "Other", Source = "API", Identifier = 131185 },
-           { Type = "Performance", Source = "API", Identifier = 131186 } --Buffer performance warning: "copied/moved from VIDEO memory to HOST memory"
+           { Type = "Performance", Source = "API", Identifier = 131186 }, --Buffer performance warning: "copied/moved from VIDEO memory to HOST memory"
            { Type = "Deprecated", Source = "API", Identifier = 7} -- API_ID_LINE_WIDTH deprecated behavior warning has been generated
        },
 --      FilterSeverity = { }
