### 2017-06-09 82f74bb
Missing code for colored textures
