### 2017-06-09 90c7eef
Remove explicit coordinate transformation from RenderableConstellationBounds (closes #300)
