### 2017-06-09 03bec73
Prevent a hard crash in case the scene file contains a syntax error
