+++ src/network/parallelconnection.cpp
@@ -62,27 +62,26 @@
 #endif
 
 #include <openspace/network/parallelconnection.h>
+
+#include <openspace/openspace.h>
 #include <openspace/engine/openspaceengine.h>
 #include <openspace/engine/wrapper/windowwrapper.h>
 #include <openspace/interaction/interactionhandler.h>
 #include <openspace/interaction/interactionmode.h>
 #include <openspace/scene/scenegraphnode.h>
-#include <openspace/util/timemanager.h>
-#include <openspace/interaction/interactionmode.h>
-#include <openspace/util/time.h>
-#include <openspace/openspace.h>
 #include <openspace/scripting/script_helper.h>
+#include <openspace/util/time.h>
+#include <openspace/util/timemanager.h>
 
 #include <ghoul/logging/logmanager.h>
 
-//lua functions
 #include "parallelconnection_lua.inl"
 
 namespace {
     const uint32_t ProtocolVersion = 2;
     const size_t MaxLatencyDiffs = 64;
-    const std::string _loggerCat = "ParallelConnection";
-}
+    const char* _loggerCat = "ParallelConnection";
+} // namespace
 
 namespace openspace {
 
@@ -94,8 +93,20 @@ ParallelConnection::ParallelConnection()
     , _address("address", "Address", "localhost")
     , _name("name", "Connection name", "Anonymous")
     , _bufferTime("bufferTime", "Buffer Time", 1, 0.5, 10)
-    , _timeKeyframeInterval("timeKeyframeInterval", "Time keyframe interval", 0.1f, 0.f, 1.f)
-    , _cameraKeyframeInterval("cameraKeyframeInterval", "Camera Keyframe interval", 0.1f, 0.f, 1.f)
+    , _timeKeyframeInterval(
+        "timeKeyframeInterval",
+        "Time keyframe interval",
+        0.1f,
+        0.f,
+        1.f
+    )
+    , _cameraKeyframeInterval(
+        "cameraKeyframeInterval",
+        "Camera Keyframe interval",
+        0.1f,
+        0.f,
+        1.f
+    )
     , _timeTolerance("timeTolerance", "Time tolerance", 1.f, 0.5f, 5.f)
     , _lastTimeKeyframeTimestamp(0)
     , _lastCameraKeyframeTimestamp(0)
@@ -127,50 +138,51 @@ ParallelConnection::ParallelConnection()
     addProperty(_timeTolerance);
 
     _connectionEvent = std::make_shared<ghoul::Event<>>();
-    _handlerThread = std::make_unique<std::thread>(&ParallelConnection::threadManagement, this);
+    _handlerThread = std::make_unique<std::thread>(
+        &ParallelConnection::threadManagement,
+        this
+    );
 }
         
-ParallelConnection::~ParallelConnection(){
-            
-    //signal that a disconnect should occur
+ParallelConnection::~ParallelConnection() {
+    // signal that a disconnect should occur
     signalDisconnect();
             
-    //signal that execution has stopped
+    // signal that execution has stopped
     _isRunning.store(false);
             
-    //join handler
+    // join handler
     _handlerThread->join();
 }
         
-void ParallelConnection::threadManagement(){
+void ParallelConnection::threadManagement() {
     // The _disconnectCondition.wait(unqlock) stalls
     // How about moving this out of the thread and into the destructor? ---abock
     
-    //while we're still running
+    // while we're still running
     while(_isRunning){
-        {
-            //lock disconnect mutex mutex
-            //not really needed since no data is modified but conditions need a mutex
-            std::unique_lock<std::mutex> disconnectLock(_disconnectMutex);
-            //wait for a signal to disconnect
-            _disconnectCondition.wait(disconnectLock, [this]() { return _disconnect.load(); });
[…]
