### 2017-06-09 cd7d1ab
Fix hardcoded path in solar projection
