+++ modules/solarbrowsing/rendering/renderablesolarimageryprojection.h
@@ -67,6 +67,7 @@ private:
 
     std::vector<SceneGraphNode*> _solarImageryDependencies;
     std::string _nodeName;
+    std::string path;
     std::unique_ptr<ghoul::opengl::Texture> _magnetogramTexture;
     float _planetRadius;
 };
