### 2017-06-09 d3f574b
render TouchMarkers, currently causes crash on runtime
