+++ modules/touch/src/TouchMarker.cpp
@@ -25,11 +25,12 @@
 #include <modules/touch/include/TouchMarker.h>
 
 #include <openspace/engine/openspaceengine.h>
-#include <openspace/engine/settingsengine.h>
-#include <openspace/engine/wrapper/windowwrapper.h>
-#include <openspace/interaction/interactionhandler.h>
 #include <openspace/rendering/renderengine.h>
-#include <openspace/rendering/screenspacerenderable.h>
+#include <ghoul/filesystem/filesystem.h>
+#include <ghoul/io/texture/texturereader.h>
+#include <ghoul/opengl/programobject.h>
+#include <ghoul/opengl/texture.h>
+#include <ghoul/opengl/textureunit.h>
 
 #include <ghoul/logging/logmanager.h>
 
@@ -38,7 +39,141 @@ namespace {
 }
 namespace openspace {
 
-TouchMarker::TouchMarker() 
-	: properties::PropertyOwner("TouchMarker") { }
+TouchMarker::TouchMarker()
+	: properties::PropertyOwner("TouchMarker")
+	, _visible("TouchMarkers visible", "Toggle visibility of markers", false)
+	, _radiusSize("Marker size", "Set marker size in radius", 0.5, 0, 1)
+	, _texturePath("texturePath", "Color Texture")
+	, _shader(nullptr)
+	, _texture(nullptr)
+	, _listIsDirty(false)
+	, _textureIsDirty(false)
+	, _numFingers(0)
+{
+	addProperty(_visible);
+	addProperty(_radiusSize);
+	addProperty(_texturePath);
+	_texturePath.onChange(std::bind(&TouchMarker::loadTexture, this));
+
+	
+}
+
+bool TouchMarker::isReady() const {
+	if (_shader == nullptr) {
+		std::cout << "something went wrong\n";
+	}
+	return (_shader != nullptr); // && _texture;
+}
+
+bool TouchMarker::initialize(const std::vector<TUIO::TuioCursor> list) {
+	glGenVertexArrays(1, &_quad); // generate array
+	glGenBuffers(1, &_vertexPositionBuffer); // generate buffer
+	createVertexList(list);
+
+	_shader = OsEng.renderEngine().buildRenderProgram("MarkerProgram",
+		"${MODULE_TOUCH}/shaders/marker_vs.glsl",
+		"${MODULE_TOUCH}/shaders/marker_fs.glsl"
+	);
+
+	//loadTexture();
+
+	return isReady();
+}
+
+bool TouchMarker::deinitialize() {
+	glDeleteVertexArrays(1, &_quad);
+	_quad = 0;
+
+	glDeleteBuffers(1, &_vertexPositionBuffer);
+	_vertexPositionBuffer = 0;
+
+	_texture = nullptr;
+
+	RenderEngine& renderEngine = OsEng.renderEngine();
+	if (_shader) {
+		renderEngine.removeRenderProgram(_shader);
+		_shader = nullptr;
+	}
+
+	return true;
+}
+
+void TouchMarker::render(const std::vector<TUIO::TuioCursor> list) {
+	if (_visible) {
+		createVertexList(list);
+		_shader->activate();
+
+		// Bind texture
+		/*ghoul::opengl::TextureUnit unit;
+		unit.activate();
+		_texture->bind();
+		_shader->setUniform("texture1", unit);*/
+		_shader->setUniform("radius", _radiusSize);
+		
+		glEnable(GL_PROGRAM_POINT_SIZE); // Enable gl_PointSize in vertex shader
+		glPolygonMode(GL_FRONT_AND_BACK, GL_POINT);
[…]
