+++ modules/touch/CMakeLists.txt
@@ -25,7 +25,7 @@
 include(${OPENSPACE_CMAKE_EXT_DIR}/module_definition.cmake)
 
 set(HEADER_FILES
-	${CMAKE_CURRENT_SOURCE_DIR}/ext/levmarq.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/ext/levmarq.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/TuioEar.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/TouchInteraction.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/TouchMarker.h
@@ -33,17 +33,23 @@ set(HEADER_FILES
 source_group("Header Files" FILES ${HEADER_FILES})
 
 set(SOURCE_FILES
-	${CMAKE_CURRENT_SOURCE_DIR}/ext/levmarq.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/ext/levmarq.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/TuioEar.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/TouchInteraction.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/TouchMarker.cpp
 )
 source_group("Source Files" FILES ${SOURCE_FILES})
+
+set(SHADER_FILES
+	${CMAKE_CURRENT_SOURCE_DIR}/shaders/marker_fs.glsl
+	${CMAKE_CURRENT_SOURCE_DIR}/shaders/marker_vs.glsl
+)
+source_group("Shader Files" FILES ${SHADER_FILES})
 	
 create_new_module(
     "Touch"
     touch_module
-    ${HEADER_FILES} ${SOURCE_FILES}
+    ${HEADER_FILES} ${SOURCE_FILES} ${SHADER_FILES}
 )
 	
 include_external_library(${touch_module} libTUIO ${CMAKE_CURRENT_SOURCE_DIR}/ext/libTUIO)
