### 2017-06-09 23f524d
cleanup and shader compile fix
