### 2017-06-09 70f798c
Fixed ATM in RenderablePlanet. No jittering anymore. Still needs reflectivity and clouds.
