+++ modules/space/rendering/renderableplanet.cpp
@@ -27,9 +27,6 @@
 #include <openspace/documentation/documentation.h>
 #include <openspace/documentation/verifier.h>
 
-#include <openspace/engine/configurationmanager.h>
-#include <openspace/engine/openspaceengine.h>
-#include <openspace/rendering/renderengine.h>
 #include <modules/space/rendering/planetgeometry.h>
 #include <openspace/util/time.h>
 #include <openspace/util/spicemanager.h>
@@ -43,6 +40,14 @@
 #include <ghoul/opengl/textureunit.h>
 #include <ghoul/misc/invariants.h>
 
+#ifdef OPENSPACE_MODULE_ATMOSPHERE_ENABLED
+#include <openspace/rendering/deferredcastermanager.h>
+#include <modules/atmosphere/rendering/atmospheredeferredcaster.h>
+#include <openspace/engine/configurationmanager.h>
+#include <openspace/engine/openspaceengine.h>
+#include <openspace/rendering/renderengine.h>
+#endif
+
 #include <memory>
 #include <fstream>
 
@@ -51,15 +56,13 @@
 
 
 namespace {
-    const char* KeyGeometry = "Geometry";
-    const char* KeyRadius = "Radius";
-    const char* KeyColorTexture = "Textures.Color";
-    const char* KeyNightTexture = "Textures.Night";
+    const char* KeyGeometry      = "Geometry";
+    const char* KeyRadius        = "Radius";
+    const char* KeyColorTexture  = "Textures.Color";
+    const char* KeyNightTexture  = "Textures.Night";
     const char* KeyHeightTexture = "Textures.Height";
-    const char* KeyShading = "PerformShading";
-
-    
-    
+    const char* KeyShading       = "PerformShading";
+       
     static const std::string _loggerCat = "RenderablePlanet";
 
     const char* keyFrame                         = "Frame";
@@ -67,6 +70,27 @@ namespace {
     const char* keyShadowSource                  = "Source";
     const char* keyShadowCaster                  = "Caster";
     const char* keyBody                          = "Body";
+
+#ifdef OPENSPACE_MODULE_ATMOSPHERE_ENABLED
+    const std::string keyAtmosphere               = "Atmosphere";
+    const std::string keyAtmosphereRadius         = "AtmoshereRadius";
+    const std::string keyPlanetRadius             = "PlanetRadius";
+    const std::string keyAverageGroundReflectance = "PlanetAverageGroundReflectance";
+    const std::string keyRayleigh                 = "Rayleigh";
+    const std::string keyRayleighHeightScale      = "H_R";
+    const std::string keyOzone                    = "Ozone";
+    const std::string keyOzoneHeightScale         = "H_O";
+    const std::string keyMie                      = "Mie";
+    const std::string keyMieHeightScale           = "H_M";
+    const std::string keyMiePhaseConstant         = "G";
+    const std::string keyImage                    = "Image";
+    const std::string keyToneMappingOp            = "ToneMapping";
+    const std::string keyExposure                 = "Exposure";
+    const std::string keyGamma                    = "Gamma";
+    const std::string keyATMDebug                 = "Debug";
+    const std::string keyTextureScale             = "PreCalculatedTextureScale";
+    const std::string keySaveTextures             = "SaveCalculatedTextures";
+#endif
 } // namespace
 
 namespace openspace {
@@ -126,17 +150,58 @@ RenderablePlanet::RenderablePlanet(const ghoul::Dictionary& dictionary)
[…]
