### 2017-06-09 0b289a4
Some cleanup
