+++ src/rendering/abufferrenderer.cpp
@@ -23,10 +23,11 @@
  ****************************************************************************************/
 
 #include <openspace/rendering/abufferrenderer.h>
-#include <openspace/rendering/raycastermanager.h>
 
 #include <openspace/engine/openspaceengine.h>
 #include <openspace/engine/wrapper/windowwrapper.h>
+#include <openspace/performance/performancemeasurement.h>
+#include <openspace/rendering/raycastermanager.h>
 #include <openspace/rendering/renderengine.h>
 #include <openspace/rendering/renderable.h>
 #include <openspace/rendering/volumeraycaster.h>
@@ -34,23 +35,19 @@
 #include <openspace/util/camera.h>
 #include <openspace/util/updatestructures.h>
 #include <openspace/util/timemanager.h>
-#include <openspace/performance/performancemeasurement.h>
-
 
 #include <ghoul/opengl/programobject.h>
 #include <ghoul/opengl/textureunit.h>
 #include <ghoul/misc/dictionary.h>
 #include <ghoul/misc/exception.h>
 
-
-
 #include <string>
 #include <iterator>
 
 namespace {
-    const std::string _loggerCat = "ABufferRenderer";
-    const std::string BoundsFragmentShaderPath = "${SHADERS}/abuffer/boundsabuffer.frag";
-    const std::string RenderFragmentShaderPath = "${SHADERS}/abuffer/renderabuffer.frag";
+    const char* _loggerCat = "ABufferRenderer";
+    const char* BoundsFragmentShaderPath = "${SHADERS}/abuffer/boundsabuffer.frag";
+    const char* RenderFragmentShaderPath = "${SHADERS}/abuffer/renderabuffer.frag";
     const int MaxRaycasters = 32;
     const int MaxLayers = 32;
     const int MaxAverageLayers = 8;
@@ -92,7 +89,14 @@ void ABufferRenderer::initialize() {
     glBindBuffer(GL_ARRAY_BUFFER, _vertexPositionBuffer);
 
     glBufferData(GL_ARRAY_BUFFER, sizeof(vertex_data), vertex_data, GL_STATIC_DRAW);
-    glVertexAttribPointer(0, 4, GL_FLOAT, GL_FALSE, sizeof(GLfloat)*4, reinterpret_cast<void*>(0));
+    glVertexAttribPointer(
+        0,
+        4,
+        GL_FLOAT,
+        GL_FALSE,
+        sizeof(GLfloat) * 4,
+        reinterpret_cast<void*>(0)
+    );
     glEnableVertexAttribArray(0);
 
     glGenTextures(1, &_anchorPointerTexture);
@@ -110,16 +114,26 @@ void ABufferRenderer::initialize() {
     GLint defaultFbo;
     glGetIntegerv(GL_FRAMEBUFFER_BINDING, &defaultFbo);
 
-
-
     updateResolution();
     updateRendererData();
     updateRaycastData();
     updateResolveDictionary();
 
     glBindFramebuffer(GL_FRAMEBUFFER, _mainFramebuffer);
-    glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D_MULTISAMPLE, _mainColorTexture, 0);
-    glFramebufferTexture2D(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_TEXTURE_2D_MULTISAMPLE, _mainDepthTexture, 0);
+    glFramebufferTexture2D(
+        GL_FRAMEBUFFER,
+        GL_COLOR_ATTACHMENT0,
+        GL_TEXTURE_2D_MULTISAMPLE,
[…]
