### 2017-06-09 9e4e681
More cleanup
