+++ src/rendering/abufferrenderer.cpp
@@ -153,7 +153,7 @@ void ABufferRenderer::initialize() {
             "${SHADERS}/abuffer/resolveabuffer.frag",
             dict
         );
-    } catch (const ghoul::RuntimeErcomponent& e) {
+    } catch (const ghoul::RuntimeError& e) {
         LERRORC(e.component, e.message);
     }
 
@@ -401,14 +401,23 @@ void ABufferRenderer::clear() {
     glBindBuffer(GL_PIXEL_UNPACK_BUFFER, _anchorPointerTextureInitializer);
     glBindTexture(GL_TEXTURE_2D, _anchorPointerTexture);
 
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_R32UI, _resolution.x, _resolution.y, 0, GL_RED_INTEGER, GL_UNSIGNED_INT, NULL);
+    glTexImage2D(
+        GL_TEXTURE_2D,
+        0,
+        GL_R32UI,
+        _resolution.x,
+        _resolution.y,
+        0,
+        GL_RED_INTEGER,
+        GL_UNSIGNED_INT,
+        nullptr
+    );
     glBindBuffer(GL_PIXEL_UNPACK_BUFFER, 0);
 
     static const GLuint zero = 1;
     glBindBufferBase(GL_ATOMIC_COUNTER_BUFFER, 0, _atomicCounterBuffer);
     glBufferSubData(GL_ATOMIC_COUNTER_BUFFER, 0, sizeof(zero), &zero);
     glBindBufferBase(GL_ATOMIC_COUNTER_BUFFER, 0, 0);
-
 }
 
 void ABufferRenderer::updateResolution() {
@@ -416,21 +425,43 @@ void ABufferRenderer::updateResolution() {
 
     int totalPixels = _resolution.x * _resolution.y;
     glBindTexture(GL_TEXTURE_2D, _anchorPointerTexture);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_R32UI, _resolution.x, _resolution.y, 0, GL_RED_INTEGER, GL_UNSIGNED_INT, NULL);
+    glTexImage2D(
+        GL_TEXTURE_2D,
+        0,
+        GL_R32UI,
+        _resolution.x,
+        _resolution.y,
+        0,
+        GL_RED_INTEGER,
+        GL_UNSIGNED_INT,
+        nullptr
+    );
 
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 
     glBindBuffer(GL_PIXEL_UNPACK_BUFFER, _anchorPointerTextureInitializer);
-    glBufferData(GL_PIXEL_UNPACK_BUFFER, totalPixels * sizeof(GLuint), NULL, GL_STATIC_DRAW);
+    glBufferData(
+        GL_PIXEL_UNPACK_BUFFER,
+        totalPixels * sizeof(GLuint),
+        nullptr,
+        GL_STATIC_DRAW
+    );
 
-    GLuint* data = (GLuint*)glMapBuffer(GL_PIXEL_UNPACK_BUFFER, GL_WRITE_ONLY);
+    GLuint* data = reinterpret_cast<GLuint*>(
+        glMapBuffer(GL_PIXEL_UNPACK_BUFFER, GL_WRITE_ONLY)
+    );
     memset(data, 0x00, totalPixels * sizeof(GLuint));
     glUnmapBuffer(GL_PIXEL_UNPACK_BUFFER);
[…]
