### 2017-06-09 5ef64aa
gl_position is vec4, cleanup/tests
