+++ modules/globebrowsing/models/roverpathfilereader.cpp
@@ -151,9 +151,9 @@ std::vector<std::shared_ptr<Subsite>> RoverPathFileReader::extractSubsitesWithMo
 		// Convert the site and drive string to match the folder structure
 		std::string site = convertString(subsite->site, "site");
 		std::string drive = convertString(subsite->drive, "drive");
-		pathToDriveFolderLevel1 = absPathToTModels + "/level1/" + "site" + site + "/" + "drive" + drive;
-		pathToDriveFolderLevel2 = absPathToTModels + "/level2/" + "site" + site + "/" + "drive" + drive;
-		pathToDriveFolderLevel3 = absPathToTModels + "/level3/" + "site" + site + "/" + "drive" + drive;
+		pathToDriveFolderLevel1 = absPathToTModels + "level1/" + "site" + site + "/" + "drive" + drive;
+		pathToDriveFolderLevel2 = absPathToTModels + "level2/" + "site" + site + "/" + "drive" + drive;
+		pathToDriveFolderLevel3 = absPathToTModels + "level3/" + "site" + site + "/" + "drive" + drive;
 
 		// If the folder exists it means there are models for this subsite, then check if that
 		// specific site/drive combination has already been added. If the models haven't already been 
