### 2017-06-12 13ec943
Dereference JSON iterator in server
This makes sure that it compiles.
