### 2017-06-12 35ea2aa
Dereference JSON iterator in server
This makes sure that it compiles.
