### 2017-06-12 302de49
[Server] Move Topic to separate file
