### 2017-06-12 92b7dc0
[Server] Move Topic to separate file
