### 2017-06-12 697b318
Add LoggerCat to private scope. (again?)
