### 2017-06-12 b55bc58
Color tables and limits should now be set in Mod file. No longer harcoded
