+++ modules/fieldlinessequence/rendering/renderablefieldlinessequence.cpp
@@ -75,6 +75,10 @@ namespace {
 
     const char* keyExtraVariables               = "ExtraVariables";
     const char* keyExtraMagnitudeVariables      = "ExtraMagnitudeVariables";
+    const char* keyExtraMinMax                  = "ExtraMinMax";
+    const char* keyExtraMinMaxLimits            = "ExtraMinMaxLimits";
+    const char* keyExtraColorTablePaths         = "ColorTablePaths";
+    const char* keyExtraColorTableMinMaxs       = "ColorTableMinMax";
 
     // const char* keyVertexListFileType       = "FileType";
     // const char* keyVertexListFileTypeJson   = "Json";
@@ -154,10 +158,6 @@ RenderableFieldlinesSequence::RenderableFieldlinesSequence(const ghoul::Dictiona
     _loggerCat += " [" + name + "]";
 
     _dictionary = dictionary;
-
-    // TODO: REMOVE HARDCODED PATH
-    _transferFunctionPath = "${OPENSPACE_DATA}/colortables/uniform_heatmap_bcgyr.txt";
-    _transferFunction = std::make_shared<TransferFunction>(fsManager.getAbsPath(_transferFunctionPath));
 }
 
 bool RenderableFieldlinesSequence::isReady() const {
@@ -386,7 +386,7 @@ bool RenderableFieldlinesSequence::initialize() {
     }
 
     bool allowUnitColoring = false;
-    int numQuanityColorVariables = 0;
+    size_t numExtraVariables = 0;
     if (_numberOfStates > 0) {
         // Approximate the end time of last state (and for the sequence as a whole)
         _seqStartTime = _startTimes[0];
@@ -471,8 +471,8 @@ bool RenderableFieldlinesSequence::initialize() {
         _domainLimY.setValue(glm::vec2(ymin,ymax));
         _domainLimZ.setValue(glm::vec2(zmin,zmax));
 
-        numQuanityColorVariables = _states[0]._extraVariables.size();
-        if (numQuanityColorVariables > 0) {
+        numExtraVariables = _states[0]._extraVariables.size();
+        if (numExtraVariables > 0) {
             allowUnitColoring = (_states[0]._extraVariables[0].size() > 0) ? true : false;
         }
 
@@ -587,7 +587,7 @@ bool RenderableFieldlinesSequence::initialize() {
     if (allowUnitColoring) {
         _colorMethod.addOption(colorMethod::UNIFORM, "Uniform Color");
         _colorMethod.addOption(colorMethod::QUANTITY_DEPENDENT, "Quantity Dependent");
-        _colorMethod = UNIFORM;
+        _colorMethod = QUANTITY_DEPENDENT;
 
         // ASSUMING ALL STATES HAVE THE SAME COLOR VARIABLES
         for (size_t i = 0; i < _states[0]._extraVariableNames.size(); ++i) {
@@ -595,22 +595,90 @@ bool RenderableFieldlinesSequence::initialize() {
         }
 
         // Set tranferfunction min/max to min/max in the given range
+        if (_dictionary.hasKeyAndValue<ghoul::Dictionary>(keyExtraColorTablePaths)) {
+            // TODO: create helper function to extract all values of type from dictionary
+            ghoul::Dictionary colTabPathsDic =
+                    _dictionary.value<ghoul::Dictionary>(keyExtraColorTablePaths);
+            size_t numColorTables = colTabPathsDic.size();
+            if (numColorTables == 0) {
+                LERROR("Must specify ColorTablePaths!");
+                return false;
+            }
+            for (size_t i = 1; i <= numColorTables; ++i) {
+                _colorTablePaths.push_back(
+                        colTabPathsDic.value<std::string>( std::to_string(i) ) );
+            }
+            // If numberOfColorTables don't add up to numberOfExtraVariables extend or
[…]
