### 2017-06-13 d4f7a40
Add set RenderBin:overlay to makes lines render at correct depth when using additive blending
