### 2017-06-13 ed10a7c
[Server] Break out Connection to separate file, add sub-Topics
