### 2017-06-13 a5d9d03
Fix resolution bug with STEREO COR1
