+++ modules/solarbrowsing/rendering/renderablesolarimagery.cpp
@@ -114,11 +114,11 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     }
 
     //TODO(mnoven): Can't pass in an int to dictionary?
-    float tmpResolution;
-    if (!dictionary.getValue("Resolution", tmpResolution)){
-        throw ghoul::RuntimeError("Resolution has to be specified");
-    }
-    _fullResolution = static_cast<unsigned int>(tmpResolution);
+    // float tmpResolution;
+    // if (!dictionary.getValue("Resolution", tmpResolution)){
+    //     throw ghoul::RuntimeError("Resolution has to be specified");
+    // }
+    //_fullResolution = static_cast<unsigned int>(tmpResolution);
 
     if (dictionary.hasKeyAndValue<ghoul::Dictionary>("Instruments")) {
         ghoul::Dictionary instrumentDic = dictionary.value<ghoul::Dictionary>("Instruments");
@@ -162,8 +162,15 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         }
     }
 
-    _imageSize = _fullResolution / (pow(2, _resolutionLevel));
-    _pbo = std::make_unique<PixelBufferObject>(_imageSize * _imageSize * sizeof(IMG_PRECISION));
+    // Get image size
+    auto& stateSequenceStart = _imageMetadataMap2[_currentActiveInstrument];
+    auto& stateStart = stateSequenceStart.getState(OsEng.timeManager().time().j2000Seconds());
+    std::shared_ptr<ImageMetadata> imStart = stateStart.contents();
+    //_imageSize = _imageMetadataMap2[_currentActiveInstrument]. //_fullResolution / (pow(2, _resolutionLevel));
+    _imageSize = imStart->fullResolution / (pow(2, _resolutionLevel));
+
+    // Always give PBO maximum size
+    _pbo = std::make_unique<PixelBufferObject>(4096 * 4096 * sizeof(IMG_PRECISION));
 
     // TODO(mnoven): Faster to send GL_RGBA32F as internal format - GPU Pads anyways?
     _texture =  std::make_unique<Texture>(
@@ -184,9 +191,20 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     _lastUpdateRealTime = _realTime;
 
     _activeInstruments.onChange([this]() {
+        _updatingCurrentActiveChannel = true;
         _pboIsDirty = false;
         _currentActiveInstrument
                = _activeInstruments.getDescriptionByValue(_activeInstruments.value());
+
+        // Update image size
+        auto& stateSequence = _imageMetadataMap2[_currentActiveInstrument];
+        auto& state = stateSequence.getState(OsEng.timeManager().time().j2000Seconds());
+       // const double& timeObserved = state.timeObserved();
+        std::shared_ptr<ImageMetadata> im = state.contents();
+        //_imageSize = _imageMetadataMap2[_currentActiveInstrument]. //_fullResolution / (pow(2, _resolutionLevel));
+        _imageSize = im->fullResolution / (pow(2, _resolutionLevel));
+        _pbo->setSize(_imageSize * _imageSize * sizeof(IMG_PRECISION));
+
         // Upload asap
         updateTextureGPU(/*asyncUpload=*/false);
         if (_useBuffering) {
@@ -202,6 +220,8 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         } /*else {
             uploadImageDataToPBO();
         }*/
+
+        _updatingCurrentActiveChannel = false;
     });
 
     _minRealTimeUpdateInterval.onChange([this]() {
[…]
