### 2017-06-13 5666c6a
Fix frustum bug and consider center pixel
