### 2017-06-13 c53bcaa
markers working, alpha and textures are not
