+++ modules/touch/src/TouchMarker.cpp
@@ -36,13 +36,14 @@
 
 namespace {
 	const std::string _loggerCat = "TouchMarker";
+	const int MAX_FINGERS = 20;
 }
 namespace openspace {
 
 TouchMarker::TouchMarker()
 	: properties::PropertyOwner("TouchMarker")
-	, _visible("TouchMarkers visible", "Toggle visibility of markers", false)
-	, _radiusSize("Marker size", "Set marker size in radius", 0.5, 0, 1)
+	, _visible("TouchMarkers visible", "Toggle visibility of markers", true)
+	, _radiusSize("Marker size", "Set marker size in radius", 10, 0, 30)
 	, _texturePath("texturePath", "Color Texture")
 	, _shader(nullptr)
 	, _texture(nullptr)
@@ -57,15 +58,19 @@ TouchMarker::TouchMarker()
 	
 }
 
-bool TouchMarker::initialize(const std::vector<TUIO::TuioCursor> list) {
+bool TouchMarker::initialize() {
 	glGenVertexArrays(1, &_quad); // generate array
 	glGenBuffers(1, &_vertexPositionBuffer); // generate buffer
-	createVertexList(list);
 
-	_shader = OsEng.renderEngine().buildRenderProgram("MarkerProgram",
-		"${MODULE_TOUCH}/shaders/marker_vs.glsl",
-		"${MODULE_TOUCH}/shaders/marker_fs.glsl"
-	);
+	try {
+		_shader = OsEng.renderEngine().buildRenderProgram("MarkerProgram",
+			"${MODULE_TOUCH}/shaders/marker_vs.glsl",
+			"${MODULE_TOUCH}/shaders/marker_fs.glsl"
+		);
+	}
+	catch (const ghoul::opengl::ShaderObject::ShaderCompileError& e) {
+		LERRORC(e.component, e.what());
+	}
 
 	//loadTexture();
 
@@ -91,7 +96,7 @@ bool TouchMarker::deinitialize() {
 }
 
 void TouchMarker::render(const std::vector<TUIO::TuioCursor> list) {
-	if (_visible) {
+	if (_visible && !list.empty()) {
 		createVertexList(list);
 		_shader->activate();
 
@@ -100,7 +105,7 @@ void TouchMarker::render(const std::vector<TUIO::TuioCursor> list) {
 		unit.activate();
 		_texture->bind();
 		_shader->setUniform("texture1", unit);*/
-		//_shader->setUniform("radius", _radiusSize);
+		_shader->setUniform("radius", _radiusSize);
 		
 		glEnable(GL_PROGRAM_POINT_SIZE); // Enable gl_PointSize in vertex shader
 		glPolygonMode(GL_FRONT_AND_BACK, GL_POINT);
@@ -123,45 +128,41 @@ void TouchMarker::update() {
 
 void TouchMarker::loadTexture() {
 	if (_texturePath.value() != "") {
-		std::unique_ptr<ghoul::opengl::Texture> texture =
-			ghoul::io::TextureReader::ref().loadTexture(absPath(_texturePath));
+		_texture = ghoul::io::TextureReader::ref().loadTexture(absPath(_texturePath));
 
-		if (texture) {
+		if (_texture) {
 			LDEBUGC(
 				"TouchMarker",
 				"Loaded texture from '" << absPath(_texturePath) << "'"
 			);
-			texture->uploadTexture();
-
-			// Textures of planets looks much smoother with AnisotropicMipMap rather than linear
[…]
