### 2017-06-13 bfbde46
Separate functionality for fieldline colors and range of interest. (depending on extra variables)
