+++ modules/fieldlinessequence/rendering/renderablefieldlinessequence.cpp
@@ -128,6 +128,7 @@ RenderableFieldlinesSequence::RenderableFieldlinesSequence(const ghoul::Dictiona
       _timeMultiplier("fieldlineParticleSpeed", "FL Particle Speed", 20, 1, 1000),
       _colorMethod("fieldlineColorMethod", "Color Method", properties::OptionProperty::DisplayType::Radio),
       _colorizingQuantity("fieldlineColorQuantity", "Quantity", properties::OptionProperty::DisplayType::Dropdown),
+      _domainQuantity("fieldlineDomainQuantity", "Quantity", properties::OptionProperty::DisplayType::Dropdown),
       _colorGroup("Color"),
       _domainGroup("Domain Limits"),
       _particleGroup("Particles"),
@@ -385,7 +386,7 @@ bool RenderableFieldlinesSequence::initialize() {
             return false;
     }
 
-    bool allowUnitColoring = false;
+    bool allowExtraVariables = false;
     size_t numExtraVariables = 0;
     if (_numberOfStates > 0) {
         // Approximate the end time of last state (and for the sequence as a whole)
@@ -473,7 +474,8 @@ bool RenderableFieldlinesSequence::initialize() {
 
         numExtraVariables = _states[0]._extraVariables.size();
         if (numExtraVariables > 0) {
-            allowUnitColoring = (_states[0]._extraVariables[0].size() > 0) ? true : false;
+            allowExtraVariables = (_states[0]._extraVariables[0].size() > 0) ? true : false;
+            _hasExtraVariables = true;
         }
 
     } else {
@@ -584,14 +586,15 @@ bool RenderableFieldlinesSequence::initialize() {
         addProperty(_isMorphing);
     }
 
-    if (allowUnitColoring) {
+    if (allowExtraVariables) {
         _colorMethod.addOption(colorMethod::UNIFORM, "Uniform Color");
         _colorMethod.addOption(colorMethod::QUANTITY_DEPENDENT, "Quantity Dependent");
         _colorMethod = QUANTITY_DEPENDENT;
 
-        // ASSUMING ALL STATES HAVE THE SAME COLOR VARIABLES
+        // ASSUMING ALL STATES HAVE THE SAME EXTRA VARIABLES
         for (size_t i = 0; i < _states[0]._extraVariableNames.size(); ++i) {
             _colorizingQuantity.addOption(i, _states[0]._extraVariableNames[i]);
+            _domainQuantity.addOption(i, _states[0]._extraVariableNames[i]);
         }
 
         // Set tranferfunction min/max to min/max in the given range
@@ -693,11 +696,16 @@ bool RenderableFieldlinesSequence::initialize() {
             _updateColorBuffer = true;
             _transferFunctionMinVal = std::to_string(_transferFunctionLimits[_colorizingQuantity].x);
             _transferFunctionMaxVal = std::to_string(_transferFunctionLimits[_colorizingQuantity].y);
-            _unitInterestRange = _tFInterestRange[_colorizingQuantity];
-            _unitInterestRange.setMinValue(glm::vec3(_tFInterestRangeLimits[_colorizingQuantity].x));
-            _unitInterestRange.setMaxValue(glm::vec3(_tFInterestRangeLimits[_colorizingQuantity].y));
             _activeColorTable = &_colorTablePaths[_colorizingQuantity];
             _transferFunctionPath = *_activeColorTable;
+        });
+
+        _domainQuantity.onChange([this] {
+            LDEBUG("CHANGED DOMAIN QUANTITY");
+            _updateDomainBuffer = true;
+            _unitInterestRange = _tFInterestRange[_domainQuantity];
[…]
