### 2017-06-13 674ca77
Add a float property (range [0,1]) which scales the alpha channel of the fieldline
