### 2017-06-14 cb151fc
More work on point globe
