+++ modules/globebrowsing/globes/pointglobe.cpp
@@ -38,7 +38,18 @@ namespace globebrowsing {
 
 PointGlobe::PointGlobe(const RenderableGlobe& owner)
     : _owner(owner)
-{}
+    , _intensityClamp(
+        "intensityClamp",
+        "Intensity clamp",
+        1, 0, 1)
+    , _lightIntensity(
+        "lightIntensity",
+        "Light intensity",
+        1, 0, 50)
+{
+    addProperty(_intensityClamp);
+    addProperty(_lightIntensity);
+}
 
 PointGlobe::~PointGlobe() {
     glDeleteBuffers(1, &_vertexBufferID);
@@ -57,20 +68,26 @@ bool PointGlobe::initialize() {
     glBindVertexArray(_vaoID);
 
     // Vertex data is only one point in the origin
-    glm::vec3 data = glm::vec3(0,0,0);
+    std::vector<glm::vec2> data;
+    data.push_back(glm::vec2(-1.0f, -1.0f));
+    data.push_back(glm::vec2(1.0f, -1.0f));
+    data.push_back(glm::vec2(-1.0f, 1.0f));
+    data.push_back(glm::vec2(-1.0f, 1.0f));
+    data.push_back(glm::vec2(1.0f, -1.0f));
+    data.push_back(glm::vec2(1.0f, 1.0f));
 
     // Vertex buffer
     glBindBuffer(GL_ARRAY_BUFFER, _vertexBufferID);
     glBufferData(
         GL_ARRAY_BUFFER,
-        sizeof(glm::vec3),
-        &data,
+        sizeof(glm::vec2) * data.size(),
+        &data[0],
         GL_STATIC_DRAW
     );
 
     // Position at location 0
     glEnableVertexAttribArray(0);
-    glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(glm::vec3), 0);
+    glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE, sizeof(glm::vec2), 0);
 
     glBindVertexArray(0);
 
@@ -91,32 +108,55 @@ void PointGlobe::render(const RenderData& data) {
     // Calculate variables to be used as uniform variables in shader
     glm::dvec3 bodyPosition = data.modelTransform.translation;
 
+    glm::dmat4 rotationTransform = glm::lookAt(
+        glm::dvec3(0.0f),
+        data.camera.positionVec3() - bodyPosition,
+        glm::normalize(glm::dvec3(1000000.0f) - bodyPosition));
+  
+    glm::dvec3 camToBody = bodyPosition - data.camera.positionVec3();
+    float distanceToBody = glm::length(camToBody);
+
+    float avgRadius = _owner.ellipsoid().averageRadius();
+    float lightIntensity = _lightIntensity.value() * data.modelTransform.scale * avgRadius / distanceToBody;
+    float lightIntensityClamped = glm::min(lightIntensity, _intensityClamp.value());
+    float lightOverflow = glm::max(lightIntensity - lightIntensityClamped, 0.0f);
+
+    float billboardRadius = lightIntensityClamped * distanceToBody;
+    glm::dmat4 scaleTransform = glm::scale(glm::dmat4(1.0), glm::dvec3(billboardRadius));
+  
+    setBoundingSphere(billboardRadius);
+
     // Model transform and view transform needs to be in double precision
     glm::dmat4 modelTransform =
         glm::translate(glm::dmat4(1.0), bodyPosition) * // Translation
-        glm::scale(glm::dmat4(1.0), glm::dvec3(data.modelTransform.scale)); // Scale
+        glm::inverse(rotationTransform) * 
[…]
