### 2017-06-14 1802c44
Fixed bug causing camer info vector reader to crash
