### 2017-06-14 5742e09
Removed unused files and started cleanup of atm module for future merging with master.
