+++ modules/atmosphere/rendering/atmosphereraycaster.cpp
@@ -1,131 +1,0 @@
-/*****************************************************************************************
- *                                                                                       *
- * OpenSpace                                                                             *
- *                                                                                       *
- * Copyright (c) 2014-2016                                                               *
- *                                                                                       *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
- * software and associated documentation files (the "Software"), to deal in the Software *
- * without restriction, including without limitation the rights to use, copy, modify,    *
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
- * permit persons to whom the Software is furnished to do so, subject to the following   *
- * conditions:                                                                           *
- *                                                                                       *
- * The above copyright notice and this permission notice shall be included in all copies *
- * or substantial portions of the Software.                                              *
- *                                                                                       *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
- * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
- ****************************************************************************************/
-
-#include <modules/atmosphere/rendering/atmosphereraycaster.h>
-
-#include <glm/glm.hpp>
-#include <ghoul/opengl/ghoul_gl.h>
-#include <sstream>
-#include <ghoul/opengl/programobject.h>
-#include <openspace/util/powerscaledcoordinate.h>
-#include <openspace/util/updatestructures.h>
-#include <openspace/rendering/renderable.h>
-
-namespace {
-    const std::string GlslRaycastPath = "${MODULES}/atmosphere/shaders/raycast.glsl";
-    const std::string GlslBoundsVsPath = "${MODULES}/atmosphere/shaders/boundsvs.glsl";
-    const std::string GlslBoundsFsPath = "${MODULES}/atmosphere/shaders/boundsfs.glsl";
-}
-
-namespace openspace {
-
-AtmosphereRaycaster::AtmosphereRaycaster(glm::vec4 color)
-    : _boundingBox(glm::vec3(1.0))
-    , _color(color) {}
-    
-AtmosphereRaycaster::~AtmosphereRaycaster() {}
-
-void AtmosphereRaycaster::initialize() {
-    _boundingBox.initialize();
-}
-    
-void AtmosphereRaycaster::deinitialize() {
-}
-    
-void AtmosphereRaycaster::renderEntryPoints(const RenderData& data, ghoul::opengl::ProgramObject& program) {
-    program.setUniform("modelTransform", _modelTransform);
-    program.setUniform("viewProjection", data.camera.viewProjectionMatrix());
-    Renderable::setPscUniforms(program, data.camera, data.position);
-    
-    // Cull back face
-    glEnable(GL_CULL_FACE);
-    glCullFace(GL_BACK);
-    
[…]

[…]
