### 2017-06-14 b1be00a
added transparency, thickness and color control to the markers
