### 2017-06-14 cb6d74e
Save metadata as text files to speed up preprocessing a lot
