+++ modules/solarbrowsing/rendering/renderablesolarimagery.cpp
@@ -44,6 +44,8 @@
 
 #include <openspace/util/spicemanager.h>
 
+#include <fstream>
+
 #include <chrono>
 #include <math.h>
 
@@ -133,7 +135,10 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         }
     }
 
-    SpacecraftImageryManager::ref().loadImageMetadata(rootPath, _imageMetadataMap2, _instrumentFilter);
+
+    //SpacecraftImageryManager::ref().loadImageMetadata(rootPath, _imageMetadataMap2, _instrumentFilter);
+    //saveMetadata(rootPath);
+    loadMetadata(rootPath);
 
     // Add GUI names
     unsigned int i = 0;
@@ -144,11 +149,11 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     _currentActiveInstrument
           = _activeInstruments.getDescriptionByValue(_activeInstruments.value());
 
-    std::string tfRootPath;
-    if (dictionary.getValue("TransferfunctionPath", tfRootPath)) {
-        SpacecraftImageryManager::ref().loadTransferFunctions(tfRootPath, _tfMap, _instrumentFilter);
+    SpacecraftImageryManager::ref().loadTransferFunctions(rootPath + "/colortables", _tfMap, _instrumentFilter);
+    //std::string tfRootPath;
+    //if (dictionary.getValue("TransferfunctionPath", tfRootPath)) {
         //throw ghoul::RuntimeError("TransferfunctionPath has to be specified");
-    }
+    //}
 
     // Some sanity checks
     if (_imageMetadataMap2.size() == 0) {
@@ -274,11 +279,11 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     // }
 
 
+
+
     // _currentActiveImageTime
     //       = getDecodeDataFromOsTime(OsEng.timeManager().time().j2000Seconds())
     //               .timeObserved;
-
-   // _imageMetadataMap2[_currentActiveInstrument].displayStateTimes();
 
    // LDEBUG("Init current active image time " <<  SpiceManager::ref().dateFromEphemerisTime(OsEng.timeManager().time().j2000Seconds()));
     //LDEBUG("Init current active image time " << SpiceManager::ref().dateFromEphemerisTime(_currentActiveImageTime));
@@ -313,6 +318,119 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     addProperty(_target);
     addProperty(_moveFactor);
     addProperty(_verboseMode);
+}
+
+// MUST do this conversion before passing in the spice manager again - WTF.
+std::string RenderableSolarImagery::ISO8601(std::string& datetime) {
+    std::string month = datetime.substr(5, 3);
+
+    std::string MM = "";
+    if (month == "JAN") MM = "01";
+    else if (month == "FEB") MM = "02";
+    else if (month == "MAR") MM = "03";
+    else if (month == "APR") MM = "04";
+    else if (month == "MAY") MM = "05";
+    else if (month == "JUN") MM = "06";
+    else if (month == "JUL") MM = "07";
+    else if (month == "AUG") MM = "08";
+    else if (month == "SEP") MM = "09";
+    else if (month == "OCT") MM = "10";
+    else if (month == "NOV") MM = "11";
+    else if (month == "DEC") MM = "12";
+    else ghoul_assert(false, "Bad month");
+
+    datetime.replace(4, 5, "-" + MM + "-");
+    return datetime;
+}
+
[…]
