### 2017-06-14 955f511
[Server] Use correct address and port method in logging
