### 2017-06-15 3b0dcfc
Add virtual properties
