### 2017-06-15 be47939
Fix bug in initializing plane while not within the frustum of the camera
