### 2017-06-15 febb526
Switch to kakadu decoder
