### 2017-06-15 04d0685
Add functionality in stream buffer to be able to process decode queue on multiple threads
