### 2017-06-15 77015e4
Added option to change rendering mode for ATM (temporarilly for tests). Fixed CMakelists in ATM module.
