+++ src/rendering/framebufferrenderer.cpp
@@ -45,6 +45,9 @@
 
 #include <ghoul/opengl/programobject.h>
 
+//#define _OLD_RENDERING_
+#define _NEW_RENDERING_
+
 namespace {
     const std::string _loggerCat = "FramebufferRenderer";
     const std::string ExitFragmentShaderPath = "${SHADERS}/framebuffer/exitframebuffer.frag";
@@ -136,8 +139,12 @@ void FramebufferRenderer::initialize() {
     }
 
     glBindFramebuffer(GL_FRAMEBUFFER, _deferredFramebuffer);
-    //glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D_MULTISAMPLE, _deferredColorTexture, 0);
+#ifdef _OLD_RENDERING_
+    glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D_MULTISAMPLE, _deferredColorTexture, 0);
+#endif
+#ifdef _NEW_RENDERING_
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, _deferredColorTexture, 0);
+#endif
 
     status = glCheckFramebufferStatus(GL_FRAMEBUFFER);
     if (status != GL_FRAMEBUFFER_COMPLETE) {
@@ -274,8 +281,9 @@ void FramebufferRenderer::updateResolution() {
         GLsizei(_resolution.y),
         true);
 
-    // DEBUG: deferred g-buffer
-    /*
+    // G-buffer
+#ifdef _OLD_RENDERING_
+    
     glBindTexture(GL_TEXTURE_2D_MULTISAMPLE, _deferredColorTexture);
 
     glTexImage2DMultisample(
@@ -284,8 +292,9 @@ void FramebufferRenderer::updateResolution() {
         GL_RGBA,
         GLsizei(_resolution.x),
         GLsizei(_resolution.y),
-        true);
-    */
+        true);    
+#endif
+#ifdef _NEW_RENDERING_
     glBindTexture(GL_TEXTURE_2D, _deferredColorTexture);
 
     glTexImage2D(
@@ -301,7 +310,8 @@ void FramebufferRenderer::updateResolution() {
 
     glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
     glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-
+#endif
+#ifdef _OLD_RENDERING_
     glBindTexture(GL_TEXTURE_2D_MULTISAMPLE, _mainOtherDataTexture);
 
     glTexImage2DMultisample(
@@ -311,7 +321,7 @@ void FramebufferRenderer::updateResolution() {
         GLsizei(_resolution.x),
         GLsizei(_resolution.y),
         true);
-
+#endif
     glBindTexture(GL_TEXTURE_2D_MULTISAMPLE, _mainPositionTexture);
 
     glTexImage2DMultisample(
@@ -695,35 +705,21 @@ void FramebufferRenderer::render(float blackoutFactor, bool doPerformanceMeasure
             LWARNING("Deferredcaster is not attached when trying to perform deferred task");
         }
     }
-    /*
-    glBindFramebuffer(GL_FRAMEBUFFER, defaultFbo);
-    _resolveProgram->activate();
-
-    ghoul::opengl::TextureUnit mainColorTextureUnit;
-    mainColorTextureUnit.activate();*/
-
-    // DEBUG: g-buffer
+    
+#ifdef _NEW_RENDERING_    
[…]
