### 2017-06-15 589e85c
Fixed CMakelists in ATM module.
