### 2017-06-15 02770ce
Refactor transfer function gradient calculation. Possible breaking change.
