### 2017-06-15 5deea2c
Fix pointer/addressing mistake for getTfGradient
