+++ modules/multiresvolume/rendering/tfbrickselector.cpp
@@ -289,8 +289,8 @@ float TfBrickSelector::splitPoints(unsigned int brickIndex, BrickSelection::Spli
 
 bool TfBrickSelector::calculateBrickErrors() {
 
-    std::vector<float> * gradients = getTfGradients();
-    if (!gradients) return false;
+    std::vector<float> gradients = getTfGradients();
+    if (!gradients.size()) return false;
 
     unsigned int nHistograms = _tsp->numTotalNodes();
     _brickErrors = std::vector<float>(nHistograms);
@@ -301,12 +301,12 @@ bool TfBrickSelector::calculateBrickErrors() {
         } else {
             const Histogram* histogram = _histogramManager->getHistogram(brickIndex);
             float error = 0;
-            for (int i = 0; i < gradients->size(); i++) {
+            for (int i = 0; i < gradients.size(); i++) {
                 float x = (i + 0.5) / _transferFunction->width();
                 float sample = histogram->interpolate(x);
                 assert(sample >= 0);
-                assert(gradients->at(i) >= 0);
-                error += sample * gradients->at(i);
+                assert(gradients[i] >= 0);
+                error += sample * gradients[i];
             }
             _brickErrors[brickIndex] = error;
         }
