### 2017-06-15 7103220
Prep to simplify LocalTfBrickSelector's calculateBrickErrors. Add TSP::NodeType enum, extend LocalTfBrickSelector::Error struct, create LocalErrorHistogramManager getter using NodeType
