### 2017-06-15 02ae57e
Refactoring repetetive code in calculateBrickErrors
