### 2017-06-16 d0375b6
Add function for clearing buffer
