### 2017-06-16 adb6724
Add simple test loop
