+++ modules/solarbrowsing/rendering/renderablesolarimageryprojection.h
@@ -69,6 +69,7 @@ private:
     std::string _nodeName;
     std::string path;
     std::unique_ptr<ghoul::opengl::Texture> _magnetogramTexture;
+    properties::BoolProperty _startLoop1;
     float _planetRadius;
 };
 
