### 2017-06-16 a5c0932
Increase maximum number of spacecraft observatory
