+++ modules/solarbrowsing/rendering/renderablesolarimageryprojection.cpp
@@ -194,7 +194,7 @@ void RenderableSolarImageryProjection::render(const RenderData& data) {
     );
 
     const int numPlanes = _solarImageryDependencies.size();
-    const int MAX_SPACECRAFT_OBSERVATORY = 6;
+    const int MAX_SPACECRAFT_OBSERVATORY = 12;
     int solarImageryCount = 0;
 
     ghoul::opengl::TextureUnit txUnits[MAX_SPACECRAFT_OBSERVATORY];
