### 2017-06-16 7984053
Include radial level of detail for mesh generation
