+++ modules/globebrowsing/tasks/collapsedmeshgeneration.cpp
@@ -27,6 +27,9 @@
 
 #include <modules/globebrowsing/tasks/imgreader.h>
 #include <modules/globebrowsing/tasks/meshwriter.h>
+#include <modules/globebrowsing/tasks/pointcloudfilter.h>
+#include <modules/globebrowsing/tasks/randompointcloudfilter.h>
+#include <modules/globebrowsing/tasks/levelofdetail.h>
 
 #include <ghoul/glm.h>
 #include <glm/gtx/quaternion.hpp>
@@ -42,6 +45,10 @@
 #include <pcl/surface/poisson.h>
 #include <pcl/surface/mls.h>
 #include <pcl/surface/simplification_remove_unused_vertices.h>
+#include <pcl/filters/statistical_outlier_removal.h>
+#include <pcl/features/normal_3d_omp.h>
+#include <pcl/filters/passthrough.h>
+#include <pcl/filters/extract_indices.h>
 
 #include <fstream>
 
@@ -52,76 +59,113 @@ namespace {
 namespace openspace {
 	namespace globebrowsing {
 		void CollapsedMeshGeneration::generateMeshFromBinary(ghoul::Dictionary dictionary) {
-			//const std::string binary_path, std::string output_path
+			//const std::string pathToBinaryFile, std::string output_path
 
-			std::string file_name_stripped2;
-			std::string file_name_stripped_obj2;
-			std::string output_path2;
+			std::string outputFile;
+			dictionary.getValue("OutputFile", outputFile);
+			outputFile = "obj";
+			std::string pathToModelsFolder;
+			dictionary.getValue("OutputPath", pathToModelsFolder);
 			ImgReader::PointCloudInfo mInfo2;
+			std::string pathToBinaryFolder;
+			dictionary.getValue("BinaryPath", pathToBinaryFolder);
+			std::string level;
+			dictionary.getValue("Level", level);
 
+			std::unordered_map<std::string, LevelOfDetail>::const_iterator levelOfDetailVars = _levelOfDetailMap.find(level);
+
+			LevelOfDetail levelOfDetail;
+
+			if (levelOfDetailVars == _levelOfDetailMap.end()) {
+				LINFO("USING LOWEST LEVEL OF DETAIL SINCE NOT SPECIFIED, OR NO EXISTING PREDEFINED FOUND");
+				level = "level1";
+				std::unordered_map<std::string, LevelOfDetail>::const_iterator levelOfDetail1 = _levelOfDetailMap.find(level);
+				levelOfDetail = levelOfDetail1->second;
+			}
+			else {
+				levelOfDetail = levelOfDetailVars->second;
+			}
+
+			ghoul::Dictionary filters;
+			dictionary.getValue("Filters.Filters", filters);
+
+			std::vector<std::unique_ptr<PointCloudFilter>> upSamplingFilters;
+			LINFO("FILTERS USED: ");
+			for (size_t i = 0; i < filters.size(); ++i) {
+				std::string dictKey = std::to_string(i + 1);
+				ghoul::Dictionary filterDic = filters.value<ghoul::Dictionary>(dictKey);
+				std::string filtersUsed;
+				filterDic.getValue("Type", filtersUsed);
+				LINFO(filtersUsed);
+
+				upSamplingFilters.push_back(std::move(PointCloudFilter::createFromDictionary(filterDic)));
+			}
+			
+			std::string pathToDriveFolder;
[…]
