### 2017-06-16 8b861a5
Add offset and start instrument to allow multiple overlapping image planes
