+++ modules/solarbrowsing/rendering/spacecraftcameraplane.h
@@ -40,7 +40,7 @@ public:
                 const float& planeOpacity, const float& contrastValue,
                 const float& gammaValue, const bool& disableBorder,
                 const bool& disableFrustum, const glm::vec2& currentCenterPixel,
-                const float& currentScale);
+                const float& currentScale, const float& multipleImageryOffset);
     void update();
     void createPlaneAndFrustum(const double& moveDistance);
     // TODO(mnoven) : Pre process image, no need to set uniforms
