### 2017-06-16 02de168
Don't perform any texture updates when not enabled
