### 2017-06-16 28f3546
Added the option to change maximum LOD during runtime and changed range of height property
