+++ modules/globebrowsing/models/renderableroversurface.h
@@ -51,6 +51,7 @@ public:
 		properties::BoolProperty useMastCam;
 		properties::BoolProperty enableDepth;
 		properties::FloatProperty heightProp;
+		properties::IntProperty maxLod;
 	};
 
 	RenderableRoverSurface(const ghoul::Dictionary& dictionary);
@@ -65,7 +66,7 @@ public:
 	
 private:
 	std::vector<std::shared_ptr<SubsiteModels>> calculateSurfacePosition(std::vector<std::shared_ptr<SubsiteModels>> vector);
-	void lockSubsite(const int level, std::vector<std::shared_ptr<Subsite>> subsites);
+	void lockSubsite(std::vector<std::shared_ptr<Subsite>> subsites);
 
 	std::vector<std::shared_ptr<Subsite>> _subsites;
 	std::vector<std::shared_ptr<Subsite>> _prevSubsites;
