+++ modules/globebrowsing/models/renderableroversurface.cpp
@@ -56,23 +56,24 @@ namespace openspace {
 using namespace properties;
 
 namespace globebrowsing {
-	RenderableRoverSurface::RenderableRoverSurface(const ghoul::Dictionary & dictionary)
-		: Renderable(dictionary)
-		, _generalProperties({
-				BoolProperty("enable", "Enabled", true),
-				BoolProperty("enablePath", "Enable path", true),
-				BoolProperty("lockSubsite", "Lock subsite", false),
-				BoolProperty("useMastCam", "Show mastcam coloring", false),
-				BoolProperty("enableDepth", "Enable depth", true),
-				FloatProperty("heightProp", "Site height", 1.f, 1.0f, 100.f)
-		})
-		, _debugModelRotation("modelrotation", "Model Rotation", glm::vec3(0.0f), glm::vec3(0.0f), glm::vec3(360.0f))
-		, _modelSwitch()
-		, _prevLevel(3)
-		, _isFirst(true)
-		, _isFirstLow(true)
-		, _isFirstHigh(true)
-		, _renderableSitePropertyOwner("LayersOfThis")
+RenderableRoverSurface::RenderableRoverSurface(const ghoul::Dictionary & dictionary)
+	: Renderable(dictionary)
+	, _generalProperties({
+			BoolProperty("enable", "Enabled", true),
+			BoolProperty("enablePath", "Enable path", true),
+			BoolProperty("lockSubsite", "Lock subsite", false),
+			BoolProperty("useMastCam", "Show mastcam coloring", false),
+			BoolProperty("enableDepth", "Enable depth", true),
+			FloatProperty("heightProp", "Site height", 1.0f, 0.0f, 3.0f),
+			IntProperty("maxLod", "Max LOD",3.0, 1.0, 3.0)
+	})
+	, _debugModelRotation("modelrotation", "Model Rotation", glm::vec3(0.0f), glm::vec3(0.0f), glm::vec3(360.0f))
+	, _modelSwitch()
+	, _prevLevel(3)
+	, _isFirst(true)
+	, _isFirstLow(true)
+	, _isFirstHigh(true)
+	, _renderableSitePropertyOwner("LayersOfThis")
 {
 	if (!dictionary.getValue(keyRoverLocationPath, _roverLocationPath))
 		throw ghoul::RuntimeError(std::string(keyRoverLocationPath) + " must be specified!");
@@ -108,6 +109,7 @@ namespace globebrowsing {
 	addProperty(_generalProperties.useMastCam);
 	addProperty(_generalProperties.enableDepth);
 	addProperty(_generalProperties.heightProp);
+	addProperty(_generalProperties.maxLod);
 	addProperty(_debugModelRotation);
 	
 	_siteManager = std::make_shared<SiteManager>(marsRoverModels, _subsitesWithModels);
@@ -208,7 +210,10 @@ void RenderableRoverSurface::render(const RenderData& data) {
 			break;
 	}
 
-	lockSubsite(level, ss);
+	if (_generalProperties.maxLod.value() < level)
+		level = _generalProperties.maxLod.value();
+
+	lockSubsite(ss);
 
 	std::vector<std::shared_ptr<SubsiteModels>> vectorOfsubsiteModels;
 	if(_generalProperties.lockSubsite.value())
@@ -387,12 +392,12 @@ std::vector<std::shared_ptr<SubsiteModels>> RenderableRoverSurface::calculateSur
 	return vector;
 }
 
-void RenderableRoverSurface::lockSubsite(const int level, std::vector<std::shared_ptr<Subsite>> subsites) {
-	if (_generalProperties.lockSubsite.value() && level == 3 && _pressedOnce == false) {
[…]
