### 2017-06-16 a57dc7b
Clean up
