### 2017-06-16 2b99d1a
Fixed bug introduced with new rendering method for G-Buffer.
