### 2017-06-16 64dabf7
Recalculating path every time the LOD changes
