### 2017-06-16 5ad00f6
Added fine controls for Mie Scattering as requested by Carter and John.
