+++ modules/solarbrowsing/rendering/renderablesolarimageryprojection.cpp
@@ -55,6 +55,7 @@ RenderableSolarImageryProjection::RenderableSolarImageryProjection(
     : Renderable(dictionary)
     , _shader(nullptr)
     , _startLoop1("startLoop1", "Start Loop 1", false)
+    , _startLoop2("startLoop2", "Start Loop 2", false)
     , _sphere(nullptr)
 {
     if (!dictionary.getValue("Name", _nodeName)) {
@@ -73,7 +74,16 @@ RenderableSolarImageryProjection::RenderableSolarImageryProjection(
         }
     });
 
+    _startLoop2.onChange([this]() {
+        if (_startLoop2) {
+            auto& time = OsEng.timeManager().time();
+            time.setTime("2012-07-12T15:00:00.00");
+            time.setDeltaTime(3600);
+        }
+    });
+
     addProperty(_startLoop1);
+    addProperty(_startLoop2);
 }
 
 bool RenderableSolarImageryProjection::initialize() {
@@ -151,7 +161,7 @@ void RenderableSolarImageryProjection::update(const UpdateData& data) {
 
     auto& time = OsEng.timeManager().time();
     if (_startLoop1) {
-        auto endTime = time.convertTime("2012-07-13T02:00:00.00");
+        auto endTime = time.convertTime("2012-07-12T18:00:00.00");
         if (time.j2000Seconds() > endTime) {
             time.setTime("2012-07-12T15:00:00.00");
             for (int i = 0; i < _solarImageryDependencies.size(); ++i) {
@@ -163,6 +173,18 @@ void RenderableSolarImageryProjection::update(const UpdateData& data) {
         }
     }
 
+    else if (_startLoop2) {
+        auto endTime = time.convertTime("2012-07-13T03:00:00.00");
+        if (time.j2000Seconds() > endTime) {
+            time.setTime("2012-07-12T15:00:00.00");
+            for (int i = 0; i < _solarImageryDependencies.size(); ++i) {
+                auto* solarImagery = static_cast<RenderableSolarImagery*>(
+                      _solarImageryDependencies[i]->renderable());
+                solarImagery->clearBuffer();
+            }
+
+        }
+    }
 }
 
 void RenderableSolarImageryProjection::render(const RenderData& data) {
@@ -204,7 +226,10 @@ void RenderableSolarImageryProjection::render(const RenderData& data) {
         auto* solarImagery = static_cast<RenderableSolarImagery*>(
               _solarImageryDependencies[i]->renderable());
 
+      //  if (!solarImagery->isEnabled()) continue;
+
         bool isCoronaGraph = solarImagery->_isCoronaGraph;
+        bool enabled = solarImagery->isEnabled();
 
         const SpacecraftCameraPlane& plane = solarImagery->cameraPlane();
         const glm::dvec3 planePos = plane.worldPosition();
@@ -216,6 +241,7 @@ void RenderableSolarImageryProjection::render(const RenderData& data) {
         // }
 
         _shader->setUniform("isCoronaGraph[" + std::to_string(i) + "]", isCoronaGraph);
+        _shader->setUniform("isEnabled[" + std::to_string(i) + "]", enabled);
 
         //_shader->setUniform("magicPlaneFactor[" + std::to_string(i) + "]", solarImagery->_magicPlaneFactor);
 
@@ -240,7 +266,7 @@ void RenderableSolarImageryProjection::render(const RenderData& data) {
         tfUnits[i].activate();
 
         auto lut = solarImagery->getTransferFunction();
-        if (lut) {
+        if (lut && solarImagery->isEnabled()) {
             lut->bind();
             _shader->setUniform("hasLut[" + std::to_string(i) + "]", true);
         } else {
