### 2017-06-16 4f6e942
Changed height of path and models
