### 2017-06-19 e02f1b8
Provide option to load binary files at runtime. Specified in modfile
