+++ modules/fieldlinessequence/rendering/renderablefieldlinessequence.cpp
@@ -48,6 +48,7 @@
 #include <iostream>
 #include <iterator>
 #include <sstream>
+#include <thread>
 
 #include <chrono>
 
@@ -57,6 +58,8 @@ namespace {
     const char* keyTracingMethodPreTracedBinary = "PreTracedBinary";
     const char* keyTracingMethodPreProcess      = "PreProcess";
     const char* keyTracingMethodLiveTrace       = "LiveTrace";
+
+    const char* keyLoadBinariesAtRuntime        = "LoadBinariesAtRuntime";
 
     const char* keySourceFolder                 = "SourceFolder";
     const char* keyStartStateOffset             = "StartStateOffset";
@@ -169,7 +172,7 @@ bool RenderableFieldlinesSequence::isReady() const {
 bool RenderableFieldlinesSequence::initialize() {
     //--Extract common  info from LUA, applicable to all types of renderable fieldline sequences--//
     std::string sourceFileExt = "";
-    std::vector<std::string> validSourceFilePaths;
+    // std::vector<std::string> validSourceFilePaths;
     TracingMethod tracingMethod;
 
     // Determine what type of tracing method to use.
@@ -196,11 +199,14 @@ bool RenderableFieldlinesSequence::initialize() {
         }
     }
 
-    if (!getSourceFilesFromDictionary(sourceFileExt, validSourceFilePaths)) {
+    if (!getSourceFilesFromDictionary(sourceFileExt, _validSourceFilePaths)) {
         return false;
     }
 
-    _numberOfStates = validSourceFilePaths.size();
+    _numberOfStates = _validSourceFilePaths.size();
+    LDEBUG("Found " << _numberOfStates << " valid files!");
+
+    // TODO this makes no sense if binaries are loaded at runtime!
     _states.reserve(_numberOfStates);
     _startTimes.reserve(_numberOfStates);
 
@@ -301,11 +307,11 @@ bool RenderableFieldlinesSequence::initialize() {
             // TODO Only one loop.. Not one for each tracing method!!
             for (size_t i = 0; i < _numberOfStates; ++i) {
                 std::unique_ptr<ccmc::Kameleon> kameleon =
-                        fsManager.createKameleonObject(validSourceFilePaths[i]);
+                        fsManager.createKameleonObject(_validSourceFilePaths[i]);
                 if (kameleon == nullptr) {
                     continue;
                 }
-                LDEBUG(validSourceFilePaths[i] << " is now being traced.");
+                LDEBUG(_validSourceFilePaths[i] << " is now being traced.");
                 _states.push_back(FieldlinesState(_seedPoints.size()));
                 fsManager.getFieldlinesState(kameleon.get(),
                                              tracingVariable,
@@ -345,38 +351,80 @@ bool RenderableFieldlinesSequence::initialize() {
             numResamples = 0;       // TODO: implement morphing
 
             for (size_t i = 0; i < _numberOfStates; ++i) {
-                LDEBUG(validSourceFilePaths[i] << " is now being processed.");
+                LDEBUG(_validSourceFilePaths[i] << " is now being processed.");
                 FieldlinesState tmpState;
                 _states.push_back(tmpState);
-                fsManager.getFieldlinesState(validSourceFilePaths[i],
-                                                                   _isMorphing,
-                                                                   numResamples,
-                                                                   resamplingOption,
-                                                                   _states[i]);
+                fsManager.getFieldlinesState(_validSourceFilePaths[i],
[…]
