### 2017-06-19 86eeddd
Update temporal tile provider to be able to take gdal descriptions without file path.
