### 2017-06-19 f73d152
Add adjustment property to layers.
