### 2017-06-19 875a83a
Add more looping time periods
