+++ modules/solarbrowsing/rendering/renderablesolarimageryprojection.h
@@ -69,8 +69,8 @@ private:
     std::string _nodeName;
     std::string path;
     std::unique_ptr<ghoul::opengl::Texture> _magnetogramTexture;
-    properties::BoolProperty _startLoop1;
-    properties::BoolProperty _startLoop2;
+    properties::BoolProperty _activateLooping;
+    properties::IntProperty _loopId;
     float _planetRadius;
 };
 
