+++ modules/solarbrowsing/rendering/renderablesolarimageryprojection.cpp
@@ -47,6 +47,15 @@ using namespace ghoul::opengl;
 namespace {
     static const std::string _loggerCat = "RendearbleSpacecraftCameraSphere";
     const char* keyRadius = "Radius";
+    const std::vector<std::pair<int, std::pair<std::string, std::string>>> loopTimes
+          = {
+             {3600, {"2012-07-12T15:00:00.00", "2012-07-12T18:00:00.00"}},
+             {3600, {"2012-07-12T15:00:00.00", "2012-07-13T03:00:00.00"}},
+             {21600, {"2012-07-08T00:00:00.00", "2012-07-13T00:00:00.00"}},
+             {7200, {"2012-07-17T12:45:00.00", "2012-07-19T17:00:00.00"}},
+             {7200, {"2012-07-17T06:00:00.00", "2012-07-19T17:00:00.00"}},
+             {7200, {"2012-07-23T00:00:00.00", "2012-07-23T16:00:00.00"}}
+            };
 }
 
 namespace openspace {
@@ -54,8 +63,8 @@ RenderableSolarImageryProjection::RenderableSolarImageryProjection(
       const ghoul::Dictionary& dictionary)
     : Renderable(dictionary)
     , _shader(nullptr)
-    , _startLoop1("startLoop1", "Start Loop 1", false)
-    , _startLoop2("startLoop2", "Start Loop 2", false)
+    , _loopId("loopId", "Loop Id", 0, 0, 5)
+    , _activateLooping("activateLooping", "Activate Looping", false)
     , _sphere(nullptr)
 {
     if (!dictionary.getValue("Name", _nodeName)) {
@@ -66,24 +75,35 @@ RenderableSolarImageryProjection::RenderableSolarImageryProjection(
         throw ghoul::RuntimeError("HMIPath has to be specified");
     }
 
-    _startLoop1.onChange([this]() {
-        if (_startLoop1) {
+    _activateLooping.onChange([this]() {
+        if (_activateLooping) {
             auto& time = OsEng.timeManager().time();
-            time.setTime("2012-07-12T15:00:00.00");
-            time.setDeltaTime(3600);
+            const auto& timePair = loopTimes[_loopId.value()];
+            time.setTime(timePair.second.first);
+            time.setDeltaTime(timePair.first);
         }
     });
 
-    _startLoop2.onChange([this]() {
-        if (_startLoop2) {
-            auto& time = OsEng.timeManager().time();
-            time.setTime("2012-07-12T15:00:00.00");
-            time.setDeltaTime(3600);
-        }
-    });
+    // _startLoop1.onChange([this]() {
+    //     if (_startLoop1) {
+    //         auto& time = OsEng.timeManager().time();
+    //         time.setTime("2012-07-12T15:00:00.00");
+    //         time.setDeltaTime(3600);
+    //     }
+    // });
 
-    addProperty(_startLoop1);
-    addProperty(_startLoop2);
+    // _startLoop2.onChange([this]() {
+    //     if (_startLoop2) {
+    //         auto& time = OsEng.timeManager().time();
+    //         time.setTime("2012-07-12T15:00:00.00");
+    //         time.setDeltaTime(3600);
+    //     }
+    // });
+
+    addProperty(_activateLooping);
+    addProperty(_loopId);
+    //addProperty(_startLoop1);
+    //addProperty(_startLoop2);
 }
 
 bool RenderableSolarImageryProjection::initialize() {
@@ -159,32 +179,46 @@ void RenderableSolarImageryProjection::update(const UpdateData& data) {
         _shader->rebuildFromFile();
     }
 
-    auto& time = OsEng.timeManager().time();
-    if (_startLoop1) {
[…]
