### 2017-06-19 8c0eef5
Add slightly more generalized functionality when reading states from json
