### 2017-06-20 583b432
Added proper CPACK structure and added generation of openspace.h header to add verisoning information in CMake.
