### 2017-06-20 f5989fa
Make buffer consists of a fix set of PBOs which are decoded to directly
