+++ modules/solarbrowsing/util/pixelbufferobject.h
@@ -42,15 +42,16 @@ public:
      * Otherwise glMapBuffer will stall for GPU to finish its job.
      */
     template<typename T>
-    T* mapToClientMemory(bool shouldOrphanData);
+    T* mapToClientMemory(bool shouldOrphanData, int size);
     /**
      * Size can be set dynamically to be able to get a smaller block than the
      * full size PBO when getting the mapped pointer/
      */
-    void setSize(const unsigned int& size);
+    //void setSize(const unsigned int& size);
     void releaseMappedBuffer();
     void activate();
     void deactivate();
+    const GLuint& id();
 
 private:
     void initialize();
