+++ modules/solarbrowsing/rendering/renderablesolarimagery.cpp
@@ -49,7 +49,7 @@
 #include <chrono>
 #include <math.h>
 
-#define BUFFER_MAX_SIZE 100
+#define BUFFER_SIZE 5
 
 using namespace ghoul::opengl;
 using namespace std::chrono;
@@ -195,7 +195,11 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     _imageSize = imStart->fullResolution / (pow(2, _resolutionLevel));
 
     // Always give PBO maximum size
-    _pbo = std::make_unique<PixelBufferObject>(4096 * 4096 * sizeof(IMG_PRECISION));
+    //_pbo = std::make_unique<PixelBufferObject>(4096 * 4096 * sizeof(IMG_PRECISION));
+
+    for (int i = 0; i < BUFFER_SIZE; ++i) {
+        _pbos[i] = std::make_unique<PixelBufferObject>(4096 * 4096 * sizeof(IMG_PRECISION));
+    }
 
     // TODO(mnoven): Faster to send GL_RGBA32F as internal format - GPU Pads anyways?
     _texture =  std::make_unique<Texture>(
@@ -236,7 +240,7 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
         std::shared_ptr<ImageMetadata> im = state.contents();
         //_imageSize = _imageMetadataMap2[_currentActiveInstrument]. //_fullResolution / (pow(2, _resolutionLevel));
         _imageSize = im->fullResolution / (pow(2, _resolutionLevel));
-        _pbo->setSize(_imageSize * _imageSize * sizeof(IMG_PRECISION));
+        //_pbo->setSize(_imageSize * _imageSize * sizeof(IMG_PRECISION));
 
         // Upload asap
         updateTextureGPU(/*asyncUpload=*/false);
@@ -278,7 +282,7 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
 
         //_imageSize = _fullResolution / (pow(2, _resolutionLevel));
 
-        _pbo->setSize(_imageSize * _imageSize * sizeof(IMG_PRECISION));
+        //_pbo->setSize(_imageSize * _imageSize * sizeof(IMG_PRECISION));
         updateTextureGPU(/*asyncUpload=*/false, /*resChanged=*/true);
         if (_useBuffering) {
 
@@ -337,7 +341,7 @@ RenderableSolarImagery::RenderableSolarImagery(const ghoul::Dictionary& dictiona
     addProperty(_sharpenValue);
     addProperty(_gammaValue);
     addProperty(_contrastValue);
-    addProperty(_bufferSize);
+    //addProperty(_bufferSize);
     addProperty(_displayTimers);
     addProperty(_planeSize);
     addProperty(_resolutionLevel);
@@ -374,9 +378,15 @@ std::string RenderableSolarImagery::ISO8601(std::string& datetime) {
 }
 
 void RenderableSolarImagery::clearBuffer() {
+    _pboIsDirty = false;
     _streamBuffer.clear();
     _frameSkipCount = 0;
     _bufferCountOffset = 1;
+    _busyPbos.clear();
+
+    while (!_pboQueue.empty()) {
+        _pboQueue.pop();
+    }
 }
 
 void RenderableSolarImagery::loadMetadata(const std::string& rootPath) {
@@ -412,16 +422,12 @@ void RenderableSolarImagery::loadMetadata(const std::string& rootPath) {
                         ImageMetadata im;
 
                         myfile >> std::ws; // skip the rest of the line
-                        //LDEBUG("HEJ");
                         std::string date;
-                        //myfile >> date;
                         std::getline(myfile, date);
[…]
