### 2017-06-20 e731abe
Fixed compiler warnings
