### 2017-06-20 fd7a8f6
Packaging(ZIP on Windows, TGZ on UNIX) and installer creation(Windows Only via NSIS) is now working.
