+++ support/cmake/packaging.cmake
@@ -22,7 +22,7 @@
 # OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         #
 #########################################################################################
 
-set(CPACK_MONOLITHIC_INSTALL FALSE)
+set(CPACK_MONOLITHIC_INSTALL TRUE)
 include(InstallRequiredSystemLibraries)
 set(CPACK_PACKAGE_NAME "OpenSpace")
 set(CPACK_PACKAGE_DESCRIPTION_FILE "${OPENSPACE_BASE_DIR}/README.md")
@@ -30,50 +30,70 @@ set(CPACK_RESOURCE_FILE_LICENSE "${OPENSPACE_BASE_DIR}/LICENSE.md")
 set(CPACK_PACKAGE_VERSION_MAJOR "${OPENSPACE_VERSION_MAJOR}")
 set(CPACK_PACKAGE_VERSION_MINOR "${OPENSPACE_VERSION_MINOR}")
 set(CPACK_PACKAGE_VERSION_PATCH "${OPENSPACE_VERSION_PATCH}")
-set(OPENSPACE_VERSION "${OPENSPACE_VERSION_MAJOR}.${OPENSPACE_VERSION_MINOR}.${OPENSPACE_VERSION_PATCH} ${OPENSPACE_VERSION_STRING}")
+set(OPENSPACE_VERSION_NUMBER "${OPENSPACE_VERSION_MAJOR}.${OPENSPACE_VERSION_MINOR}.${OPENSPACE_VERSION_PATCH}")
 set(CPACK_PACKAGE_INSTALL_DIRECTORY "OpenSpace ${OPENSPACE_VERSION}")
-
-if(WIN32)
-	# Need backslash for correct subdirectory paths with NSIS
-	if(CMAKE_SIZEOF_VOID_P EQUAL 8 )
-		set(CPACK_PACKAGE_FILE_NAME         "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION} 64bit")
-		set(CPACK_PACKAGE_INSTALL_DIRECTORY "${CPACK_PACKAGE_NAME}\\\\${OPENSPACE_VERSION}\\\\64bit")
-		set(CPACK_NSIS_DISPLAY_NAME         "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION} 64-bit")
-	else()
-		set(CPACK_PACKAGE_FILE_NAME         "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION} 32bit")
-		set(CPACK_PACKAGE_INSTALL_DIRECTORY "${CPACK_PACKAGE_NAME}\\\\${OPENSPACE_VERSION}\\\\32bit")
-		set(CPACK_NSIS_DISPLAY_NAME         "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION} 32-bit")
-	endif()
-else()
-	set(CPACK_PACKAGE_FILE_NAME         "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION}")
-	set(CPACK_PACKAGE_INSTALL_DIRECTORY "CPACK_PACKAGE_NAME/${OPENSPACE_VERSION}")
-endif()
-
-set(CPACK_PACKAGE_EXECUTABLES "openspace;OpenSpace")
-set(CPACK_GENERATOR ZIP)
+set(CPACK_PACKAGE_ICON "${OPENSPACE_BASE_DIR}\\\\apps\\\\OpenSpace\\\\openspace.png")
[…]
