### 2017-06-20 f021b1c
Changed from using a static scenegraph node to a dynamic
