+++ modules/globebrowsing/models/multimodelprovider.cpp
@@ -39,14 +39,15 @@ MultiModelProvider::MultiModelProvider(const ghoul::Dictionary& dictionary)
 	: ModelProvider(dictionary) {
 }
 
-std::vector<std::shared_ptr<Subsite>> MultiModelProvider::calculate(const std::vector<std::vector<std::shared_ptr<Subsite>>> subsites, const RenderData& data) {
+std::vector<std::shared_ptr<Subsite>> MultiModelProvider::calculate(const std::vector<std::vector<std::shared_ptr<Subsite>>> subsites,
+	const RenderData& data, const SceneGraphNode* parent) {
+
 	std::vector<std::shared_ptr<Subsite>> subsitesInsideRadius;
 	float radius = 20;
 
-	SceneGraphNode* _parent = OsEng.renderEngine().scene()->sceneGraphNode("Mars");
-	RenderableGlobe* rg = (RenderableGlobe*)_parent->renderable();
+	RenderableGlobe* rg = (RenderableGlobe*)parent->renderable();
 
-	glm::dvec3 center = _parent->worldPosition();
+	glm::dvec3 center = parent->worldPosition();
 	glm::dmat4 globeModelTransform = rg->modelTransform();
 	glm::dmat4 globeModelInverseTransform = rg->inverseModelTransform();
 	glm::dvec3 cameraPos = data.camera.positionVec3();
@@ -68,5 +69,6 @@ std::vector<std::shared_ptr<Subsite>> MultiModelProvider::calculate(const std::v
 bool MultiModelProvider::initialize() {
 	return true;
 }
-}
-}
+
+} // namespace globebrowsing
+} // namespace openspace
