+++ modules/globebrowsing/models/singlemodelprovider.h
@@ -28,6 +28,7 @@
 
 #include <modules/globebrowsing/models/modelprovider.h>
 #include <modules/globebrowsing/models/renderableroversurface.h>
+#include <openspace/scene/scenegraphnode.h>
 
 namespace openspace {
 namespace globebrowsing {
@@ -35,7 +36,8 @@ class SingleModelProvider : public ModelProvider {
 public:
 	SingleModelProvider(const ghoul::Dictionary& dictionary);
 
-	virtual std::vector<std::shared_ptr<Subsite>> calculate(const std::vector<std::vector<std::shared_ptr<Subsite>>> subsites, const RenderData& data);
+	virtual std::vector<std::shared_ptr<Subsite>> calculate(const std::vector<std::vector<std::shared_ptr<Subsite>>> subsites,
+		const RenderData& data, const SceneGraphNode* parent);
 
 	virtual bool initialize();
 };
