### 2017-06-20 6af9162
Added a background exposure control for atm scenes.
