### 2017-06-20 c39c07d
Fixes for packing an archive on OSX.
