+++ support/cmake/packaging.cmake
@@ -32,11 +32,10 @@ set(CPACK_PACKAGE_VERSION_MINOR "${OPENSPACE_VERSION_MINOR}")
 set(CPACK_PACKAGE_VERSION_PATCH "${OPENSPACE_VERSION_PATCH}")
 set(OPENSPACE_VERSION_NUMBER "${OPENSPACE_VERSION_MAJOR}.${OPENSPACE_VERSION_MINOR}.${OPENSPACE_VERSION_PATCH}")
 set(CPACK_PACKAGE_INSTALL_DIRECTORY "OpenSpace ${OPENSPACE_VERSION}")
-set(CPACK_PACKAGE_ICON "${OPENSPACE_BASE_DIR}\\\\apps\\\\OpenSpace\\\\openspace.png")
 set(CPACK_PACKAGE_FILE_NAME "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION_NUMBER} ${OPENSPACE_VERSION_STRING}")
 set(CPACK_STRIP_FILES 1)
 
-install(DIRECTORY ${OPENSPACE_BASE_DIR}/bin/openspace/${CMAKE_BUILD_TYPE}/ DESTINATION bin)
+install(DIRECTORY ${OPENSPACE_BASE_DIR}/bin/openspace/${CMAKE_BUILD_TYPE}/ DESTINATION bin USE_SOURCE_PERMISSIONS)
 install(DIRECTORY ${OPENSPACE_BASE_DIR}/config/ DESTINATION config)
 
 # Data needs to be separately distributed due to its size... but then it crashes with no data folder...
