### 2017-06-20 789381e
Checking if model exists instead of checking if path to model folder exixts
