### 2017-06-20 22d37f2
More virtual properties and hardcoded stuff for the event
