### 2017-06-20 725ae7a
More virtual properties and hardcoded stuff for the event
