### 2017-06-20 7f39ebb
cleanup and commented all code
