+++ openspace.cfg
@@ -7,12 +7,12 @@ return {
     -- occurs in a single window, a fisheye projection, or a dome cluster system
 
     -- A regular 1280x720 window
-    SGCTConfig = sgct.config.single{},
+    --SGCTConfig = sgct.config.single{},
 
     -- SGCTConfig = sgct.config.single{res={1920, 1080}, shared=true},
 
     -- A regular 1920x1080 window
-    --SGCTConfig = sgct.config.single{1920, 1080, border=false, windowPos={0,0}},   
+    SGCTConfig = sgct.config.single{1920, 1080, border=false, windowPos={0,0}},   
 
     -- A 1k fisheye rendering
     -- SGCTConfig = sgct.config.fisheye{1024, 1024},

[…]
