### 2017-06-20 05844c1
Beginning to refactor brick selection algorithm to separate it from the ranking/points algorithms. This may not be the right direction.
