+++ modules/multiresvolume/rendering/brickselector.cpp
@@ -22,39 +22,77 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
  ****************************************************************************************/
 
-#ifndef __OPENSPACE_MODULE_MULTIRESVOLUME___TSPBRICKSELECTOR___H__
-#define __OPENSPACE_MODULE_MULTIRESVOLUME___TSPBRICKSELECTOR___H__
-
-#include <vector>
-#include <modules/multiresvolume/rendering/tsp.h>
 #include <modules/multiresvolume/rendering/brickselector.h>
 #include <modules/multiresvolume/rendering/brickselection.h>
-#include <modules/multiresvolume/rendering/brickcover.h>
+#include <modules/multiresvolume/rendering/errorhistogrammanager.h>
+#include <openspace/rendering/transferfunction.h>
+#include <openspace/util/histogram.h>
+#include <algorithm>
+#include <cassert>
+
+namespace {
+    const std::string _loggerCat = "BrickSelector";
+}
 
 namespace openspace {
+BrickSelector::BrickSelector() {}
 
-class TransferFunction;
+BrickSelector::BrickSelector(int memoryBudget, int streamingBudget)
+    : _memoryBudget(memoryBudget), _streamingBudget(streamingBudget) {};
 
-class TSPBrickSelector : public BrickSelector {
-public:
-    TSPBrickSelector();
-    TSPBrickSelector(TSP* tsp);
-    TSPBrickSelector(TSP* tsp, int memoryBudget, int streamingBudget);
-    TSPBrickSelector(TSP* tsp, TransferFunction* tf, int memoryBudget, int streamingBudget);
-    virtual ~TSPBrickSelector();
+BrickSelector::~BrickSelector() {}
 
-    virtual int getTimeStepCount();
-    virtual int getBrickCount();
+bool BrickSelector::initialize() { return true; }
 
-protected:
-    TSP* _tsp;
-    TransferFunction* _transferFunction;
+void BrickSelector::setMemoryBudget(int memoryBudget) { _memoryBudget = memoryBudget; };
 
-    int linearCoords(int x, int y, int z);
-    void writeSelection(BrickSelection brickSelection, std::vector<int>& bricks);
-    std::vector<float> getTfGradients();
-};
+void BrickSelector::setStreamingBudget(int streamingBudget) { _streamingBudget = streamingBudget; };
+
+int BrickSelector::getTimeStepCount() {
+    return 0;
+}
+
+int BrickSelector::getBrickCount() {
+    return 0;
+}
+
+float BrickSelector::getRank() {
+    return -1.;
+}
+
+void BrickSelector::selectBricks(int timestep, std::vector<int>& bricks) {
+    int numTimeSteps = getTimeStepCount();
+    int numBricksPerDim = getBrickCount();
+
+    unsigned int rootNode = 0;
+    BrickSelection::SplitType splitType = BrickSelection::SplitType::None;
+    float rootSplitPoints = getRank();
+
+
+    BrickSelection brickSelection = BrickSelection(numBricksPerDim, numTimeSteps, splitType, rootSplitPoints);
+
+    std::vector<BrickSelection> priorityQueue;
+    std::vector<BrickSelection> leafSelections;
+    std::vector<BrickSelection> temporalSplitQueue;
+    std::vector<BrickSelection> deadEnds;
+
+    if (splitType != BrickSelection::SplitType::None) {
+        priorityQueue.push_back(brickSelection);
+    } else {
+        leafSelections.push_back(brickSelection);
+    }
+
+    int memoryBudget = _memoryBudget;
+    int totalStreamingBudget = _streamingBudget * numTimeSteps;
[…]
