+++ modules/multiresvolume/CMakeLists.txt
@@ -50,6 +50,7 @@ source_group("Header Files" FILES ${HEADER_FILES})
 set(SOURCE_FILES
     ${CMAKE_CURRENT_SOURCE_DIR}/rendering/atlasmanager.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rendering/brickmanager.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/rendering/brickselector.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rendering/multiresvolumeraycaster.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rendering/tspbrickselector.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/rendering/shenbrickselector.cpp
