### 2017-06-21 661677e
Changed names of variables
