+++ modules/globebrowsing/models/roverpathfilereader.cpp
@@ -143,39 +143,39 @@ std::vector<std::shared_ptr<Subsite>> RoverPathFileReader::extractSubsitesWithMo
 
 	std::vector<std::shared_ptr<Subsite>> subsitesWithModels;
 	for (auto subsite : allSubsites) {
-		std::string pathToDriveFolderLevel1;
-		std::string pathToDriveFolderLevel2;
-		std::string pathToDriveFolderLevel3;
+		std::string pathToMeshFolderLevel1;
+		std::string pathToMeshFolderLevel2;
+		std::string pathToMeshFolderLevel3;
 
 		// Convert the site and drive string to match the folder structure
 		std::string site = convertString(subsite->site, "site");
 		std::string drive = convertString(subsite->drive, "drive");
-		pathToDriveFolderLevel1 = absPathToTModels + "level1/" + "site" + site + "/" + "drive" + drive;
-		pathToDriveFolderLevel2 = absPathToTModels + "level2/" + "site" + site + "/" + "drive" + drive;
-		pathToDriveFolderLevel3 = absPathToTModels + "level3/" + "site" + site + "/" + "drive" + drive;
+		pathToMeshFolderLevel1 = absPathToTModels + "level1/" + "site" + site + "/" + "drive" + drive;
+		pathToMeshFolderLevel2 = absPathToTModels + "level2/" + "site" + site + "/" + "drive" + drive;
+		pathToMeshFolderLevel3 = absPathToTModels + "level3/" + "site" + site + "/" + "drive" + drive;
 
 		// If the folder exists it means there are models for this subsite, then check if that
 		// specific site/drive combination has already been added. If the models haven't already been 
 		// added, loop through the text file with file names and add those to the subsite.
 		// Also store information about which levels are available for this specific subsite.
 		std::string restPath = "/OBJ.obj";
-		bool level1Exists = FileSys.fileExists(pathToDriveFolderLevel1 + restPath);
-		bool level2Exists = FileSys.fileExists(pathToDriveFolderLevel2 + restPath);
-		bool level3Exists = FileSys.fileExists(pathToDriveFolderLevel3 + restPath);
+		bool level1Exists = FileSys.fileExists(pathToMeshFolderLevel1 + restPath);
+		bool level2Exists = FileSys.fileExists(pathToMeshFolderLevel2 + restPath);
+		bool level3Exists = FileSys.fileExists(pathToMeshFolderLevel3 + restPath);
 
 		// TODO: refactor like hell!!!
 		std::string pathToDriveFolder;
 		if (level1Exists) {
 			subsite->availableLevels.push_back(1);
-			pathToDriveFolder = pathToDriveFolderLevel1;
+			pathToDriveFolder = pathToMeshFolderLevel1;
 		}
 		if (level2Exists) {
 			subsite->availableLevels.push_back(2);
-			pathToDriveFolder = pathToDriveFolderLevel2;
+			pathToDriveFolder = pathToMeshFolderLevel2;
 		}
 		if (level3Exists) {
 			subsite->availableLevels.push_back(3);
-			pathToDriveFolder = pathToDriveFolderLevel3;
+			pathToDriveFolder = pathToMeshFolderLevel3;
 		}
 		bool modelExists = false;
 		if(level1Exists || level2Exists || level3Exists) {
