### 2017-06-21 f6e5a0b
Minor performance fix
