### 2017-06-21 c61eefa
Cleanup and add blend mode Color
