### 2017-06-21 fcc0782
Add culling to gui
