### 2017-06-21 b7fc6e7
Get rid of magnetogram
