### 2017-06-21 ccd08c2
Update gotosol, update CMakeLists with everything from meshgeneration, might cause crash
