### 2017-06-21 8dc57e1
Add a fieldline utils file which for now helps to calculate topology
