### 2017-06-21 f8cb0d5
Enable setting of layeradjustment and blend mode from mod files.
