### 2017-06-21 87d27a1
Fix thread related issues.
