+++ modules/fieldlinessequence/rendering/renderablefieldlinessequence.cpp
@@ -83,6 +83,9 @@ namespace {
     const char* keyExtraColorTablePaths         = "ColorTablePaths";
     const char* keyExtraColorTableMinMaxs       = "ColorTableMinMax";
 
+    const char* keyCartesianDomainLimits        = "CartesianDomainLimits";
+    const char* keyRadialDomainLimits           = "RadialDomainLimits";
+
     // const char* keyVertexListFileType       = "FileType";
     // const char* keyVertexListFileTypeJson   = "Json";
     // const char* keyVertexListFileTypeBinary = "Binary";
@@ -162,6 +165,7 @@ RenderableFieldlinesSequence::RenderableFieldlinesSequence(const ghoul::Dictiona
 
     _loggerCat += " [" + name + "]";
 
+    // TODO Check dictionary for main fields and set bools accordingly?
     _dictionary = dictionary;
 }
 
@@ -304,43 +308,303 @@ bool RenderableFieldlinesSequence::initialize() {
             std::vector<std::string> extraMagVars{std::istream_iterator<std::string>{iss},
                                                   std::istream_iterator<std::string>{}};
 
-            // TODO Only one loop.. Not one for each tracing method!!
-            for (size_t i = 0; i < _numberOfStates; ++i) {
-                std::unique_ptr<ccmc::Kameleon> kameleon =
-                        fsManager.createKameleonObject(_validSourceFilePaths[i]);
-                if (kameleon == nullptr) {
-                    continue;
-                }
-                LDEBUG(_validSourceFilePaths[i] << " is now being traced.");
-                _states.push_back(FieldlinesState(_seedPoints.size()));
-                fsManager.getFieldlinesState(kameleon.get(),
-                                             tracingVariable,
-                                             _seedPoints,
-                                             tracingStepLength,
-                                             maxSteps,
-                                             _isMorphing,
-                                             numResamples,
-                                             resamplingOption,
-                                             extraFloatVars,
-                                             extraMagVars,
-                                             _states[i]);
-                _startTimes.push_back(_states[i]._triggerTime);
+// //---------------------- HARDCODE DELUXEEE START!-------------------------------------//
+    std::vector<std::string> sourcefiles;
+    std::vector<std::string> tmpfiles;
+    if (!fsManager.getAllFilePathsOfType(
+            "/media/ccarlbau/ff854077-f4bf-4644-a94b-1fe2d8824c64/Oskar_Carlbaum_061617_SH_1/3D_CDF", ".cdf", tmpfiles)) {
+            // "/media/ccarlbau/42a9eae3-e5e0-4d94-b439-f634d48b0d8c/Oskar_Carlbaum_060717_SH_4/3D_CDF", ".cdf", tmpfiles)) {
+        LERROR("Failed to get valid "<< ".cdf" << " file paths from '"
+                << "*pathToSourceFolder 061617_SH_1*" << "'" );
+        return false;
+    }
+    sourcefiles.insert(sourcefiles.end(),tmpfiles.begin(),tmpfiles.end());
+    // tmpfiles.clear();
+    // if (!fsManager.getAllFilePathsOfType(
+    //         "/media/ccarlbau/42a9eae3-e5e0-4d94-b439-f634d48b0d8c/Oskar_Carlbaum_060717_SH_3/3D_CDF", ".cdf", tmpfiles)) {
+    //     LERROR("Failed to get valid "<< ".cdf" << " file paths from '"
+    //             << "*pathToSourceFolder SH3*" << "'" );
+    //     return false;
+    // }
[…]
