### 2017-06-21 fd2f6bc
Clean up how time is extracted from a CCMC json file. Add (for now) hardcoded stuff to separate PFSS from BATRSUS.
