### 2017-06-21 ac2a223
Windows compile fixes and fix crash on exit bug
