+++ modules/solarbrowsing/rendering/renderablesolarimagery.h
@@ -77,7 +77,7 @@ public:
 
     float _currentScale;
     glm::vec2 _currentCenterPixel;
-    SolarImageData _currentSolarImageData;
+    //SolarImageData _currentSolarImageData;
     double _currentActiveImageTime;
     bool _isCoronaGraph;
     bool _shouldRenderPlane = true;
