### 2017-06-21 ae7fb7a
Update with level of detail based on radius for level 3 models
