+++ modules/globebrowsing/tasks/collapsedmeshgeneration.cpp
@@ -49,11 +49,13 @@
 #include <pcl/features/normal_3d_omp.h>
 #include <pcl/filters/passthrough.h>
 #include <pcl/filters/extract_indices.h>
+#include <pcl/surface/marching_cubes_rbf.h>
+#include <pcl/surface/vtk_smoothing/vtk_mesh_smoothing_laplacian.h>
 
 #include <fstream>
 
 namespace {
-	const std::string _loggerCat = "MeshGeneration";
+	const std::string _loggerCat = "CollapsedMeshGeneration";
 }
 
 namespace openspace {
@@ -86,6 +88,8 @@ namespace openspace {
 				levelOfDetail = levelOfDetailVars->second;
 			}
 
+			LERROR("Level : " << level);
+
 			ghoul::Dictionary filters;
 			dictionary.getValue("Filters.Filters", filters);
 
@@ -100,13 +104,13 @@ namespace openspace {
 
 				upSamplingFilters.push_back(std::move(PointCloudFilter::createFromDictionary(filterDic)));
 			}
-			
+
 			std::string pathToDriveFolder;
 			pcl::PointCloud<pcl::PointXYZ>::Ptr fullSite(new pcl::PointCloud<pcl::PointXYZ>);
 
 			std::vector<pcl::PointCloud<pcl::PointXYZ>::Ptr> tempVec;
 			glm::dmat4 outputMatrix;
-			
+
 			for (size_t i = 0; i < dictionary.size(); ++i) {
 				std::string pathToBinaryFile;
 
@@ -130,7 +134,7 @@ namespace openspace {
 					FileSys.createDirectory(pathToDriveFolder, k);
 				}
 				else if (FileSys.fileExists(pathToDriveFolder + outputFile + ".obj")) {
-					
+
 					LERROR("FILE ALREADY EXISTS " << pathToDriveFolder + outputFile + ".obj");
 					return;
 				}
@@ -145,12 +149,12 @@ namespace openspace {
 				pcl::PointCloud<pcl::PointXYZ>::Ptr fullCloud(new pcl::PointCloud<pcl::PointXYZ>);
 
 				MeshGeneration::extractCoordinatesFromArray(fullCloud, xyz, mInfo);
-			
+
 				tempVec.push_back(fullCloud);
 
 				pcl::PointCloud<pcl::PointXYZ>::Ptr prevCloud(new pcl::PointCloud<pcl::PointXYZ>());
 				prevCloud = fullCloud;
-				
+
 				for (size_t i = 0; i < upSamplingFilters.size(); ++i) {
 					pcl::PointCloud<pcl::PointXYZ>::Ptr outputCloud(new pcl::PointCloud<pcl::PointXYZ>());
 					upSamplingFilters.at(i)->setInputCloud(prevCloud);
@@ -165,7 +169,7 @@ namespace openspace {
 
 				glm::dmat4 rot = glm::dmat4(glm::toMat4(q3));
 
-				glm::dvec3 angle = glm::dvec3(0, M_PI, M_PI / 2.0);
+				glm::dvec3 angle = glm::dvec3(0, M_PI, -M_PI / 2.0);
 
 				glm::dmat4 unitMat4(1);
 
@@ -190,6 +194,15 @@ namespace openspace {
 				MeshGeneration::writeTxtFile(file_name_stripped, pathToDriveFolder, mInfo);
 			}
 
+
+			Eigen::Vector4f maxDist;
+
+			pcl::getMaxDistance(*fullSite, Eigen::Vector4f(0,0,0,1), maxDist);
+			LERROR("Distance from camera is : " << maxDist.norm());
+			LERROR("Max distance from camera: " << maxDist.x() << ", " << maxDist.y() << ", " << maxDist.z());
+
[…]
