### 2017-06-21 5550ad4
Windows cmake fixes and get rid of openjpeg
