### 2017-06-21 17c1916
New mesh writer, w/o texture coords
