### 2017-06-21 29cf644
Commented temporary code
