### 2017-06-21 f94d58d
improvement to tap interpretation by time instead of path size
