### 2017-06-21 7bb83ab
constructor looks more clean
