### 2017-06-21 f3bd167
Add missing scene file and fix typo
