+++ openspace.cfg
@@ -27,7 +27,7 @@ return {
 
     -- Sets the scene that is to be loaded by OpenSpace. A scene file is a description
     -- of all entities that will be visible during an instance of OpenSpace
-    Scene = "${SCENE}/satellite-trails.scene",
+    Scene = "${SCENE}/sun_earth_event.scene",
     -- Scene = "${SCENE}/globebrowsing.scene",
     -- Scene = "${SCENE}/rosetta.scene",
     -- Scene = "${SCENE}/dawn.scene",
