### 2017-06-21 2425681
Fix bug with lambda going out of scope
