### 2017-06-21 a9396a0
Normalize time varying volume values to range [0, 1]
