### 2017-06-22 190cb22
Fix bug with scaling/translation of basic volume raycaster
