+++ modules/volume/rendering/basicvolumeraycaster.cpp
@@ -79,12 +79,8 @@ glm::dmat4 BasicVolumeRaycaster::modelViewTransform(const RenderData& data) {
     glm::dmat4 modelTransform =
         glm::translate(glm::dmat4(1.0), data.modelTransform.translation) *
         glm::dmat4(data.modelTransform.rotation) *
-        glm::dmat4(
-            glm::scale(
-                glm::dmat4(_modelTransform),
-                glm::dvec3(data.modelTransform.scale)
-            )
-        );
+        glm::scale(glm::dmat4(1.0), glm::dvec3(data.modelTransform.scale)) *
+        glm::dmat4(_modelTransform);
 
     return data.camera.combinedViewMatrix() * modelTransform;
 }
