### 2017-06-22 bdb9684
Add possibility to change grid types for time varying volumes
