### 2017-06-22 a1eae95
Reduce number of texture units for sun projection
