### 2017-06-22 0d7fc02
Fix missing variable in fieldlinesequencemanager
