### 2017-06-22 fe12183
Add missing files in packaging Add install setting to the globebrowsing module to copy required gdal_data Update Credits
