+++ support/cmake/packaging.cmake
@@ -23,76 +23,92 @@
 #########################################################################################
 
 set(CPACK_MONOLITHIC_INSTALL TRUE)
+
 include(InstallRequiredSystemLibraries)
+
 set(CPACK_PACKAGE_NAME "OpenSpace")
 set(CPACK_PACKAGE_DESCRIPTION_FILE "${OPENSPACE_BASE_DIR}/README.md")
 set(CPACK_RESOURCE_FILE_LICENSE "${OPENSPACE_BASE_DIR}/LICENSE.md")
 set(CPACK_PACKAGE_VERSION_MAJOR "${OPENSPACE_VERSION_MAJOR}")
 set(CPACK_PACKAGE_VERSION_MINOR "${OPENSPACE_VERSION_MINOR}")
 set(CPACK_PACKAGE_VERSION_PATCH "${OPENSPACE_VERSION_PATCH}")
-set(OPENSPACE_VERSION_NUMBER "${OPENSPACE_VERSION_MAJOR}.${OPENSPACE_VERSION_MINOR}.${OPENSPACE_VERSION_PATCH}")
-set(CPACK_PACKAGE_INSTALL_DIRECTORY "OpenSpace ${OPENSPACE_VERSION}")
-set(CPACK_PACKAGE_FILE_NAME "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION_NUMBER} ${OPENSPACE_VERSION_STRING}")
+set(OPENSPACE_VERSION_NUMBER
+    "${OPENSPACE_VERSION_MAJOR}.${OPENSPACE_VERSION_MINOR}.${OPENSPACE_VERSION_PATCH}"
+)
+set(CPACK_PACKAGE_INSTALL_DIRECTORY "OpenSpace ${OPENSPACE_VERSION_NUMBER}")
+set(CPACK_PACKAGE_FILE_NAME
+    "${CPACK_PACKAGE_NAME} ${OPENSPACE_VERSION_NUMBER} ${OPENSPACE_VERSION_STRING}"
+)
 set(CPACK_STRIP_FILES 1)
 
-install(DIRECTORY ${OPENSPACE_BASE_DIR}/bin/openspace/${CMAKE_BUILD_TYPE}/ DESTINATION bin USE_SOURCE_PERMISSIONS)
+install(DIRECTORY
+    ${OPENSPACE_BASE_DIR}/bin/openspace/${CMAKE_BUILD_TYPE}/
+    DESTINATION bin
+    USE_SOURCE_PERMISSIONS
+)
 install(DIRECTORY ${OPENSPACE_BASE_DIR}/config/ DESTINATION config)
 
-# Data needs to be separately distributed due to its size... but then it crashes with no data folder...
 install(DIRECTORY ${OPENSPACE_BASE_DIR}/data/ DESTINATION data)
 
-install(DIRECTORY ${OPENSPACE_BASE_DIR}/documentation/ DESTINATION documentation)
 install(DIRECTORY ${OPENSPACE_BASE_DIR}/modules/
     DESTINATION modules
     FILES_MATCHING
-	PATTERN "*.glsl"
-	PATTERN "*.hglsl"
-	PATTERN "*.lua"
+    PATTERN "*.glsl"
+    PATTERN "*.hglsl"
+    PATTERN "*.fs"
+    PATTERN "*.vs"
+    PATTERN "*.lua"
 )
 install(DIRECTORY ${OPENSPACE_BASE_DIR}/scripts/ DESTINATION scripts)
 install(DIRECTORY ${OPENSPACE_BASE_DIR}/shaders/ DESTINATION shaders)
[…]
