### 2017-06-22 348526c
Added sol number to drop down menu
