### 2017-06-22 b1399f7
Changed name from goToSol to goToSubsite
