+++ modules/globebrowsing/models/drivepropertyowner.h
@@ -36,7 +36,7 @@ public:
 	DrivePropertyOwner(std::string drive, std::string sol, glm::dvec2 driveCoords);
 private:
 
-	void goToSubSite(std::string drive);
+	void goToSubsite(std::string drive);
 
 	std::string _drive;
 	glm::dvec2 _driveCoords;
