### 2017-06-22 aefb9c3
Removed lua function goToSubsite, using goToGeo instead
