+++ include/openspace/interaction/interactionhandler.h
@@ -69,7 +69,6 @@ public:
     
     void goToChunk(int x, int y, int level);
     void goToGeo(double latitude, double longitude);
-	void goToSubsite(double latitude, double longitude);
     
     void resetKeyBindings();
 
