### 2017-06-22 257f440
Fix bug with to big leaf size
