+++ modules/globebrowsing/tasks/collapsedmeshgeneration.cpp
@@ -260,6 +260,26 @@ namespace openspace {
 
 			
 			////////////////////////////////////////////////////////////////////////////
+			pcl::PointXYZ camera;
+			camera.x = -0.7;
+			camera.y = 0;
+			camera.z = 0;
+
+			pcl::search::KdTree<pcl::PointXYZ>::Ptr kdTreeFar(new pcl::search::KdTree<pcl::PointXYZ>);
+			kdTreeFar->setInputCloud(normalCloudRotated);
+
+			std::vector<int> indicesRadius;
+			std::vector<float> indiciesDistance;
+
+			kdTreeFar->radiusSearch(camera, 50, indicesRadius, indiciesDistance);
+
+			boost::shared_ptr<std::vector<int>> indicesRadiusPtr(new std::vector<int>(indicesRadius));
+
+			pcl::ExtractIndices<pcl::PointXYZ> innerFilter(true); // Initializing with true will allow us to extract the removed indices
+			innerFilter.setInputCloud(normalCloudRotated);
+			innerFilter.setIndices(indicesRadiusPtr);
+			
+			innerFilter.filterDirectly(normalCloudRotated);
 
 			pcl::PointCloud<pcl::PointXYZ>::Ptr xf_cloud(new pcl::PointCloud<pcl::PointXYZ>);
 			pcl::PointCloud<pcl::PointXYZ>::Ptr yf_cloud(new pcl::PointCloud<pcl::PointXYZ>);
@@ -285,97 +305,93 @@ namespace openspace {
 			if (level == "level3") {
 				pcl::search::KdTree<pcl::PointXYZ>::Ptr kdTree2(new pcl::search::KdTree<pcl::PointXYZ>);
 				kdTree2->setInputCloud(zf_cloud);
-				pcl::PointXYZ camera;
-				camera.x = -0.7;
-				camera.y = 0;
-				camera.z = 0;
+				
 				std::vector<int> indicesInnerRadius;
 				std::vector<float> indiciesDist;
 				kdTree2->radiusSearch(camera, 5.0, indicesInnerRadius, indiciesDist);
 				boost::shared_ptr<std::vector<int>> indicesInnerRadiusPtr(new std::vector<int>(indicesInnerRadius));
-
-				typename pcl::PointCloud<pcl::PointXYZ>::Ptr normalCloudRotated2(new pcl::PointCloud<pcl::PointXYZ>);
-				//The smaller one was with 10 i radius
-				pcl::copyPointCloud(*zf_cloud, *normalCloudRotated2);
-
-				typename pcl::PointCloud<pcl::PointXYZ>::Ptr normalCloudRotated3(new pcl::PointCloud<pcl::PointXYZ>);
-				//The smaller one was with 10 i radius
-				pcl::copyPointCloud(*zf_cloud, *normalCloudRotated3);
-
 				/////////////////////////////////////INNER RADIUS
 
 				pcl::PointCloud<pcl::PointXYZ>::Ptr innerRadiusFiltered(new pcl::PointCloud<pcl::PointXYZ>);
+				pcl::PointCloud<pcl::PointXYZ>::Ptr withoutInnerRadius(new pcl::PointCloud<pcl::PointXYZ>);
+
 				pcl::ExtractIndices<pcl::PointXYZ> innerFilter(true); // Initializing with true will allow us to extract the removed indices
 				innerFilter.setInputCloud(zf_cloud);
 				innerFilter.setIndices(indicesInnerRadiusPtr);
+				
 				innerFilter.filter(*innerRadiusFiltered);
+
+				innerFilter.setNegative(true);
[…]
