### 2017-06-22 2daaac0
Disable TcpSocketServer temporarily
This was, for some reason not yet known, blocking the web socket server.
