### 2017-06-22 d8c1a2a
Change level of detail radius
