### 2017-06-22 fa26d66
Increase number of neighbours investigated for mesh generation
