### 2017-06-22 6c36d46
Windows compile fixes and new key bindings
