### 2017-06-22 a123b1e
Fix bug that caused crash when pcl only close to camera
