### 2017-06-23 2bb733e
Removed log
