### 2017-06-23 5caf871
improve LMA converging ability on zoom cases and limit max iterations to 3k instead of 5k
