### 2017-06-23 d01a900
Implementing Shen-style selector. Starting to try to optimize Sand TSP temporal errors.
