### 2017-06-24 4f48cad
More loops, fixes and hardcoded stuff for the event
