### 2017-06-24 c6090ad
Improve support for spherical grids
