+++ modules/volume/rendering/renderabletimevaryingvolume.cpp
@@ -52,7 +52,6 @@ namespace {
     const char* KeyStepSize = "StepSize";
     const char* KeyTransferFunction = "TransferFunction";
     const char* KeySourceDirectory = "SourceDirectory";
-    const char* KeyVariable = "Variable";
     const char* KeyLowerDomainBound = "LowerDomainBound";
     const char* KeyUpperDomainBound = "UpperDomainBound";
     const char* KeyLowerValueBound = "LowerValueBound";
@@ -60,9 +59,7 @@ namespace {
     const char* KeyClipPlanes = "ClipPlanes";
     const char* KeySecondsBefore = "SecondsBefore";
     const char* KeySecondsAfter = "SecondsAfter";
-    const char* KeyCache = "Cache";
     const char* KeyGridType = "GridType";
-    const char* ValueSphericalGridType = "Spherical";
     const char* KeyMinValue = "MinValue";
     const char* KeyMaxValue = "MaxValue";
     const char* KeyTime = "Time";
@@ -307,13 +304,22 @@ void RenderableTimeVaryingVolume::update(const UpdateData& data) {
         Timestep* t = currentTimestep();
         _currentTimestep = timestepIndex(t);
         if (t && t->texture) {
-            glm::dvec3 scale = t->upperDomainBound - t->lowerDomainBound;
-            glm::dvec3 translation = (t->lowerDomainBound + t->upperDomainBound) * 0.5f;
+            if (_raycaster->gridType() == volume::VolumeGridType::Cartesian) {
+                glm::dvec3 scale = t->upperDomainBound - t->lowerDomainBound;
+                glm::dvec3 translation = (t->lowerDomainBound + t->upperDomainBound) * 0.5f;
 
-            glm::dmat4 modelTransform = glm::translate(glm::dmat4(1.0), translation);
-            glm::dmat4 scaleMatrix = glm::scale(glm::dmat4(1.0), scale);
-            modelTransform = modelTransform * scaleMatrix;
-            _raycaster->setModelTransform(glm::mat4(modelTransform));
+                glm::dmat4 modelTransform = glm::translate(glm::dmat4(1.0), translation);
+                glm::dmat4 scaleMatrix = glm::scale(glm::dmat4(1.0), scale);
+                modelTransform = modelTransform * scaleMatrix;
+                _raycaster->setModelTransform(glm::mat4(modelTransform));
+            } else {
+                _raycaster->setModelTransform(
+                    glm::scale(
+                        glm::dmat4(1.0),
+                        glm::dvec3(t->upperDomainBound[0])
+                    )
+                );
+            }
             _raycaster->setVolumeTexture(t->texture);
         } else {
             _raycaster->setVolumeTexture(nullptr);
