### 2017-06-24 a92a1c3
Improve support for spherical grids
