### 2017-06-24 87caf8b
Improve time handling in Kameleon to Raw converter
