### 2017-06-24 01944f6
Improve time handling in Kameleon to Raw converter
