### 2017-06-24 2d39e7e
Clean up time conversion
