### 2017-06-24 50dc8fa
Clean up time conversion
