### 2017-06-24 b46edaf
Add opacity slider for time varying volumes
