+++ modules/volume/rendering/basicvolumeraycaster.h
@@ -81,6 +81,8 @@ public:
     std::shared_ptr<ghoul::opengl::Texture> volumeTexture() const;
     void setTransferFunction(std::shared_ptr<TransferFunction> transferFunction);
     void setStepSize(float stepSize);
+    float opacity() const;
+    void setOpacity(float opacity);
     VolumeGridType gridType() const;
     void setGridType(VolumeGridType gridType);
     void setModelTransform(const glm::mat4& transform);
@@ -93,6 +95,7 @@ private:
     BoxGeometry _boundingBox;
     VolumeGridType _gridType;
     glm::mat4 _modelTransform;
+    float _opacity;
 
     std::unique_ptr<ghoul::opengl::TextureUnit> _tfUnit;
     std::unique_ptr<ghoul::opengl::TextureUnit> _textureUnit;
