### 2017-06-24 b446a5f
Add opacity slider for time varying volumes
