+++ modules/volume/rendering/renderabletimevaryingvolume.h
@@ -85,6 +85,7 @@ private:
     std::shared_ptr<VolumeClipPlanes> _clipPlanes;
 
     properties::FloatProperty _stepSize;
+    properties::FloatProperty _opacity;
     properties::FloatProperty _secondsBefore;
     properties::FloatProperty _secondsAfter;
     properties::StringProperty _sourceDirectory;
