+++ modules/volume/rendering/renderabletimevaryingvolume.h
@@ -86,6 +86,7 @@ private:
 
     properties::FloatProperty _stepSize;
     properties::FloatProperty _opacity;
+    properties::FloatProperty _rNormalization;
     properties::FloatProperty _secondsBefore;
     properties::FloatProperty _secondsAfter;
     properties::StringProperty _sourceDirectory;
