+++ modules/volume/rendering/renderabletimevaryingvolume.cpp
@@ -82,6 +82,7 @@ RenderableTimeVaryingVolume::RenderableTimeVaryingVolume(const ghoul::Dictionary
     , _jumpToTimestep("jumpToTimestep", "Jump to timestep", 0, 0, 256)
     , _currentTimestep("currentTimestep", "Current timestep", 0, 0, 256)
     , _opacity("opacity", "Opacity", 20.0f, 0.0f, 50.0f)
+    , _rNormalization("rNormalization", "Radius normalization", 0.0f, 0.0f, 2.0f)
     , _raycaster(nullptr)
     , _transferFunction(nullptr)
 {
@@ -205,6 +206,7 @@ bool RenderableTimeVaryingVolume::initialize() {
     addProperty(_jumpToTimestep);
     addProperty(_currentTimestep);
     addProperty(_opacity);
+    addProperty(_rNormalization);
 
     _raycaster->setGridType((_gridType.value() == 1) ? VolumeGridType::Spherical : VolumeGridType::Cartesian);
     _gridType.onChange([this] {
@@ -327,6 +329,7 @@ void RenderableTimeVaryingVolume::update(const UpdateData& data) {
         }
         _raycaster->setStepSize(_stepSize);
         _raycaster->setOpacity(_opacity);
+        _raycaster->setRNormalization(_rNormalization);
     }
 }
 
