+++ modules/volume/rendering/basicvolumeraycaster.h
@@ -85,6 +85,7 @@ public:
     void setOpacity(float opacity);
     float rNormalization() const;
     void setRNormalization(float opacity);
+    void setValueRemapping(float mapZeroTo, float mapOneTo);
     VolumeGridType gridType() const;
     void setGridType(VolumeGridType gridType);
     void setModelTransform(const glm::mat4& transform);
@@ -99,6 +100,7 @@ private:
     glm::mat4 _modelTransform;
     float _opacity;
     float _rNormalization;
+    glm::vec2 _valueRemapping;
 
     std::unique_ptr<ghoul::opengl::TextureUnit> _tfUnit;
     std::unique_ptr<ghoul::opengl::TextureUnit> _textureUnit;
