+++ modules/volume/rendering/basicvolumeraycaster.cpp
@@ -53,6 +53,7 @@ BasicVolumeRaycaster::BasicVolumeRaycaster(
     , _boundingBox(glm::vec3(1.0))
     , _opacity(20.0)
     , _rNormalization(0.0)
+    , _valueRemapping(0.0, 1.0)
     {}
     
 BasicVolumeRaycaster::~BasicVolumeRaycaster() {}
@@ -141,6 +142,7 @@ void BasicVolumeRaycaster::preRaycast(
     program.setUniform("clipOffsets_" + id, clipOffsets.data(), nClips);
     program.setUniform("opacity_" + id, _opacity);
     program.setUniform("rNormalization_" + id, _rNormalization);
+    program.setUniform("valueRemapping_" + id, _valueRemapping);
 }
     
 void BasicVolumeRaycaster::postRaycast(
@@ -218,6 +220,10 @@ void BasicVolumeRaycaster::setRNormalization(float rNormalization) {
 
 float BasicVolumeRaycaster::rNormalization() const {
     return _rNormalization;
+}
+
+void BasicVolumeRaycaster::setValueRemapping(float mapZeroTo, float mapOneTo) {
+    _valueRemapping = glm::vec2(mapZeroTo, mapOneTo);
 }
 
 VolumeGridType BasicVolumeRaycaster::gridType() const {
