### 2017-06-25 765ad67
Handle ratios in kameleon volume reader
Conflicts:
	modules/kameleonvolume/kameleonvolumereader.cpp
