### 2017-06-26 d7b69c7
Send closing reason to sockets
