### 2017-06-26 4a5e82c
[Server] Fix topic emplacing in Connection
