### 2017-06-27 4e9913d
Fix bugs in kameleon volume conversion
Conflicts:
	ext/ghoul
