### 2017-06-27 de961e1
[Server] Add BounceTopic for pinging purposes
Also, improve Topic structure a tad to support future development.
