### 2017-06-27 166cd65
[Server] Add BounceTopic for pinging purposes
Also, improve Topic structure a tad to support future development.
