### 2017-06-27 db1afca
Change so it won't crash when no pcl is found
