### 2017-06-27 34dcfeb
Add upper bound for radius
Conflicts:
	modules/volume/shaders/raycast.glsl
