+++ modules/volume/rendering/basicvolumeraycaster.h
@@ -84,7 +84,9 @@ public:
     float opacity() const;
     void setOpacity(float opacity);
     float rNormalization() const;
-    void setRNormalization(float opacity);
+    void setRNormalization(float rNormalization);
+    float rUpperBound() const;
+    void setRUpperBound(float rNormalization);
     void setValueRemapping(float mapZeroTo, float mapOneTo);
     VolumeGridType gridType() const;
     void setGridType(VolumeGridType gridType);
@@ -100,6 +102,7 @@ private:
     glm::mat4 _modelTransform;
     float _opacity;
     float _rNormalization;
+    float _rUpperBound;
     glm::vec2 _valueRemapping;
 
     std::unique_ptr<ghoul::opengl::TextureUnit> _tfUnit;
