### 2017-06-27 6f8f922
Initial Ozone Layer absorption.
