### 2017-06-27 61bd3b0
Adding boost to use memory mapping and hopefully speed up tsp error calcs. Possible online variance calculator
