+++ modules/multiresvolume/rendering/sandtsp.h
@@ -36,6 +36,7 @@ public:
 
     bool construct();
     bool calculateSpatialError();
+    float onlineVariance(std::vector<float> data);
     bool calculateTemporalError();
 
 protected:
