### 2017-06-27 b2f125f
[Server] Add working-ish AuthorizationTopic
It creates and waits for a 4-digit code from the client. This should
probably be set in a better way somehow. Authorization is not enforced
anywhere.
