### 2017-06-28 72dd671
[Server] Rename AuthorizationTopic -> AuthenticationTopic
