+++ modules/server/CMakeLists.txt
@@ -29,7 +29,7 @@ set(HEADER_FILES
     include/connectionpool.h
     include/connection.h
     include/topic.h
-    include/authorizationtopic.h
+    include/authenticationtopic.h
     ${CMAKE_CURRENT_SOURCE_DIR}/servermodule.h
 )
 source_group("Header Files" FILES ${HEADER_FILES})
@@ -38,7 +38,7 @@ set(SOURCE_FILES
     src/connectionpool.cpp
     src/connection.cpp
     src/topics/topic.cpp
-    src/topics/authorizationtopic.cpp
+    src/topics/authenticationtopic.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/servermodule.cpp
 )
 source_group("Source Files" FILES ${SOURCE_FILES})
