### 2017-06-28 35c9bff
[Server] Rename AuthorizationTopic -> AuthenticationTopic
