### 2017-06-28 6791df1
Add Config option for authentication requirement
