+++ openspace.cfg
@@ -73,6 +73,7 @@ return {
     KeyboardShortcuts = "${DOCUMENTATION}/KeyboardMapping.html",
     Documentation = "${DOCUMENTATION}/Documentation.html",
     FactoryDocumentation = "${DOCUMENTATION}/FactoryDocumentation.html",
+    RequireSocketAuthentication = true,
 
     ShutdownCountdown = 3,
     -- OnScreenTextScaling = "framebuffer",
