### 2017-06-28 1a96cc2
Add Config option for authentication requirement
