### 2017-06-28 ebdaea9
Clean up
