### 2017-06-28 4ba8bad
Safety check-in. Found memory mapping offset
