### 2017-06-28 eaeac80
CalculateTemporalError using memory mapping for 2 order of magnitude speedup
