### 2017-06-28 e74d7cd
Return of the calculateSpatialError. Used memory mapping for SpatialError speedup as well.
