### 2017-06-28 b2a665e
[WebGui] Clean webpack dev config, use SCSS for styles
