+++ modules/webgui/web/webpack.config.js
@@ -2,69 +2,61 @@ const { resolve } = require('path');
 const webpack = require('webpack');
 
 module.exports = {
-    context: resolve(__dirname, 'src'),
-
-    entry: [
-        'react-hot-loader/patch',
-        // activate HMR for React
-
-        'webpack-dev-server/client?http://localhost:8080',
-        // bundle the client for webpack-dev-server
-        // and connect to the provided endpoint
-
-        'webpack/hot/only-dev-server',
-        // bundle the client for hot reloading
-        // only- means to only hot reload for successful updates
-
-        './index.jsx',
-        // the entry point of our app
-    ],
-    output: {
-        filename: 'bundle.js',
-        // the output bundle
-
-        path: resolve(__dirname, 'dist'),
-
-        publicPath: '/static/',
-        // necessary for HMR to know where to load the hot update chunks
-    },
-
-    devtool: 'inline-source-map',
-
-    devServer: {
-        historyApiFallback: true,
-        hot: true,
-        // enable HMR on the server
-
-        contentBase: resolve(__dirname, 'dist'),
-
-        // match the output path
-        publicPath: '/static/',
-    },
-
-    module: {
-        rules: [
-            {
-                test: /\.jsx?$/,
-                use: [ 'babel-loader', ],
-                exclude: /node_modules/
+  context: resolve(__dirname, 'src'),
+  entry: [
+    'react-hot-loader/patch',
+    'webpack-dev-server/client?http://localhost:8080',
+    'webpack/hot/only-dev-server',
+    './index.jsx',
+  ],
+  output: {
+    filename: 'bundle.js',
+    path: resolve(__dirname, 'dist'),
+    publicPath: '/static/',
+  },
+  devtool: 'inline-source-map',
+  devServer: {
+    historyApiFallback: true,
+    hot: true,
+    contentBase: resolve(__dirname, 'dist'),
+    publicPath: '/static/',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        use: [ 'babel-loader', ],
+        exclude: /node_modules/
+      }, {
+        test: /\.scss$/,
+        use: [{
+            loader: 'style-loader',
+          }, {
+            loader: 'css-loader?modules',
+            options: {
+              sourceMap: true,
+              importLoaders: 2,
+              modules: true,
+              localIdentName: '[path][name]-[local]',
+              minimize: false,
             },
-            {
-                test: /\.css$/,
-                use: [ 'style-loader', 'css-loader?modules', ],
+          }, {
+            loader: 'sass-loader',
+            options: {
+              sourceMap: true,
             },
+          },
         ],
-    },
-
-    resolve: {
-        extensions: ['.js', '.jsx'],
-    },
-
-    plugins: [
-        new webpack.HotModuleReplacementPlugin(),
-        // enable HMR globally
-
-        new webpack.NamedModulesPlugin(),
-        // prints more readable module names in the browser console on HMR updates
+      },
     ],
+  },
+  resolve: {
+    extensions: ['.js', '.jsx'],
+  },
+  plugins: [
+    new webpack.HotModuleReplacementPlugin(),
+    // enable HMR globally
+    new webpack.NamedModulesPlugin(),
+    // prints more readable module names in the browser console on HMR updates
+  ],
 };
