### 2017-06-29 6a2e9d8
No need for grayscale layers. Reading grayscale in to rgb instead for color layers.
