### 2017-06-29 d4606e7
Remove unused layer groups
