### 2017-06-29 079b845
Acquiring min and max value when reading float volumes
