+++ modules/kameleonvolume/kameleonvolumereader.cpp
@@ -63,16 +63,31 @@ KameleonVolumeReader::KameleonVolumeReader(const std::string& path)
 std::unique_ptr<volume::RawVolume<float>> KameleonVolumeReader::readFloatVolume(
     const glm::uvec3 & dimensions,
     const std::string & variable,
-    const glm::vec3 & lowerBound,
-    const glm::vec3 & upperBound) const
+    const glm::vec3 & lowerDomainBound,
+    const glm::vec3 & upperDomainBound) const
 {
+    float min, max;
+    return readFloatVolume(dimensions, variable, lowerDomainBound, upperDomainBound, min, max);
+}
+
+std::unique_ptr<volume::RawVolume<float>> KameleonVolumeReader::readFloatVolume(
+    const glm::uvec3 & dimensions,
+    const std::string & variable,
+    const glm::vec3 & lowerBound,
+    const glm::vec3 & upperBound,
+    float& minValue,
+    float& maxValue) const
+{
+    minValue = FLT_MAX;
+    maxValue = FLT_MIN;
+
     auto volume = std::make_unique<volume::RawVolume<float>>(dimensions);
     const glm::vec3 dims = volume->dimensions();
     const glm::vec3 diff = upperBound - lowerBound;
 
     std::function<float(const std::string&, glm::vec3)> interpolate =
         [this](const std::string& variable, glm::vec3 volumeCoords) {
-            return _kameleonInterpolator->interpolate(
+            return _interpolator->interpolate(
                 variable,
                 volumeCoords[0],
                 volumeCoords[1],
@@ -107,11 +122,11 @@ std::unique_ptr<volume::RawVolume<float>> KameleonVolumeReader::readFloatVolume(
 
         data[index] = sample(volumeCoords);
 
-        if (data[index] < newMinValue) {
-            newMinValue = data[index];
+        if (data[index] < minValue) {
+            minValue = data[index];
         }
-        if (data[index] > newMaxValue) {
-            newMaxValue = data[index];
+        if (data[index] > maxValue) {
+            maxValue = data[index];
         }
     }
 
