+++ modules/kameleonvolume/kameleonvolumereader.h
@@ -49,6 +49,15 @@ public:
         const std::string& variable,
         const glm::vec3& lowerDomainBound,
         const glm::vec3& upperDomainBound) const;
+
+    std::unique_ptr<volume::RawVolume<float>> readFloatVolume(
+        const glm::uvec3& dimensions,
+        const std::string& variable,
+        const glm::vec3& lowerBound,
+        const glm::vec3& upperBound,
+        float& minValue,
+        float& maxValue) const;
+
     ghoul::Dictionary readMetaData() const;
 
     std::string time() const;
