### 2017-06-29 196cf78
Correctly read to grayscale layers
