+++ modules/globebrowsing/rendering/layer/layermanager.cpp
@@ -38,7 +38,14 @@ namespace {
 LayerManager::LayerManager(const ghoul::Dictionary& layerGroupsDict)
     : properties::PropertyOwner("Layers")
 {
+	// First create empty layer groups in case not all are specified
     _layerGroups.resize(layergroupid::NUM_LAYER_GROUPS);
+    for (int i = 0; i < _layerGroups.size(); ++i) {
+		ghoul::Dictionary emptyDict;
+        _layerGroups[i] = std::make_shared<LayerGroup>(
+			static_cast<layergroupid::GroupID>(i), emptyDict
+		);
+    }
 
     std::vector<std::string> layerGroupNamesInDict = layerGroupsDict.keys();
 
