### 2017-06-29 30f7bc9
Rename ColorOverlays to Overlays.
