### 2017-06-29 9761550
Cache success of calculating errors. Cleanup TSP functions and objects.
