### 2017-06-29 6ce0027
Moved error calculations into base TSP class. Activated tolerance sliders.
