+++ modules/multiresvolume/rendering/sandtsp.cpp
@@ -30,9 +30,6 @@
 #include <ghoul/logging/logmanager.h>
 #include <ghoul/glm.h>
 
-// boost
-#include <boost/iostreams/device/mapped_file.hpp>
-
 // std
 #include <algorithm>
 #include <math.h>
@@ -57,241 +54,6 @@ SandTSP::~SandTSP() {
 bool SandTSP::construct() {
     LDEBUG("Constructing TSP...");
     return TSP::construct();
-}
-
-bool SandTSP::calculateSpatialError() {
-    const unsigned int numBrickVals = paddedBrickDim_*paddedBrickDim_*paddedBrickDim_;
-
-    boost::iostreams::mapped_file_source mfile;
-    mfile.open(_filename);
-
-    if (!mfile.is_open()) {
-        return false;
-    }
-
-    const float * voxelData = (float *)mfile.data();
-    const long long headerOffset = dataPosition() / sizeof(float);
-
-    std::vector<float> averages(numTotalNodes_);
-    std::vector<float> stdDevs(numTotalNodes_);
-
-    // First pass: Calculate average color for each brick
-    LDEBUG("Calculating spatial error, first pass");
-    for (size_t brick = 0; brick<numTotalNodes_; ++brick) {
-        // Offset in file
-        const auto brickStart = headerOffset + static_cast<long long>(brick*numBrickVals);
-        double average = 0.0;
-
-        for (size_t i = 0; i < numBrickVals; i++) {
-            average += voxelData[brickStart + i];
-        }
-        averages[brick] = average / static_cast<double>(numBrickVals);
-    }
-
-    // Spatial SNR stats
-    float minError = 1e20f;
-    float maxError = 0.f;
-    std::vector<float> medianArray(numTotalNodes_);
-
-    // Second pass: For each brick, compare the covered leaf voxels with
-    // the brick average
-    LDEBUG("Calculating spatial error, second pass");
-    for (size_t brick = 0; brick<numTotalNodes_; ++brick) {
-
-        // Fetch mean intensity 
-        float brickAvg = averages[brick];
-
-        // Sum  for std dev computation
-        float stdDev = 0.f;
-
-        // Get a list of leaf bricks that the current brick covers
-        std::list<unsigned int> coveredLeafBricks =
-            CoveredLeafBricks(brick);
-
-        // If the brick is already a leaf, assign a negative error.
-        // Ad hoc "hack" to distinguish leafs from other nodes that happens
-        // to get a zero error due to rounding errors or other reasons.
-        if (coveredLeafBricks.size() == 1) {
-            stdDev = -0.1f;
-        }
-        else {
-
-            // Calculate "standard deviation" corresponding to leaves
-            for (auto lb = coveredLeafBricks.begin(); lb != coveredLeafBricks.end(); ++lb) {
-
-                // Add to sum
-                const auto leafStart = headerOffset + static_cast<long long>((*lb)*numBrickVals);
-
-                for (size_t i = 0; i < numBrickVals; i++) {
-                    stdDev += pow(voxelData[leafStart + i] - brickAvg, 2.f);
-                }
-            }
-
-            // Finish calculation
-            if (sizeof(float) != sizeof(int)) {
-                LERROR("Float and int sizes don't match, can't reintepret");
-                return false;
-            }
-
-            stdDev /= static_cast<float>(coveredLeafBricks.size()*numBrickVals);
-            stdDev = sqrt(stdDev);
-
-        } // if not leaf
-
-        if (stdDev < minError) {
-            minError = stdDev;
-        }
[…]
