+++ modules/multiresvolume/rendering/sandtsp.h
@@ -35,8 +35,6 @@ public:
     ~SandTSP();
 
     bool construct();
-    bool calculateSpatialError();
-    bool calculateTemporalError();
 
 protected:
 
