### 2017-06-29 19028b7
Move calculating brick averages into virtual helper function.
