### 2017-06-29 df2f35d
Move calculating standard deviations into virtual helper function. Refactoring to be able to replace error metrics in TSP subclasses.
