### 2017-06-29 32ea6b1
Cleanup calculateBrickStdDevs
