### 2017-06-29 1d875e4
Replace the ShenTSP calculateSpatialError with Shen method instead of Sand.
