+++ modules/multiresvolume/rendering/shentsp.cpp
@@ -30,6 +30,9 @@
 #include <ghoul/logging/logmanager.h>
 #include <ghoul/glm.h>
 
+// boost
+#include <boost/iostreams/device/mapped_file.hpp>
+
 // std
 #include <algorithm>
 #include <math.h>
@@ -54,121 +57,93 @@ bool ShenTSP::construct() {
     return TSP::construct();
 }
 
-bool ShenTSP::calculateSpatialError() {
-    unsigned int numBrickVals = paddedBrickDim_*paddedBrickDim_*paddedBrickDim_;
-
-    if (!_file.is_open())
-        return false;
-
-    std::vector<float> buffer(numBrickVals);
-    std::vector<float> averages(numTotalNodes_);
-    std::vector<float> stdDevs(numTotalNodes_);
-
-    // First pass: Calculate average color for each brick
-    LDEBUG("Calculating spatial error, first pass");
-    for (unsigned int brick = 0; brick<numTotalNodes_; ++brick) {
-
-        // Offset in file
-        std::streampos offset = dataPosition() + static_cast<long long>(brick*numBrickVals*sizeof(float));
-        _file.seekg(offset);
-
-        _file.read(reinterpret_cast<char*>(&buffer[0]),
-            static_cast<size_t>(numBrickVals)*sizeof(float));
-
-        double average = 0.0;
-        for (auto it = buffer.begin(); it != buffer.end(); ++it) {
-            average += *it;
-        }
-
-        averages[brick] = average / static_cast<double>(numBrickVals);
+std::vector<float> ShenTSP::calculateBrickStdDevs(std::vector<float> brickAverages) {
+    // Sanity check
+    if (sizeof(float) != sizeof(int)) {
+        LERROR("Float and int sizes don't match, can't reintepret");
+        return {};
     }
 
-    // Spatial SNR stats
-    float minError = 1e20f;
-    float maxError = 0.f;
-    std::vector<float> medianArray(numTotalNodes_);
+    boost::iostreams::mapped_file_source mfile;
+    mfile.open(_filename);
+
+    if (!mfile.is_open()) {
+        return {};
+    }
+
+    const float * voxelData = (float *)mfile.data();
+    const long long headerOffset = dataPosition() / sizeof(float);
+    const unsigned int numBrickVals = paddedBrickDim_*paddedBrickDim_*paddedBrickDim_;
+    std::vector<float> stdDevs(numTotalNodes_);
 
     // Second pass: For each brick, compare the covered leaf voxels with
     // the brick average
     LDEBUG("Calculating spatial error, second pass");
-    for (unsigned int brick = 0; brick<numTotalNodes_; ++brick) {
-
-        // Fetch mean intensity 
-        float brickAvg = averages[brick];
-
-        // Sum  for std dev computation
-        float stdDev = 0.f;
-
-        // Get a list of leaf bricks that the current brick covers
-        std::list<unsigned int> coveredLeafBricks =
-            CoveredLeafBricks(brick);
+    for (size_t brick = 0; brick<numTotalNodes_; ++brick) {
 
         // If the brick is already a leaf, assign a negative error.
         // Ad hoc "hack" to distinguish leafs from other nodes that happens
         // to get a zero error due to rounding errors or other reasons.
-        if (coveredLeafBricks.size() == 1) {
-            stdDev = -0.1f;
+        if (isOctreeLeaf(brick)) {
+            stdDevs[brick] = -0.1f;
+            continue;
         }
-        else {
 
-            // Calculate "standard deviation" corresponding to leaves
-            for (auto lb = coveredLeafBricks.begin();
-                lb != coveredLeafBricks.end(); ++lb) {
+        // Else for non-leaves: Calculate "standard deviation" corresponding
[…]
