### 2017-06-30 0e3e352
Fix temporal deviation calculation: mistype had it using Voxel Value instead of Voxel Average over Time.
