### 2017-06-30 6ef41ce
[WebGui] Add Storybook config
