### 2017-06-30 faa346e
Set correct slider ranges for time jumping in time varying volumes
