### 2017-06-30 93d5532
Ability to wait until next frame to set time
