### 2017-06-30 16a2674
Interaction speed is not dependent on framerate
