+++ src/interaction/interactionmode.cpp
@@ -424,13 +424,13 @@ void OrbitalInteractionMode::updateCameraStateFromMouseStates(Camera& camera, do
 
         { // Do local roll
             glm::dquat cameraRollRotation =
-                glm::angleAxis(_mouseStates->synchedLocalRollMouseVelocity().x, dvec3(0, 0, 1));
+                glm::angleAxis(_mouseStates->synchedLocalRollMouseVelocity().x * deltaTime, dvec3(0, 0, 1));
             localCameraRotation = localCameraRotation * cameraRollRotation;
         }
         if (!_rotateToFocusNodeInterpolator.isInterpolating())
         { // Do local rotation
             dvec3 eulerAngles(_mouseStates->synchedLocalRotationMouseVelocity().y, _mouseStates->synchedLocalRotationMouseVelocity().x, 0);
-            dquat rotationDiff = dquat(eulerAngles);
+            dquat rotationDiff = dquat(eulerAngles * deltaTime);
 
             localCameraRotation = localCameraRotation * rotationDiff;
         }
@@ -447,7 +447,7 @@ void OrbitalInteractionMode::updateCameraStateFromMouseStates(Camera& camera, do
         { // Do global rotation
             dvec2 smoothMouseVelocity = _mouseStates->synchedGlobalRotationMouseVelocity();
             dvec3 eulerAngles(-smoothMouseVelocity.y, -smoothMouseVelocity.x, 0);
-            dquat rotationDiffCamSpace = dquat(eulerAngles);
+            dquat rotationDiffCamSpace = dquat(eulerAngles * deltaTime);
 
             dquat newRotationCamspace = globalCameraRotation * rotationDiffCamSpace;
             dquat rotationDiffWorldSpace = newRotationCamspace * inverse(globalCameraRotation); 
@@ -470,11 +470,11 @@ void OrbitalInteractionMode::updateCameraStateFromMouseStates(Camera& camera, do
                 -directionToCenter *
                 boundingSphere;
             camPos += -(centerToCamera - centerToBoundingSphere) *
-                _mouseStates->synchedTruckMovementMouseVelocity().y;
+                _mouseStates->synchedTruckMovementMouseVelocity().y * deltaTime;
         }
         { // Roll around sphere normal
             dquat cameraRollRotation =
-                angleAxis(_mouseStates->synchedGlobalRollMouseVelocity().x, -directionToCenter);
+                angleAxis(_mouseStates->synchedGlobalRollMouseVelocity().x * deltaTime, -directionToCenter);
             globalCameraRotation = cameraRollRotation * globalCameraRotation;
         }
         { // Push up to surface
@@ -591,13 +591,13 @@ void GlobeBrowsingInteractionMode::updateCameraStateFromMouseStates(Camera& came
 
         { // Do local roll
             glm::dquat cameraRollRotation =
-                glm::angleAxis(_mouseStates->synchedLocalRollMouseVelocity().x, dvec3(0, 0, 1));
+                glm::angleAxis(_mouseStates->synchedLocalRollMouseVelocity().x * deltaTime, dvec3(0, 0, 1));
             localCameraRotation = localCameraRotation * cameraRollRotation;
         }
         if(!_rotateToFocusNodeInterpolator.isInterpolating())
         { // Do local rotation
             glm::dvec3 eulerAngles(_mouseStates->synchedLocalRotationMouseVelocity().y, _mouseStates->synchedLocalRotationMouseVelocity().x, 0);
-            glm::dquat rotationDiff = glm::dquat(eulerAngles);
+            glm::dquat rotationDiff = glm::dquat(eulerAngles * deltaTime);
 
             localCameraRotation = localCameraRotation * rotationDiff;
         }
[…]
