### 2017-06-30 f03cca1
ShenTSP calculateTemporalError. Uses stdDev of BST leaves against current brick to calculate coefficient of variation.
