### 2017-06-30 09e1ccc
Code/comment cleanup
