### 2017-06-30 e72c3ae
Remove the pow(.25) scaling factor on errors that has no apparent purpose.
