### 2017-06-30 71f4142
Add key to multiresvolume Renderable so you can choose the TSP generation method from the scene/mod files
