+++ modules/multiresvolume/rendering/renderablemultiresvolume.cpp
@@ -74,6 +74,7 @@ namespace {
     const std::string KeyErrorHistogramsSource = "ErrorHistogramsSource";
     const std::string KeyHints = "Hints";
     const std::string KeyTransferFunction = "TransferFunction";
+    const std::string KeyTspType = "TspType";
 
     const std::string KeyVolumeName = "VolumeName";
     const std::string KeyBrickSelector = "BrickSelector";
@@ -88,19 +89,27 @@ namespace {
 namespace openspace {
 
     const char* RenderableMultiresVolume::TYPE_SIMPLE = "simple";
-    const char* RenderableMultiresVolume::TYPE_TIME = "time";
-    const char* RenderableMultiresVolume::TYPE_TF = "tf";
-    const char* RenderableMultiresVolume::TYPE_LOCAL = "local";
-    const char* RenderableMultiresVolume::TYPE_SHEN = "shen";
+    const char* RenderableMultiresVolume::TYPE_TIME   = "time";
+    const char* RenderableMultiresVolume::TYPE_TF     = "tf";
+    const char* RenderableMultiresVolume::TYPE_LOCAL  = "local";
+    const char* RenderableMultiresVolume::TYPE_SHEN   = "shen";
 
-
+    const char* RenderableMultiresVolume::TSP_DEFAULT = "default";
+    const char* RenderableMultiresVolume::TSP_SAND    = "sand";
+    const char* RenderableMultiresVolume::TSP_SHEN    = "shen";
 
     const std::unordered_map<const char *, RenderableMultiresVolume::Selector> RenderableMultiresVolume::SelectorValues = {
-        {RenderableMultiresVolume::TYPE_SIMPLE , RenderableMultiresVolume::Selector::SIMPLE},
-        {RenderableMultiresVolume::TYPE_TF     , RenderableMultiresVolume::Selector::TF},
-        {RenderableMultiresVolume::TYPE_LOCAL  , RenderableMultiresVolume::Selector::LOCAL},
+        { RenderableMultiresVolume::TYPE_SIMPLE , RenderableMultiresVolume::Selector::SIMPLE},
+        { RenderableMultiresVolume::TYPE_TF     , RenderableMultiresVolume::Selector::TF},
+        { RenderableMultiresVolume::TYPE_LOCAL  , RenderableMultiresVolume::Selector::LOCAL},
         { RenderableMultiresVolume::TYPE_SHEN  , RenderableMultiresVolume::Selector::SHEN },
-        {RenderableMultiresVolume::TYPE_TIME   , RenderableMultiresVolume::Selector::TIME}
+        { RenderableMultiresVolume::TYPE_TIME   , RenderableMultiresVolume::Selector::TIME}
+    };
+
+    const std::unordered_map<const char *, RenderableMultiresVolume::TspType> RenderableMultiresVolume::TspTypes = {
+        { RenderableMultiresVolume::TSP_DEFAULT , RenderableMultiresVolume::TspType::DEFAULT },
+        { RenderableMultiresVolume::TSP_SAND    , RenderableMultiresVolume::TspType::SAND },
+        { RenderableMultiresVolume::TSP_SHEN    , RenderableMultiresVolume::TspType::SHEN }
     };
 
 RenderableMultiresVolume::RenderableMultiresVolume (const ghoul::Dictionary& dictionary)
@@ -132,6 +141,7 @@ RenderableMultiresVolume::RenderableMultiresVolume (const ghoul::Dictionary& dic
     , _rotation("rotation", "Euler rotation", glm::vec3(0.0, 0.0, 0.0), glm::vec3(0), glm::vec3(6.28))
[…]
