### 2017-07-01 e358d9d
Update to C++17
