### 2017-07-01 0d05d30
Feature/packaging (#333)
* Added proper CPACK structure and added generation of openspace.h header to add verisoning information in CMake.

* Packaging(ZIP on Windows, TGZ on UNIX) and installer creation(Windows Only via NSIS) is now working.

* Fixes for packing an archive on OSX.

* Add missing files in packaging
Add install setting to the globebrowsing module to copy required gdal_data
Update Credits

* Updated application icon
