### 2017-07-01 8a781a2
Compile fix for Windows 10 SDK
