### 2017-07-01 24485d8
Use icon for Launcher
