### 2017-07-03 cfdeea3
Compile fix in Pixelbuffer
