### 2017-07-03 de84bd0
Fix OpenGL compile error
