### 2017-07-03 80222ba
Make it possible to run OpenSpace without opening a scene
