+++ modules/onscreengui/onscreenguimodule.cpp
@@ -88,8 +88,10 @@ OnScreenGUIModule::OnScreenGUIModule()
             
             gui._property.setSource(
                 []() {
-                    const std::vector<SceneGraphNode*>& nodes =
-                        OsEng.renderEngine().scene()->allSceneGraphNodes();
+                    const Scene* scene = OsEng.renderEngine().scene();
+                    const std::vector<SceneGraphNode*>& nodes = scene ?
+                        scene->allSceneGraphNodes() :
+                        std::vector<SceneGraphNode*>();
                     return std::vector<properties::PropertyOwner*>(nodes.begin(), nodes.end());
                 }
             );
