+++ src/engine/settingsengine.cpp
@@ -65,18 +65,24 @@ void SettingsEngine::initialize() {
     // TODO: match regex with either with new ghoul readFiles or local code
     const std::string sceneDir = "${SCENE}";
     const std::vector<std::string> scenes = ghoul::filesystem::Directory(sceneDir).readFiles();
+    const size_t nScenes = scenes.size();
 
-    for (std::size_t i = 0; i < scenes.size(); ++i) {
+    for (std::size_t i = 0; i < nScenes; ++i) {
         const std::size_t found = scenes[i].find_last_of("/\\");
         _scenes.addOption(static_cast<int>(i), scenes[i].substr(found + 1));
     }
+    _scenes.addOption(scenes.size(), "None");
 
     // Set interaction to change ConfigurationManager and schedule the load
-    _scenes.onChange([this, sceneDir]() {
-        std::string sceneFile = _scenes.getDescriptionByValue(_scenes);
-        OsEng.configurationManager().setValue(
-            ConfigurationManager::KeyConfigScene, sceneFile);
-        OsEng.scheduleLoadScene(sceneDir + "/" + sceneFile);
+    _scenes.onChange([this, nScenes, sceneDir]() {
+        if (_scenes == nScenes) {
+            OsEng.scheduleLoadScene("");
+        } else {
+            std::string sceneFile = _scenes.getDescriptionByValue(_scenes);
+            OsEng.configurationManager().setValue(
+                ConfigurationManager::KeyConfigScene, sceneFile);
+            OsEng.scheduleLoadScene(sceneDir + "/" + sceneFile);
+        }
     }
     );
 }
