+++ src/util/time.cpp
@@ -111,7 +111,11 @@ void Time::setTime(std::string time, bool requireJump) {
 }
 
 std::string Time::UTC() const {
-    return SpiceManager::ref().dateFromEphemerisTime(_time);
+    try {
+        return SpiceManager::ref().dateFromEphemerisTime(_time);
+    } catch (const SpiceManager::SpiceException& c) {
+        return "0000 JAN 00 T00:00:00.00";
+    }
 }
 
 std::string Time::ISO8601() const {
