+++ src/query/query.cpp
@@ -120,15 +120,17 @@ std::vector<properties::Property*> allProperties() {
     );
 
     const Scene* graph = sceneGraph();
-    std::vector<SceneGraphNode*> nodes = graph->allSceneGraphNodes();
+    if (graph) {
+        std::vector<SceneGraphNode*> nodes = graph->allSceneGraphNodes();
 
-    for (SceneGraphNode* n : nodes) {
-        std::vector<properties::Property*> props = n->propertiesRecursive();
-        properties.insert(
-            properties.end(),
-            props.begin(),
-            props.end()
-        );
+        for (SceneGraphNode* n : nodes) {
+            std::vector<properties::Property*> props = n->propertiesRecursive();
+            properties.insert(
+                properties.end(),
+                props.begin(),
+                props.end()
+            );
+        }
     }
 
     return properties;
