### 2017-07-03 81164e0
Stub for Performance Manager & GUI Button
