### 2017-07-03 1d9bb74
Some clarification & iteration fixes.
