### 2017-07-04 c897421
Split up interaction code in files and perform smooth interpolation when changing focus
