+++ include/openspace/interaction/interactionmode.h
@@ -25,16 +25,14 @@
 #ifndef __OPENSPACE_CORE___INTERACTIONMODE___H__
 #define __OPENSPACE_CORE___INTERACTIONMODE___H__
 
+#include <openspace/interaction/interpolator.h>
+#include <openspace/interaction/inputstate.h>
+#include <openspace/interaction/delayedvariable.h>
+
 #include <openspace/network/parallelconnection.h>
 #include <openspace/util/mouse.h>
 #include <openspace/util/keys.h>
 #include <openspace/util/timeline.h>
-
-#ifdef OPENSPACE_MODULE_GLOBEBROWSING_ENABLED
-#include <modules/globebrowsing/tile/tileindex.h>
-#include <modules/globebrowsing/geometry/geodetic2.h>
-#include <modules/globebrowsing/geometry/geodetic3.h>
-#endif
 
 #include <list>
 
@@ -49,67 +47,13 @@ namespace globebrowsing {
 
 namespace interaction {
 
-    template <typename T>
-    class Interpolator
-    {
-    public:
-        Interpolator(std::function<T(double)> transferFunction)
-        : _t(0.0)
-        , _transferFunction(transferFunction) {};
-        ~Interpolator() {};
-
-        void start() { _t = 0.0; };
-        void end() { _t = 1.0; };
-        void step(double delta) { _t += delta; };
-
-        T value() { return _transferFunction(_t); };
-        bool isInterpolating() { return _t < 1.0; };
-    private:
-        std::function<T(double)> _transferFunction;
-        double _t;
-    };
-
-    class InputState
-    {
-    public:
-        InputState();
-        ~InputState();
-
-        // Callback functions
-        void keyboardCallback(Key key, KeyModifier modifier, KeyAction action);
-        void mouseButtonCallback(MouseButton button, MouseAction action);
-        void mousePositionCallback(double mouseX, double mouseY);
-        void mouseScrollWheelCallback(double mouseScrollDelta);
-
-        // Accessors
-        const std::list<std::pair<Key, KeyModifier> >& getPressedKeys() const;
-        const std::list<MouseButton>& getPressedMouseButtons() const;
-        glm::dvec2 getMousePosition() const;
-        double getMouseScrollDelta() const;
-        const std::vector<datamessagestructures::CameraKeyframe>& keyframes() const;
-
-        bool isKeyPressed(std::pair<Key, KeyModifier> keyModPair) const;
-        bool isKeyPressed(Key key) const;
-        bool isMouseButtonPressed(MouseButton mouseButton) const;
-    private:
-        // Input from keyboard and mouse
-        std::list<std::pair<Key, KeyModifier> > _keysDown;
-        std::list<MouseButton> _mouseButtonsDown;
-        glm::dvec2 _mousePosition;
-        double _mouseScrollDelta;
-    };
-
-
-
 class InteractionMode {
 public:
     InteractionMode();
     virtual ~InteractionMode();
 
-    // Mutators
     virtual void setFocusNode(SceneGraphNode* focusNode);
 
-    // Accessors
     SceneGraphNode* focusNode();
     Interpolator<double>& rotateToFocusNodeInterpolator();
     virtual bool followingNodeRotation() const = 0;
@@ -118,47 +62,6 @@ public:
     virtual void updateCameraStateFromMouseStates(Camera& camera, double deltaTime) = 0;
 
 protected:
-    /**
-        Inner class that acts as a smoothing filter to a variable. The filter has a step
-        response on a form that resembles the function y = 1-e^(-t/scale). The variable
-        will be updates as soon as it is set to a value (calling the set() function).
-    */
-    template <typename T, typename ScaleType>
-    class DelayedVariable {
-    public:
[…]
