### 2017-07-04 0901027
Abstract interaction code in to functions.
