+++ src/interaction/orbitalinteractionmode.cpp
@@ -175,102 +175,171 @@ OrbitalInteractionMode::~OrbitalInteractionMode() {
 
 }
 
+OrbitalInteractionMode::CameraRotationDecomposition
+	OrbitalInteractionMode::decomposeCameraRotation(const Camera& camera)
+{
+    // Read the current state of the camera and focus node
+    glm::dvec3 camPos = camera.positionVec3();
+    glm::dvec3 centerPos = _focusNode->worldPosition();
+    
+    glm::dquat totalRotation = camera.rotationQuaternion();
+    glm::dvec3 directionToCenter = glm::normalize(centerPos - camPos);
+    glm::dvec3 lookUp = camera.lookUpVectorWorldSpace();
+    glm::dvec3 camDirection = camera.viewDirectionWorldSpace();
+
+    // Create the internal representation of the local and global camera rotations
+    glm::dmat4 lookAtMat = glm::lookAt(
+        glm::dvec3(0, 0, 0),
+        directionToCenter,
+        normalize(camDirection + lookUp)); // To avoid problem with lookup in up direction
+    glm::dquat globalCameraRotation = glm::normalize(glm::quat_cast(glm::inverse(lookAtMat)));
+    glm::dquat localCameraRotation = glm::inverse(globalCameraRotation) * totalRotation;
+
+    return { localCameraRotation, globalCameraRotation };
+}
+
+void OrbitalInteractionMode::performRoll(double deltaTime, glm::dquat& localCameraRotation) {
+    glm::dquat rollQuat = glm::angleAxis(
+        _mouseStates->synchedLocalRollMouseVelocity().x * deltaTime,
+        glm::dvec3(0, 0, 1)
+    );
+    localCameraRotation = localCameraRotation * rollQuat;
+}
+
+void OrbitalInteractionMode::performLocalRotation(double deltaTime, glm::dquat& localCameraRotation) {
+    glm::dvec3 eulerAngles(
+        _mouseStates->synchedLocalRotationMouseVelocity().y,
+        _mouseStates->synchedLocalRotationMouseVelocity().x,
+        0
+    );
+    glm::dquat rotationDiff = glm::dquat(eulerAngles * deltaTime);
+	localCameraRotation = localCameraRotation * rotationDiff;
+}
+
+void OrbitalInteractionMode::interpolateLocalRotation(double deltaTime, glm::dquat& localCameraRotation) {
+    double t = _rotateToFocusNodeInterpolator.value();
+    _rotateToFocusNodeInterpolator.step(deltaTime);
+	localCameraRotation = glm::slerp(
+        localCameraRotation,
+        glm::dquat(glm::dvec3(0.0)),
+        glm::min(t * deltaTime, 1.0));
+}
+
+void OrbitalInteractionMode::performHorizontalTranslationAndRotation(
+	double deltaTime,
+	glm::dvec3 objectPosition,
+	glm::dvec3& cameraPosition,
+	glm::dquat& globalCameraRotation)
+{
+	glm::dvec3 centerToCamera = cameraPosition - objectPosition;
+
+    glm::dvec2 smoothMouseVelocity = _mouseStates->synchedGlobalRotationMouseVelocity();
+    glm::dvec3 eulerAngles(-smoothMouseVelocity.y, -smoothMouseVelocity.x, 0);
+    glm::dquat rotationDiffCamSpace = glm::dquat(eulerAngles * deltaTime);
+
+    glm::dquat newRotationCamspace = globalCameraRotation * rotationDiffCamSpace;
+    glm::dquat rotationDiffWorldSpace = newRotationCamspace * inverse(globalCameraRotation); 
[…]
