+++ src/interaction/interactionmode.cpp
@@ -37,12 +37,6 @@
 
 #include <glm/gtx/quaternion.hpp>
 
-#ifdef OPENSPACE_MODULE_GLOBEBROWSING_ENABLED
-#include <modules/globebrowsing/globes/renderableglobe.h>
-#include <modules/globebrowsing/globes/chunkedlodglobe.h>
-#include <modules/globebrowsing/geometry/geodetic2.h>
-#endif
-
 namespace {
     const std::string _loggerCat = "InteractionMode";
 }
@@ -50,10 +44,31 @@ namespace {
 namespace openspace {
 namespace interaction {
 
-InteractionMode::InteractionMode()
-    : _rotateToFocusNodeInterpolator(Interpolator<double>([](double t){
-        return (6 * (t + t*t) / (1 - 3 * t*t + 2 * t*t*t));
-    })) {
+InteractionMode::InteractionMode() {
+    // The transfer function is used here to get a different interpolation than the one
+    // obtained from newValue = lerp(0, currentValue, dt). That one will result in an
+    // exponentially decreasing value but we want to be able to control it. Either as
+    // a linear interpolation or a smooth step interpolation. Therefore we use
+    // newValue = lerp(0, currentValue * f(t) * dt) where f(t) is the transfer function
+    // and lerp is a linear iterpolation
+    // lerp(endValue, startValue, interpolationParameter).
+    //
+    // The transfer functions are derived from:
+    // f(t) = d/dt ( ln(1 / f_orig(t)) ) where f_orig is the transfer function that would
+    // be used if the interpolation was sinply linear between a start value and an end
+    // value instead of current value and end value (0) as we use it when inerpoláting.
+    // As an example f_orig(t) = 1 - t yields f(t) = 1 / (1 - t) which results in a linear
+    // interpolation from 1 to 0.
+    
+    auto smoothStepDerivedTranferFunction = 
+        [](double t) {
+            return (6 * (t + t*t) / (1 - 3 * t*t + 2 * t*t*t));
+        };
+    auto linearDerivedTranferFunction = 
+        [](double t) {
+            return 1 / (1 - t);
+        };
+    _rotateToFocusNodeInterpolator.setTransferFunction(smoothStepDerivedTranferFunction);
 }
 
 InteractionMode::~InteractionMode() {
