### 2017-07-04 2e2c0be
Interpolation time is dependent on angle to focus node.
