### 2017-07-05 b38409b
Fix bug regarding decomposition of camera rotation.
