### 2017-07-05 2e4b908
Work on asset loading
