+++ src/scene/assetloader.cpp
@@ -1,0 +1,336 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#include <openspace/scene/assetloader.h>
+#include <ghoul/lua/ghoul_lua.h>
+#include <ghoul/lua/luastate.h>
+#include <ghoul/lua/lua_helper.h>
+#include <ghoul/logging/logmanager.h>
+#include <ghoul/misc/onscopeexit.h>
+
+#include <ghoul/filesystem/filesystem.h>
+
+#include "assetloader_lua.inl"
+
+namespace {
+    const char* AssetFileSuffix = "asset";
+    const char* ImportFunctionName = "import";
+    const char* ExportFunctionName = "export";
+    const char* SyncedResourceFunctionName = "syncedResource";
+    const char* LocalResourceFunctionName = "localResource";
+    const char* _loggerCat = "AssetLoader";
+    const char* AssetsTableName = "_assets";
+    const char* KeySceneGraphNodes = "SceneGraphNodes";
+}
+
+namespace openspace {
+
+namespace assetloader {
+int importAsset(lua_State* state) {
+    AssetLoader *assetLoader = (AssetLoader*)lua_touserdata(state, lua_upvalueindex(1));
+    return assetLoader->importAssetLua();
+}
+
+int exportAsset(lua_State* state) {
+    AssetLoader::Asset* asset = (AssetLoader::Asset*)lua_touserdata(state, lua_upvalueindex(1));
+    return asset->exportAssetLua();
+}
+
+int resolveLocalResource(lua_State* state) {
[…]
