### 2017-07-05 e237b6f
Adding eclipse shadows to GlobeBrowsing and ATM.
