+++ modules/globebrowsing/globes/chunkedlodglobe.cpp
@@ -187,9 +187,9 @@ float ChunkedLodGlobe::getHeight(glm::dvec3 position) const {
             return 0;
         }
 
-		ghoul::opengl::Texture* tileTexture = tile.texture();
-		if (!tileTexture)
-			return 0;
+        ghoul::opengl::Texture* tileTexture = tile.texture();
+        if (!tileTexture)
+            return 0;
 
         glm::vec2 transformedUv = Tile::TileUvToTextureSamplePosition(
             uvTransform,
