+++ openspace.cfg
@@ -61,6 +61,7 @@ return {
     Logging = {
         -- LogLevel = "Trace",
         LogLevel = "Debug",
+        --LogLevel = "None",
         ImmediateFlush = true,
         Logs = {
             { Type = "html", File = "${BASE_PATH}/log.html", Append = false }
