### 2017-07-05 0cedfe6
Make orbital interaction mode behave as globe browsing and no longer use interactiondepth below ellipsoid.
