+++ src/rendering/renderable.cpp
@@ -158,6 +158,19 @@ void Renderable::render(const RenderData& data, RendererTasks&) {
 
 void Renderable::render(const RenderData&) {}
 
+Renderable::SurfacePositionHandle Renderable::calculateSurfacePositionHandle(
+    glm::dvec3 targetModelSpace)
+{
+    glm::dvec3 directionFromCenterToTarget = glm::normalize(targetModelSpace);
+	return {
+		directionFromCenterToTarget * static_cast<double>(boundingSphere()),
+		directionFromCenterToTarget,
+        directionFromCenterToTarget,
+		glm::length(targetModelSpace) - boundingSphere(),
+		0.0
+	};
+}
+
 void Renderable::setPscUniforms(ghoul::opengl::ProgramObject& program,
                                 const Camera& camera,
                                 const PowerScaledCoordinate& position)
