+++ src/interaction/orbitalinteractionmode.cpp
@@ -25,6 +25,7 @@
 #include <openspace/interaction/orbitalinteractionmode.h>
 
 #include <openspace/scene/scenegraphnode.h>
+#include <openspace/rendering/renderable.h>
 
 #include <ghoul/logging/logmanager.h>
 
@@ -182,18 +183,25 @@ OrbitalInteractionMode::CameraRotationDecomposition
 		glm::dvec3 cameraLookUp,
 		glm::dvec3 cameraViewDirection)
 {
-    // Read the current state of the camera and focus node
-    glm::dvec3 camPos = cameraPosition;
-    glm::dvec3 centerPos = _focusNode->worldPosition();
-	glm::dvec3 directionToCenter = normalize(centerPos - camPos);
+    glm::dmat4 inverseModelTransform = _focusNode->inverseModelTransform();
+    glm::dmat4 modelTransform = _focusNode->modelTransform();
+    glm::dvec3 cameraPositionModelSpace =
+        glm::dvec3(inverseModelTransform * glm::dvec4(cameraPosition, 1));
+
+    Renderable::SurfacePositionHandle posHandle =
+        _focusNode->renderable()->calculateSurfacePositionHandle(cameraPositionModelSpace);
+
+	glm::dvec3 directionFromSurfaceToCameraModelSpace = posHandle.referenceSurfaceNormal;
+	glm::dvec3 directionFromSurfaceToCamera =
+		glm::normalize(glm::dmat3(modelTransform) * directionFromSurfaceToCameraModelSpace);
 
     // Create the internal representation of the local and global camera rotations
-    glm::dmat4 lookAtMat = glm::lookAt(
-        glm::dvec3(0, 0, 0),
-        directionToCenter,
+	glm::dmat4 lookAtMat = glm::lookAt(
+		glm::dvec3(0.0, 0.0, 0.0),
+        -directionFromSurfaceToCamera,
         normalize(cameraViewDirection + cameraLookUp)); // To avoid problem with lookup in up direction
-    glm::dquat globalCameraRotation = glm::normalize(glm::quat_cast(glm::inverse(lookAtMat)));
-    glm::dquat localCameraRotation = glm::inverse(globalCameraRotation) * cameraRotation;
+	glm::dquat globalCameraRotation = glm::normalize(glm::quat_cast(inverse(lookAtMat)));
+	glm::dquat localCameraRotation = glm::inverse(globalCameraRotation) * cameraRotation;
 
     return { localCameraRotation, globalCameraRotation };
 }
@@ -259,63 +267,324 @@ void OrbitalInteractionMode::performHorizontalTranslationAndRotation(
 	globalCameraRotation = glm::normalize(quat_cast(inverse(lookAtMat)));
 }
 
+
+
+void OrbitalInteractionMode::performHorizontalTranslation(
+    double deltaTime,
+    glm::dvec3 objectPosition,
+    glm::dquat& focusNodeRotationDiff,
+    glm::dvec3& cameraPosition,
+    glm::dquat& globalCameraRotation)
+{
+	using namespace glm;
+    // Uniform variables
+    double ellipsoidShrinkTerm = 0;
+    glm::dmat4 inverseModelTransform = _focusNode->inverseModelTransform();
+    glm::dmat4 modelTransform = _focusNode->modelTransform();
+
+    // Get position handle
+    glm::dvec3 cameraPositionModelSpace =
+        glm::dvec3(inverseModelTransform * glm::dvec4(cameraPosition, 1));
+    Renderable::SurfacePositionHandle posHandle =
+        _focusNode->renderable()->calculateSurfacePositionHandle(cameraPositionModelSpace);
+
+    dvec3 surfaceNormal =
[…]
