### 2017-07-05 9015a8d
Example of access to invidvidual performance statistics
