### 2017-07-05 b1f8b87
PerformanceManager logic for ABufferRenderer
