### 2017-07-06 a49da2a
Do not always rotate with object. Depending on distance
