+++ src/interaction/orbitalinteractionmode.cpp
@@ -168,8 +168,14 @@ glm::dvec2 OrbitalInteractionMode::MouseStates::synchedGlobalRollMouseVelocity()
 OrbitalInteractionMode::OrbitalInteractionMode(
     std::shared_ptr<MouseStates> mouseStates)
     : InteractionMode()
-    , _mouseStates(mouseStates) {
-        
+    , _mouseStates(mouseStates)
+{
+    auto smoothStep = 
+        [](double t) {
+            double res = 3.0 * t*t  - 2.0 * t*t*t;
+            return glm::clamp(res, 0.0, 1.0);
+        };
+    //_followRotationInterpolator.setTransferFunction(smoothStep);
 }
 
 OrbitalInteractionMode::~OrbitalInteractionMode() {
@@ -191,7 +197,7 @@ OrbitalInteractionMode::CameraRotationDecomposition
     Renderable::SurfacePositionHandle posHandle =
         _focusNode->renderable()->calculateSurfacePositionHandle(cameraPositionModelSpace);
 
-	glm::dvec3 directionFromSurfaceToCameraModelSpace = posHandle.referenceSurfaceNormal;
+	glm::dvec3 directionFromSurfaceToCameraModelSpace = posHandle.referenceSurfaceOutDirection;
 	glm::dvec3 directionFromSurfaceToCamera =
 		glm::normalize(glm::dmat3(modelTransform) * directionFromSurfaceToCameraModelSpace);
 
@@ -289,13 +295,13 @@ void OrbitalInteractionMode::performHorizontalTranslation(
         _focusNode->renderable()->calculateSurfacePositionHandle(cameraPositionModelSpace);
 
     dvec3 surfaceNormal =
-        normalize(dmat3(modelTransform) * posHandle.referenceSurfaceNormal);
+        normalize(dmat3(modelTransform) * posHandle.referenceSurfaceOutDirection);
 
     dvec3 posDiff = cameraPosition - objectPosition;
 
     dvec3 centerToReferenceSurface = dmat3(modelTransform) * posHandle.centerToReferenceSurface;
     dvec3 centerToActualSurfaceModelSpace = posHandle.centerToReferenceSurface +
-        posHandle.referenceSurfaceNormal * posHandle.heightToSurface;
+        posHandle.referenceSurfaceOutDirection * posHandle.heightToSurface;
     dvec3 centerToActualSurface = dmat3(modelTransform) * centerToActualSurfaceModelSpace;
     dvec3 actualSurfaceToCamera = posDiff - centerToActualSurface;
     double distFromSurfaceToCamera = glm::length(actualSurfaceToCamera);
@@ -357,7 +363,7 @@ void OrbitalInteractionMode::performGlobalRotation(
     Renderable::SurfacePositionHandle posHandle =
         _focusNode->renderable()->calculateSurfacePositionHandle(cameraPositionModelSpace);
 
-    glm::dvec3 directionFromSurfaceToCameraModelSpace = posHandle.referenceSurfaceNormal;
+    glm::dvec3 directionFromSurfaceToCameraModelSpace = posHandle.referenceSurfaceOutDirection;
     glm::dvec3 directionFromSurfaceToCamera =
       normalize(dmat3(modelTransform) * directionFromSurfaceToCameraModelSpace);
 
@@ -390,7 +396,7 @@ void OrbitalInteractionMode::performVerticalTranslation(
 
     dvec3 centerToReferenceSurface = dmat3(modelTransform) * posHandle.centerToReferenceSurface;
     dvec3 centerToActualSurfaceModelSpace = posHandle.centerToReferenceSurface +
-        posHandle.referenceSurfaceNormal * posHandle.heightToSurface;
[…]
