+++ modules/globebrowsing/globes/renderableglobe.cpp
@@ -265,14 +265,15 @@ Renderable::SurfacePositionHandle RenderableGlobe::calculateSurfacePositionHandl
     glm::dvec3 centerToEllipsoidSurface = _ellipsoid.geodeticSurfaceProjection(targetModelSpace);
     glm::dvec3 ellipsoidSurfaceToTarget = targetModelSpace - centerToEllipsoidSurface;
     double heightToSurface = getHeight(targetModelSpace);
-    
-    glm::dvec3 geodeticNormal = glm::normalize(ellipsoidSurfaceToTarget);
-    if (glm::dot(geodeticNormal, centerToEllipsoidSurface) < 0) {
-        geodeticNormal *= -1.0;
+    glm::dvec3 geodeticNormal = ellipsoid().geodeticSurfaceNormal(
+                    ellipsoid().cartesianToGeodetic2(targetModelSpace));
+    glm::dvec3 ellipsoidSurfaceOutDirection = glm::normalize(ellipsoidSurfaceToTarget);
+    if (glm::dot(ellipsoidSurfaceOutDirection, centerToEllipsoidSurface) < 0) {
+		ellipsoidSurfaceOutDirection *= -1.0;
     }
 	return {
 		centerToEllipsoidSurface,
-		glm::normalize(ellipsoidSurfaceToTarget),
+		ellipsoidSurfaceOutDirection,
         geodeticNormal,
         glm::length(ellipsoidSurfaceToTarget),
 		heightToSurface
