### 2017-07-06 ee36a97
Improve asset lua interface. Simplify scene loading.
