+++ src/scene/scenemanager.cpp
@@ -1,59 +1,0 @@
-/*****************************************************************************************
- *                                                                                       *
- * OpenSpace                                                                             *
- *                                                                                       *
- * Copyright (c) 2014-2017                                                               *
- *                                                                                       *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
- * software and associated documentation files (the "Software"), to deal in the Software *
- * without restriction, including without limitation the rights to use, copy, modify,    *
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
- * permit persons to whom the Software is furnished to do so, subject to the following   *
- * conditions:                                                                           *
- *                                                                                       *
- * The above copyright notice and this permission notice shall be included in all copies *
- * or substantial portions of the Software.                                              *
- *                                                                                       *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
- * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
- ****************************************************************************************/
-
-#include <openspace/scene/scene.h>
-#include <openspace/scene/sceneloader.h>
-#include <openspace/scene/scenemanager.h>
-
-#include <algorithm>
-#include <memory>
-
-namespace openspace {
-
-SceneManager::SceneManager(std::string assetRoot, std::string syncRoot)
-    : _assetLoader(assetRoot, syncRoot)
-    , _sceneLoader(&_assetLoader)
-{}
-
-Scene* SceneManager::loadScene(const std::string& path) {
-    std::unique_ptr<Scene> scene = _sceneLoader.loadScene(path);
-    Scene* s = scene.get();
-    if (s) {
-        _scenes.push_back(std::move(scene));
-    }
-    return s;
-}
-
-void SceneManager::unloadScene(Scene& scene) {
-    _scenes.erase(std::remove_if(
-        _scenes.begin(), _scenes.end(), [&scene] (auto& s) {
-            return s.get() == &scene;
-        }), _scenes.end()
-    );
-}
-
-void SceneManager::unloadAll() {
-    _scenes.clear();
-}
-}
