### 2017-07-06 ae76924
Remove interaction depth below ellipsoid. Now able to interact without renderable
