### 2017-07-06 6f83696
Remove GlobeBrowsingInteractionMode
