### 2017-07-06 d16c0cf
Rename OrbitalInteractionMode to OrbitalNavigator and no longer extend interactionmode.
