+++ src/interaction/orbitalinteractionmode.cpp
@@ -31,16 +31,17 @@
 #include <ghoul/logging/logmanager.h>
 
 #include <glm/gtx/quaternion.hpp>
+#include <glm/gtx/vector_angle.hpp>
 
 namespace {
-    const std::string _loggerCat = "OrbitalInteractionMode";
+    const std::string _loggerCat = "OrbitalNavigator";
 }
 
 namespace openspace {
 namespace interaction {
 
-// OrbitalInteractionMode
-OrbitalInteractionMode::MouseStates::MouseStates(double sensitivity, double velocityScaleFactor)
+// OrbitalNavigator
+OrbitalNavigator::MouseStates::MouseStates(double sensitivity, double velocityScaleFactor)
     : _sensitivity(sensitivity)
     , _globalRotationMouseState(velocityScaleFactor)
     , _localRotationMouseState(velocityScaleFactor)
@@ -50,7 +51,7 @@ OrbitalInteractionMode::MouseStates::MouseStates(double sensitivity, double velo
 {
 }
 
-void OrbitalInteractionMode::MouseStates::updateMouseStatesFromInput(const InputState& inputState, double deltaTime) {
+void OrbitalNavigator::MouseStates::updateMouseStatesFromInput(const InputState& inputState, double deltaTime) {
     glm::dvec2 mousePosition = inputState.getMousePosition();
 
     bool button1Pressed = inputState.isMouseButtonPressed(MouseButton::Button1);
@@ -121,25 +122,25 @@ void OrbitalInteractionMode::MouseStates::updateMouseStatesFromInput(const Input
     }
 }
 
-void OrbitalInteractionMode::MouseStates::setRotationalFriction(double friction) {
+void OrbitalNavigator::MouseStates::setRotationalFriction(double friction) {
     _localRotationMouseState.setFriction(friction);
     _localRollMouseState.setFriction(friction);
     _globalRollMouseState.setFriction(friction);
 }
 
-void OrbitalInteractionMode::MouseStates::setHorizontalFriction(double friction) {
+void OrbitalNavigator::MouseStates::setHorizontalFriction(double friction) {
     _globalRotationMouseState.setFriction(friction);
 }
 
-void OrbitalInteractionMode::MouseStates::setVerticalFriction(double friction) {
+void OrbitalNavigator::MouseStates::setVerticalFriction(double friction) {
     _truckMovementMouseState.setFriction(friction);
 }
 
-void OrbitalInteractionMode::MouseStates::setSensitivity(double sensitivity) {
+void OrbitalNavigator::MouseStates::setSensitivity(double sensitivity) {
     _sensitivity = sensitivity;
 }
 
-void OrbitalInteractionMode::MouseStates::setVelocityScaleFactor(double scaleFactor) {
+void OrbitalNavigator::MouseStates::setVelocityScaleFactor(double scaleFactor) {
     _globalRotationMouseState.setVelocityScaleFactor(scaleFactor);
     _localRotationMouseState.setVelocityScaleFactor(scaleFactor);
     _truckMovementMouseState.setVelocityScaleFactor(scaleFactor);
@@ -147,30 +148,29 @@ void OrbitalInteractionMode::MouseStates::setVelocityScaleFactor(double scaleFac
     _globalRollMouseState.setVelocityScaleFactor(scaleFactor);
 }
 
-glm::dvec2 OrbitalInteractionMode::MouseStates::synchedGlobalRotationMouseVelocity() {
+glm::dvec2 OrbitalNavigator::MouseStates::synchedGlobalRotationMouseVelocity() {
     return _globalRotationMouseState.velocity.get();
 }
 
[…]
