### 2017-07-06 810c2ec
Enable PerformanceManager logging with ImGui checkbox.
