### 2017-07-06 002a2a0
Move properties from interaction handler to orbital navigator
