+++ src/interaction/orbitalinteractionmode.cpp
@@ -168,9 +168,18 @@ glm::dvec2 OrbitalNavigator::MouseStates::synchedGlobalRollMouseVelocity() {
     return _globalRollMouseState.velocity.get();
 }
 
-OrbitalNavigator::OrbitalNavigator(
-    std::shared_ptr<MouseStates> mouseStates)
-    : _mouseStates(mouseStates)
+OrbitalNavigator::OrbitalNavigator()
+    : properties::PropertyOwner("OrbitalNavigator")
+    , _rotationalFriction("rotationalFriction", "Rotational friction", true)
+    , _horizontalFriction("horizontalFriction", "Horizontal friction", true)
+    , _verticalFriction("verticalFriction", "Vertical friction", true)
+    , _followFocusNodeRotationDistance("followFocusNodeRotationDistance",
+        "Follow focus node rotation distance", 2.0f, 0.0f, 10.f)
+    , _minimumAllowedDistance("minimumAllowedDistance",
+        "Minimum allowed distance", 10.0f, 0.0f, 10000.f)
+    , _sensitivity("sensitivity", "Sensitivity", 20.0f, 1.0f, 50.f)
+    , _motionLag("motionLag", "Motion lag", 0.5f, 0.f, 1.f)
+    , _mouseStates(_sensitivity * pow(10.0,-4), 1 / (_motionLag + 0.0000001))
 {
 	_followFocusNodeRotationDistance = 2.0;
 	_minimumAllowedDistance = 10.0;
@@ -208,6 +217,32 @@ OrbitalNavigator::OrbitalNavigator(
             return 1 / (1 - t);
         };
     _rotateToFocusNodeInterpolator.setTransferFunction(smoothStepDerivedTranferFunction);
+
+    // Define lambda functions for changed properties
+    _rotationalFriction.onChange([&]() {
+        _mouseStates.setRotationalFriction(_rotationalFriction);
+    });
+    _horizontalFriction.onChange([&]() {
+        _mouseStates.setHorizontalFriction(_horizontalFriction);
+    });
+    _verticalFriction.onChange([&]() {
+        _mouseStates.setVerticalFriction(_verticalFriction);
+    });
+    _sensitivity.onChange([&]() {
+        _mouseStates.setSensitivity(_sensitivity * pow(10.0,-4));
+    });
+    _motionLag.onChange([&]() {
+        _mouseStates.setVelocityScaleFactor(1 / (_motionLag + 0.0000001));
+    });
+
+    // Add the properties
+    addProperty(_rotationalFriction);
+    addProperty(_horizontalFriction);
+    addProperty(_verticalFriction);
+    addProperty(_followFocusNodeRotationDistance);
+    addProperty(_minimumAllowedDistance);
+    addProperty(_sensitivity);
+    addProperty(_motionLag);
 }
 
 OrbitalNavigator::~OrbitalNavigator() {
@@ -246,7 +281,7 @@ OrbitalNavigator::CameraRotationDecomposition
 
 void OrbitalNavigator::performRoll(double deltaTime, glm::dquat& localCameraRotation) {
     glm::dquat rollQuat = glm::angleAxis(
-        _mouseStates->synchedLocalRollMouseVelocity().x * deltaTime,
+        _mouseStates.synchedLocalRollMouseVelocity().x * deltaTime,
         glm::dvec3(0, 0, 1)
     );
     localCameraRotation = localCameraRotation * rollQuat;
@@ -254,8 +289,8 @@ void OrbitalNavigator::performRoll(double deltaTime, glm::dquat& localCameraRota
 
 void OrbitalNavigator::performLocalRotation(double deltaTime, glm::dquat& localCameraRotation) {
[…]
