+++ src/interaction/interactionhandler.cpp
@@ -59,12 +59,6 @@ namespace {
     const char* MainTemplateFilename = "${OPENSPACE_DATA}/web/keybindings/main.hbs";
     const char* KeybindingTemplateFilename = "${OPENSPACE_DATA}/web/keybindings/keybinding.hbs";
     const char* JsFilename = "${OPENSPACE_DATA}/web/keybindings/script.js";
-    
-    const int IdOrbitalInteractionMode = 0;
-    const char* KeyOrbitalInteractionMode = "Orbital";
-    
-    const int IdKeyframeInteractionMode = 2;
-    const char* KeyKeyframeInteractionMode = "Keyframe";
 } // namespace
 
 #include "interactionhandler_lua.inl"
@@ -85,21 +79,7 @@ InteractionHandler::InteractionHandler()
         JsFilename
     )
     , _origin("origin", "Origin", "")
-    , _rotationalFriction("rotationalFriction", "Rotational friction", true)
-    , _horizontalFriction("horizontalFriction", "Horizontal friction", true)
-    , _verticalFriction("verticalFriction", "Vertical friction", true)
-    , _followFocusNodeRotationDistance("followFocusNodeRotationDistance",
-        "Follow focus node rotation distance", 2.0f, 0.0f, 10.f)
-    , _minimumAllowedDistance("minimumAllowedDistance",
-        "Minimum allowed distance", 10.0f, 0.0f, 10000.f)
-    , _sensitivity("sensitivity", "Sensitivity", 20.0f, 1.0f, 50.f)
-    , _motionLag("motionLag", "Motion lag", 0.5f, 0.f, 1.f)
     , _useKeyFrameInteraction("useKeyFrameInteraction", "Use keyframe interaction", false)
-    , _interactionModeOption(
-          "interactionMode",
-          "Interaction Mode",
-          properties::OptionProperty::DisplayType::Dropdown
-      )
 {
     _origin.onChange([this]() {
         SceneGraphNode* node = sceneGraphNode(_origin.value());
@@ -111,67 +91,14 @@ InteractionHandler::InteractionHandler()
         resetCameraDirection();
     });
 
-    // Create the interactionModes
     _inputState = std::make_unique<InputState>();
-    // Inject the same mouse states to both orbital and global interaction mode
-    _mouseStates = std::make_unique<OrbitalNavigator::MouseStates>(
-        _sensitivity * pow(10.0,-4), 1 / (_motionLag + 0.0000001));
-
-    _orbitalInteractionMode = std::make_unique<OrbitalNavigator>(_mouseStates);
+    _orbitalNavigator = std::make_unique<OrbitalNavigator>();
     _keyframeInteractionMode = std::make_unique<KeyframeInteractionMode>();
-
-    _interactionModeOption.addOption(IdOrbitalInteractionMode, KeyOrbitalInteractionMode);
-    _interactionModeOption.addOption(IdKeyframeInteractionMode, KeyKeyframeInteractionMode);
-
-    // Define lambda functions for changed properties
-    _rotationalFriction.onChange([&]() {
-        _mouseStates->setRotationalFriction(_rotationalFriction);
-    });
-    _horizontalFriction.onChange([&]() {
-        _mouseStates->setHorizontalFriction(_horizontalFriction);
-    });
-    _verticalFriction.onChange([&]() {
-        _mouseStates->setVerticalFriction(_verticalFriction);
-    });
-    _followFocusNodeRotationDistance.onChange([&]() {
-        _orbitalInteractionMode->setFollowFocusNodeRotationDistance(_followFocusNodeRotationDistance);
-    });
-    _minimumAllowedDistance.onChange([&]() {
-        _orbitalInteractionMode->setMinimumAllowedDistance(_minimumAllowedDistance);
-    });
[…]

[…]
