### 2017-07-06 e81b357
Use smooth step for follow rotation interpolator
