### 2017-07-06 d632cea
Rename KeyframeInteractionMode to KeyframeNavigator
