+++ src/network/parallelconnection.cpp
@@ -67,7 +67,7 @@
 #include <openspace/engine/openspaceengine.h>
 #include <openspace/engine/wrapper/windowwrapper.h>
 #include <openspace/interaction/interactionhandler.h>
-#include <openspace/interaction/interactionmode.h>
+#include <openspace/interaction/keyframenavigator.h>
 #include <openspace/scene/scenegraphnode.h>
 #include <openspace/scripting/script_helper.h>
 #include <openspace/util/time.h>
