+++ include/openspace/interaction/orbitalnavigator.h
@@ -28,6 +28,7 @@
 #include <openspace/interaction/delayedvariable.h>
 #include <openspace/interaction/inputstate.h>
 #include <openspace/interaction/interpolator.h>
+#include <openspace/interaction/mousestate.h>
 
 #include <openspace/properties/propertyowner.h>
 #include <openspace/properties/scalar/boolproperty.h>
@@ -36,73 +37,25 @@
 #include <glm/glm.hpp>
 #include <glm/gtx/quaternion.hpp>
 
-#include <tuple>
-
 namespace openspace {
 namespace interaction {
 
 class OrbitalNavigator : public properties::PropertyOwner 
 {
 public:
-	struct MouseState {
-		MouseState(double scaleFactor)
-			: velocity(scaleFactor, 1)
-			, previousPosition(0.0, 0.0) {}
-		void setFriction(double friction) {
-			velocity.setFriction(friction);
-		}
-		void setVelocityScaleFactor(double scaleFactor) {
-			velocity.setScaleFactor(scaleFactor);
-		}
-		glm::dvec2 previousPosition;
-		DelayedVariable<glm::dvec2, double> velocity;
-	};
-
-    class MouseStates
-    {
-    public:
-        /**
-        \param sensitivity
-        \param velocityScaleFactor can be set to 60 to remove the inertia of the
-        interaction. Lower value will make it harder to move the camera. 
-        */
-        MouseStates(double sensitivity, double velocityScaleFactor);
-        void updateMouseStatesFromInput(const InputState& inputState, double deltaTime);
-        void setRotationalFriction(double friction);
-        void setHorizontalFriction(double friction);
-        void setVerticalFriction(double friction);
-        void setSensitivity(double sensitivity);
-        void setVelocityScaleFactor(double scaleFactor);
-
-        glm::dvec2 synchedGlobalRotationMouseVelocity();
-        glm::dvec2 synchedLocalRotationMouseVelocity();
-        glm::dvec2 synchedTruckMovementMouseVelocity();
-        glm::dvec2 synchedLocalRollMouseVelocity();
-        glm::dvec2 synchedGlobalRollMouseVelocity();
-
-    private:
-        double _sensitivity;
-
-        MouseState _globalRotationMouseState;
-        MouseState _localRotationMouseState;
-        MouseState _truckMovementMouseState;
-        MouseState _localRollMouseState;
-        MouseState _globalRollMouseState;
-    };
-
     OrbitalNavigator();
     ~OrbitalNavigator();
 
-    void updateMouseStatesFromInput(const InputState& inputState, double deltaTime);
+    void updateMouseStatesFromInput(const InputState& inputState,
+                                    double deltaTime);
     void updateCameraStateFromMouseStates(Camera& camera, double deltaTime);
-    bool followingNodeRotation() const;
-
     void setFocusNode(SceneGraphNode* focusNode);
-
-    SceneGraphNode* focusNode();
     void startInterpolateCameraDirection(const Camera& camera);
 
-protected:
+    bool followingNodeRotation() const;    
+    SceneGraphNode* focusNode() const;
+
+private:
     struct CameraRotationDecomposition {
         glm::dquat localRotation;
         glm::dquat globalRotation;
@@ -135,35 +88,25 @@ protected:
     void performRoll(double deltaTime, glm::dquat& localCameraRotation);
     void performLocalRotation(double deltaTime, glm::dquat& localCameraRotation);
     void interpolateLocalRotation(double deltaTime, glm::dquat& localCameraRotation);
-    void performHorizontalTranslationAndRotation(
-        double deltaTime,
-        glm::dvec3 objectPosition,
-        glm::dvec3& cameraPosition,
[…]
