+++ src/interaction/orbitalnavigator.cpp
@@ -40,134 +40,6 @@ namespace {
 namespace openspace {
 namespace interaction {
 
-// OrbitalNavigator
-OrbitalNavigator::MouseStates::MouseStates(double sensitivity, double velocityScaleFactor)
-    : _sensitivity(sensitivity)
-    , _globalRotationMouseState(velocityScaleFactor)
-    , _localRotationMouseState(velocityScaleFactor)
-    , _truckMovementMouseState(velocityScaleFactor)
-    , _localRollMouseState(velocityScaleFactor)
-    , _globalRollMouseState(velocityScaleFactor)
-{
-}
-
-void OrbitalNavigator::MouseStates::updateMouseStatesFromInput(const InputState& inputState, double deltaTime) {
-    glm::dvec2 mousePosition = inputState.getMousePosition();
-
-    bool button1Pressed = inputState.isMouseButtonPressed(MouseButton::Button1);
-    bool button2Pressed = inputState.isMouseButtonPressed(MouseButton::Button2);
-    bool button3Pressed = inputState.isMouseButtonPressed(MouseButton::Button3);
-    bool keyCtrlPressed = inputState.isKeyPressed(Key::LeftControl);
-    bool keyShiftPressed = inputState.isKeyPressed(Key::LeftShift);
-    
-    // Update the mouse states
-    if (button1Pressed && !keyShiftPressed) {
-        if (keyCtrlPressed) {
-            glm::dvec2 mousePositionDelta =
-                _localRotationMouseState.previousPosition - mousePosition;
-            _localRotationMouseState.velocity.set(mousePositionDelta * _sensitivity, deltaTime);
-
-            _globalRotationMouseState.previousPosition = mousePosition;
-            _globalRotationMouseState.velocity.decelerate(deltaTime);
-        }
-        else {
-            glm::dvec2 mousePositionDelta =
-                _globalRotationMouseState.previousPosition - mousePosition;
-            _globalRotationMouseState.velocity.set(mousePositionDelta * _sensitivity, deltaTime);
-
-            _localRotationMouseState.previousPosition = mousePosition;
-            _localRotationMouseState.velocity.decelerate(deltaTime);
-        }
-    }
-    else { // !button1Pressed
-        _localRotationMouseState.previousPosition = mousePosition;
-        _localRotationMouseState.velocity.decelerate(deltaTime);
-
-        _globalRotationMouseState.previousPosition = mousePosition;
-        _globalRotationMouseState.velocity.decelerate(deltaTime);
-    }
-    if (button2Pressed) {
-        glm::dvec2 mousePositionDelta =
-            _truckMovementMouseState.previousPosition - mousePosition;
-        _truckMovementMouseState.velocity.set(mousePositionDelta * _sensitivity, deltaTime);
-    }
-    else { // !button2Pressed
-        _truckMovementMouseState.previousPosition = mousePosition;
-        _truckMovementMouseState.velocity.decelerate(deltaTime);
-    }
-    if (button3Pressed || (keyShiftPressed && button1Pressed)) {
-        if (keyCtrlPressed) {
-            glm::dvec2 mousePositionDelta =
-                _localRollMouseState.previousPosition - mousePosition;
-            _localRollMouseState.velocity.set(mousePositionDelta * _sensitivity, deltaTime);
-
-            _globalRollMouseState.previousPosition = mousePosition;
-            _globalRollMouseState.velocity.decelerate(deltaTime);
-        }
-        else {
-            glm::dvec2 mousePositionDelta =
[…]
