### 2017-07-06 646674b
Separate mousestate from orbitalnavigator
