+++ src/interaction/mousestate.cpp
@@ -1,0 +1,162 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#include <openspace/interaction/mousestate.h>
+
+namespace {
+    const std::string _loggerCat = "MouseState";
+}
+
+namespace openspace {
+namespace interaction {
+
+MouseStates::MouseStates(double sensitivity, double velocityScaleFactor)
+    : _sensitivity(sensitivity)
+    , _globalRotationMouseState(velocityScaleFactor)
+    , _localRotationMouseState(velocityScaleFactor)
+    , _truckMovementMouseState(velocityScaleFactor)
+    , _localRollMouseState(velocityScaleFactor)
+    , _globalRollMouseState(velocityScaleFactor)
+{
+}
+
+void MouseStates::updateMouseStatesFromInput(const InputState& inputState, double deltaTime) {
+    glm::dvec2 mousePosition = inputState.getMousePosition();
+
+    bool button1Pressed = inputState.isMouseButtonPressed(MouseButton::Button1);
+    bool button2Pressed = inputState.isMouseButtonPressed(MouseButton::Button2);
+    bool button3Pressed = inputState.isMouseButtonPressed(MouseButton::Button3);
+    bool keyCtrlPressed = inputState.isKeyPressed(Key::LeftControl);
+    bool keyShiftPressed = inputState.isKeyPressed(Key::LeftShift);
+    
+    // Update the mouse states
+    if (button1Pressed && !keyShiftPressed) {
+        if (keyCtrlPressed) {
+            glm::dvec2 mousePositionDelta =
+                _localRotationMouseState.previousPosition - mousePosition;
+            _localRotationMouseState.velocity.set(mousePositionDelta * _sensitivity, deltaTime);
+
[…]
