### 2017-07-06 4d4a2f6
[WebGui] Add initial data management api
