### 2017-07-06 351a246
Add support for c closures in lua
