### 2017-07-06 58a423f
Loading and unloading assets
