+++ src/engine/settingsengine.cpp
@@ -63,7 +63,7 @@ SettingsEngine::SettingsEngine()
 void SettingsEngine::initialize() {
     // Load all matching files in the Scene
     // TODO: match regex with either with new ghoul readFiles or local code
-    const std::string sceneDir = "${SCENE}";
+    const std::string sceneDir = "${ASSETS}";
     const std::vector<std::string> scenes = ghoul::filesystem::Directory(sceneDir).readFiles();
     const size_t nScenes = scenes.size();
 
