+++ src/scene/assetloader.cpp
@@ -23,25 +23,30 @@
  ****************************************************************************************/
 
 #include <openspace/scene/assetloader.h>
+
+#include <openspace/scripting/script_helper.h>
+
 #include <ghoul/lua/ghoul_lua.h>
 #include <ghoul/lua/luastate.h>
 #include <ghoul/lua/lua_helper.h>
 #include <ghoul/logging/logmanager.h>
 #include <ghoul/misc/onscopeexit.h>
-
 #include <ghoul/filesystem/filesystem.h>
+
+#include "assetloader_lua.inl"
 
 namespace {
     const char* AssetFileSuffix = "asset";
     const char* AssetGlobalVariableName = "asset";
+    const char* DataFileSuffix = "data";
 
     const char* ImportFunctionName = "import";
     const char* SyncedResourceFunctionName = "syncedResource";
     const char* LocalResourceFunctionName = "localResource";
+    const char* OnInitializeFunctionName = "onInitialize";
     const char* OnDeinitializeFunctionName = "onDeinitialize";
-    
+
     const char* AssetsTableName = "_assets";
-    const char* KeySceneGraphNodes = "SceneGraphNodes";
 
     const char* _loggerCat = "AssetLoader";
 }
@@ -49,6 +54,7 @@ namespace {
 namespace openspace {
 
 namespace assetloader {
+
 int importAsset(lua_State* state) {
     AssetLoader *assetLoader = (AssetLoader*)lua_touserdata(state, lua_upvalueindex(1));
     return assetLoader->importAssetLua();
@@ -68,55 +74,79 @@ int noOperation(lua_State* state) {
     return 0;
 }
 
-
-}
+} // namespace assetloader
 
 AssetLoader::AssetLoader(ghoul::lua::LuaState* luaState, std::string assetRoot, std::string syncRoot)
     : _luaState(luaState)
     , _rootAsset(std::make_unique<Asset>(this, std::move(assetRoot)))
     , _syncRoot(std::move(syncRoot))
 {
-    _assetStack.push_back(_rootAsset.get());
+    pushAsset(_rootAsset.get());
 
     // Create _assets table.
     lua_newtable(*_luaState);
     lua_setglobal(*_luaState, AssetsTableName);
 }
 
-AssetLoader::Asset* AssetLoader::loadAsset(const std::string& name) {
+AssetLoader::Asset* AssetLoader::importAsset(const std::string& name) {
     ghoul::filesystem::Directory directory = currentDirectory();
-    std::unique_ptr<Asset> asset = std::make_unique<Asset>(this, directory, name);
-    const std::string id = asset->id();
+    std::unique_ptr<Asset> newAsset = std::make_unique<Asset>(this, directory, name);
+    const std::string id = newAsset->id();
+
+    Asset* asset = newAsset.get();
 
     // Check if asset is already loaded.
-    if (_loadedAssets.find(id) != _loadedAssets.end()) {
-        return _loadedAssets[id].get();
+    const auto it = _importedAssets.find(id);
+    const bool loaded = it != _importedAssets.end();
+    if (loaded) {
+        asset = it->second.get();
     }
 
     // Create dependency link
     Asset* dependant = _assetStack.back();
-    dependant->addDependency(asset.get());
-    asset->addDependant(dependant);
+    dependant->addDependency(asset);
+
+    if (loaded) {
+        return asset;
+    }
 
     // Actually loading asset.
-    pushAsset(asset.get());
+    pushAsset(asset);
     ghoul::OnScopeExit e([this]() {
         popAsset();
     });
 
     try {
-        ghoul::lua::runScriptFile(*_luaState, asset->assetFilepath());
[…]
