### 2017-07-07 a165af8
Output to files. Function files aren't writing for some reason.
