### 2017-07-07 01242c4
Split keybindingmanager from interactionhandler interactionhandler
