+++ include/openspace/interaction/keybindingmanager.h
@@ -1,0 +1,90 @@
+/*****************************************************************************************
+ *                                                                                       *
+ * OpenSpace                                                                             *
+ *                                                                                       *
+ * Copyright (c) 2014-2017                                                               *
+ *                                                                                       *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
+ * software and associated documentation files (the "Software"), to deal in the Software *
+ * without restriction, including without limitation the rights to use, copy, modify,    *
+ * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
+ * permit persons to whom the Software is furnished to do so, subject to the following   *
+ * conditions:                                                                           *
+ *                                                                                       *
+ * The above copyright notice and this permission notice shall be included in all copies *
+ * or substantial portions of the Software.                                              *
+ *                                                                                       *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
+ * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
+ * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
+ * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
+ ****************************************************************************************/
+
+#ifndef __OPENSPACE_CORE___KEYBINDINGMANAGER___H__
+#define __OPENSPACE_CORE___KEYBINDINGMANAGER___H__
+
+#include <openspace/documentation/documentationgenerator.h>
+#include <openspace/scripting/lualibrary.h>
+#include <openspace/util/keys.h>
+
+#include <ghoul/misc/boolean.h>
+
+namespace openspace {
+
+class Camera;
+class SceneGraphNode;
+
+namespace interaction {
+
+class KeyBindingManager : public DocumentationGenerator
+{
+public:
+    KeyBindingManager();
+    ~KeyBindingManager();
+
+    void initialize();
+    void deinitialize();
+    
+    void resetKeyBindings();
+
+    void bindKeyLocal(
+        Key key,
+        KeyModifier modifier,
+        std::string luaCommand,
+        std::string documentation = ""
+    );
+
+    void bindKey(
+        Key key,
+        KeyModifier modifier,
+        std::string luaCommand,
+        std::string documentation = ""
+    );
+
+    static scripting::LuaLibrary luaLibrary();
+
+    // Callback functions 
+    void keyboardCallback(Key key, KeyModifier modifier, KeyAction action);
+
+private:
+    using Synchronized = ghoul::Boolean;
+
+    struct KeyInformation {
+        std::string command;
+        Synchronized synchronization;
+        std::string documentation;
+    };
+    
+    std::string generateJson() const override;
+
+    bool _cameraUpdatedFromScript = false;
+
+    std::multimap<KeyWithModifier, KeyInformation> _keyLua;
+};
+
+} // namespace interaction
+} // namespace openspace
+
+#endif // __OPENSPACE_CORE___KEYBINDINGMANAGER___H__
