### 2017-07-07 df9fd3e
Rename interactionhandler to navigationhandler
