+++ include/openspace/engine/openspaceengine.h
@@ -63,7 +63,7 @@ class TimeManager;
 class WindowWrapper;
 
 namespace interaction {
-    class InteractionHandler;
+    class NavigationHandler;
     class KeyBindingManager;
 }
 namespace gui { class GUI; }
@@ -126,7 +126,7 @@ public:
     TimeManager& timeManager();
     WindowWrapper& windowWrapper();
     ghoul::fontrendering::FontManager& fontManager();
-    interaction::InteractionHandler& interactionHandler();
+    interaction::NavigationHandler& navigationHandler();
     interaction::KeyBindingManager& keyBindingManager();
     properties::PropertyOwner& globalPropertyOwner();
     scripting::ScriptEngine& scriptEngine();
@@ -199,7 +199,7 @@ private:
     std::unique_ptr<WindowWrapper> _windowWrapper;
     std::unique_ptr<ghoul::cmdparser::CommandlineParser> _commandlineParser;
     std::unique_ptr<ghoul::fontrendering::FontManager> _fontManager;
-    std::unique_ptr<interaction::InteractionHandler> _interactionHandler;
+    std::unique_ptr<interaction::NavigationHandler> _navigationHandler;
     std::unique_ptr<interaction::KeyBindingManager> _keyBindingManager;
     std::unique_ptr<scripting::ScriptEngine> _scriptEngine;
     std::unique_ptr<scripting::ScriptScheduler> _scriptScheduler;
