+++ include/openspace/interaction/delayedvariable.h
@@ -25,6 +25,8 @@
 #ifndef __OPENSPACE_CORE___DELAYEDVARIABLE___H__
 #define __OPENSPACE_CORE___DELAYEDVARIABLE___H__
 
+#include <ghoul/glm.h>
+
 namespace openspace {
 namespace interaction {
 
@@ -44,12 +46,12 @@ public:
     void set(T value, double dt) {
         _targetValue = value;
         _currentValue = _currentValue + (_targetValue - _currentValue) *
-            std::min(_scaleFactor * dt, 1.0); // less or equal to 1.0 keeps it stable
+			glm::min(_scaleFactor * dt, 1.0); // less or equal to 1.0 keeps it stable
     }
     
     void decelerate(double dt) {
         _currentValue = _currentValue + (- _currentValue) *
-            std::min(_scaleFactor * _friction * dt, 1.0); // less or equal to 1.0 keeps it stable
+            glm::min(_scaleFactor * _friction * dt, 1.0); // less or equal to 1.0 keeps it stable
     }
     
     void setHard(T value) {
