+++ modules/globebrowsing/geometry/geodeticpatch.h
@@ -27,14 +27,14 @@
 
 #include <modules/globebrowsing/geometry/geodetic2.h>
 #include <modules/globebrowsing/tile/quad.h>
+#include <modules/globebrowsing/tile/tileindex.h>
 
 namespace openspace {
 namespace globebrowsing {
-    
-struct TileIndex;
 
 class GeodeticPatch {
 public:
+    GeodeticPatch(int hej);
     GeodeticPatch(
         double centerLat,
         double centerLon,
