### 2017-07-07 15dc2c3
Fix bugs in scene graph manipulation API
