### 2017-07-07 dbe9c23
Move functionality from navigationhandler to keyframenavigator
