### 2017-07-07 345cce7
Update scripts for navigation
