+++ src/interaction/orbitalnavigator.cpp
@@ -49,7 +49,7 @@ namespace openspace {
 namespace interaction {
 
 OrbitalNavigator::OrbitalNavigator()
-    : properties::PropertyOwner("orbitalNavigator")
+    : properties::PropertyOwner("OrbitalNavigator")
     , _rotationalFriction("rotationalFriction", "Rotational friction", true)
     , _horizontalFriction("horizontalFriction", "Horizontal friction", true)
     , _verticalFriction("verticalFriction", "Vertical friction", true)
@@ -158,7 +158,7 @@ void OrbitalNavigator::updateCameraStateFromMouseStates(Camera& camera,
 
         focusNodeRotationDiff = interpolateRotationDifferential(
             deltaTime,
-            1.0,
+            5.0,
             focusNodeRotationDiff,
             centerPos,
             camPos);
@@ -173,24 +173,24 @@ void OrbitalNavigator::updateCameraStateFromMouseStates(Camera& camera,
 
         performHorizontalTranslation(
             deltaTime,
+            camPos,
             centerPos,
             focusNodeRotationDiff,
-            camPos,
             camRot.globalRotation);
 
         followFocusNodeRotation(
+            camPos,
+            centerPos,
+            focusNodeRotationDiff);
+
+        performGlobalRotation(
+            camRot.globalRotation,
             centerPos,
             focusNodeRotationDiff,
             camPos);
 
-        performGlobalRotation(
-            centerPos,
-            focusNodeRotationDiff,
-            camPos,
-            camRot.globalRotation);
-
-        performVerticalTranslation(deltaTime, centerPos, camPos);
-        performHorizontalRotation(deltaTime, camPos, camRot.globalRotation);
+        performVerticalTranslation(deltaTime, camPos, centerPos);
+        performHorizontalRotation(deltaTime, camRot.globalRotation, camPos);
         pushToSurface(_minimumAllowedDistance, centerPos, camPos);
 
         // Update the camera state
@@ -232,10 +232,10 @@ SceneGraphNode* OrbitalNavigator::focusNode() const {
 
 OrbitalNavigator::CameraRotationDecomposition
 	OrbitalNavigator::decomposeCameraRotation(
-		glm::dvec3 cameraPosition,
-		glm::dquat cameraRotation,
-		glm::dvec3 cameraLookUp,
-		glm::dvec3 cameraViewDirection)
+		const glm::dvec3& cameraPosition,
+		const glm::dquat& cameraRotation,
+		const glm::dvec3& cameraLookUp,
+		const glm::dvec3& cameraViewDirection)
 {
     glm::dmat4 inverseModelTransform = _focusNode->inverseModelTransform();
     glm::dmat4 modelTransform = _focusNode->modelTransform();
@@ -294,10 +294,10 @@ void OrbitalNavigator::interpolateLocalRotation(double deltaTime, glm::dquat& lo
 
 void OrbitalNavigator::performHorizontalTranslation(
     double deltaTime,
-    glm::dvec3 objectPosition,
-    glm::dquat& focusNodeRotationDiff,
     glm::dvec3& cameraPosition,
-    glm::dquat& globalCameraRotation)
+    const glm::dvec3& objectPosition,
+    const glm::dquat& focusNodeRotationDiff,
+    const glm::dquat& globalCameraRotation)
 {
 	using namespace glm;
     // Uniform variables
@@ -355,9 +355,9 @@ void OrbitalNavigator::performHorizontalTranslation(
 }
 
 void OrbitalNavigator::followFocusNodeRotation(
[…]
