### 2017-07-07 0fe71b5
Move more scene loading logic to assets
