### 2017-07-07 115bb99
Remove colons and cat filename together with formatLogName.
