### 2017-07-07 29fd158
Create BoolProperty for PerformanceComponent
