### 2017-07-07 f1d468d
Appending to logs now instead of overwriting :-/
