### 2017-07-07 5213733
[WebGui] Add some minor fixes to TimePicker and client subscriptions
