### 2017-07-08 0f34b05
Increasing warning level
