### 2017-07-08 OpenSpace/OpenSpace#352 @MatthewTurk
Some issues with Linux installation instructions
Hi!  As part of https://github.com/openjournals/joss-reviews/issues/281 I'm attempting to follow the linux compilation instructions:

https://github.com/OpenSpace/OpenSpace/wiki/Guides-Compile-OpenSpace-on-Linux

It didn't quite work as described.  The `git clone` command returned that there was no upstream `develop` branch; I think it should be `master`.

A few other minor things I had to do -- I wasn't able to get cmake to be able to detect the right C++17 options; this is not an OpenSpace issue, but it might be helpful to suggest that if an error of the sort:

```
 CMake Error in CMakeLists.txt:
   Target "Launcher" requires the language dialect "CXX17" (with compiler
   extensions), but CMake does not know the compile flags to use to enable it.
```

shows up, to set the CMake variable `CMAKE_CXX_FLAGS` to `-std=c++17`.  This mostly seems to correct the issue, but I am independently attempting to determine why CMake is struggling.

Finally, Ghoul was giving some errors (specifically, arguments like `-Walloc-zero` not being recognized), which look like they are related to GCC 7 features.  I'm not sure where to report those as problems, or if perhaps something has been incorrectly set up in my git submodules.
