+++ modules/onscreengui/src/guiorigincomponent.cpp
@@ -70,7 +70,7 @@ void GuiOriginComponent::render() {
     bool hasChanged = ImGui::Combo("Origin", &currentPosition, nodeNames.c_str());
     if (hasChanged) {
         OsEng.scriptEngine().queueScript(
-            "openspace.setPropertyValue('Interaction.origin', '" +
+            "openspace.setPropertyValue('NavigationHandler.origin', '" +
             nodes[currentPosition]->name() + "');",
             scripting::ScriptEngine::RemoteScripting::Yes
         );
