+++ src/interaction/orbitalnavigator.cpp
@@ -87,13 +87,9 @@ OrbitalNavigator::OrbitalNavigator()
         [](double t) {
             return (6 * (t + t*t) / (1 - 3 * t*t + 2 * t*t*t));
         };
-    auto linearDerivedTranferFunction = 
-        [](double t) {
-            return 1 / (1 - t);
-        };
     _rotateToFocusNodeInterpolator.setTransferFunction(smoothStepDerivedTranferFunction);
 
-    // Define lambda functions for changed properties
+    // Define callback functions for changed properties
     _rotationalFriction.onChange([&]() {
         _mouseStates.setRotationalFriction(_rotationalFriction);
     });
@@ -120,9 +116,8 @@ OrbitalNavigator::OrbitalNavigator()
     addProperty(_motionLag);
 }
 
-OrbitalNavigator::~OrbitalNavigator() {
-
-}
+OrbitalNavigator::~OrbitalNavigator()
+{ }
 
 void OrbitalNavigator::updateMouseStatesFromInput(const InputState& inputState,
                                                   double deltaTime)
@@ -134,64 +129,89 @@ void OrbitalNavigator::updateCameraStateFromMouseStates(Camera& camera,
                                                         double deltaTime)
 {
     if (_focusNode) {
-        using namespace glm;
-        // Read the current state of the camera and focusnode
-        dvec3 camPos = camera.positionVec3();
-        dvec3 centerPos = _focusNode->worldPosition();
+        // Read the current state of the camera
+        glm::dvec3 camPos = camera.positionVec3();
+        glm::dvec3 centerPos = _focusNode->worldPosition();
 
         // Follow focus nodes movement
-        dvec3 focusNodeDiff = centerPos - _previousFocusNodePosition;
+        glm::dvec3 focusNodeDiff = centerPos - _previousFocusNodePosition;
         _previousFocusNodePosition = centerPos;
         camPos += focusNodeDiff;
 
+        // Calculate a position handle based on the camera position in world space
+        SurfacePositionHandle posHandle = calculateSurfacePositionHandle(camPos);
+
+        // Decompose camera rotation so that we can handle global and local rotation
+        // individually. Then we combine them again when finished.
         CameraRotationDecomposition camRot = decomposeCameraRotation(
             camPos,
             camera.rotationQuaternion(),
             camera.lookUpVectorWorldSpace(),
             camera.viewDirectionWorldSpace());
 
-        // Rotate with the globe
-        dmat3 globeStateMatrix = _focusNode->worldRotationMatrix();
-        dquat globeRotation = quat_cast(globeStateMatrix);
-        dquat focusNodeRotationDiff = _previousFocusNodeRotation * inverse(globeRotation);
-        _previousFocusNodeRotation = globeRotation;
+        // Rotate with the globe by finding a differential rotation from the previous
+        // to the current rotation
+        glm::dmat3 globeStateMatrix = _focusNode->worldRotationMatrix();
+        glm::dquat objectRotation = glm::quat_cast(globeStateMatrix);
+        glm::dquat focusNodeRotationDiff =
+            _previousFocusNodeRotation * glm::inverse(objectRotation);
+        _previousFocusNodeRotation = objectRotation;
 
+        // Interpolate rotation differential so that the camera rotates with the object
+        // only if close enough
[…]

[…]
