### 2017-07-10 fe938b4
Solve but that caused NaN values for camera position when being in center of globe and setting focus to the globe.
