### 2017-07-10 5ef5db1
Introduce asset toggles
