+++ src/scene/assetloader.cpp
@@ -41,12 +41,14 @@ namespace {
     const char* DataFileSuffix = "data";
 
     const char* ImportFunctionName = "import";
+    const char* ImportToggleFunctionName = "importToggle";
     const char* SyncedResourceFunctionName = "syncedResource";
     const char* LocalResourceFunctionName = "localResource";
     const char* OnInitializeFunctionName = "onInitialize";
     const char* OnDeinitializeFunctionName = "onDeinitialize";
 
     const char* AssetsTableName = "_assets";
+    const char* TogglesTableName = "_toggles";
 
     const char* _loggerCat = "AssetLoader";
 }
@@ -58,6 +60,11 @@ namespace assetloader {
 int importAsset(lua_State* state) {
     AssetLoader *assetLoader = (AssetLoader*)lua_touserdata(state, lua_upvalueindex(1));
     return assetLoader->importAssetLua();
+}
+
+int importAssetToggle(lua_State* state) {
+    AssetLoader *assetLoader = (AssetLoader*)lua_touserdata(state, lua_upvalueindex(1));
+    return assetLoader->importAssetToggleLua();
 }
 
 int resolveLocalResource(lua_State* state) {
@@ -88,7 +95,11 @@ AssetLoader::AssetLoader(ghoul::lua::LuaState* luaState, std::string assetRoot, 
     lua_setglobal(*_luaState, AssetsTableName);
 }
 
-AssetLoader::Asset* AssetLoader::importAsset(const std::string& name) {
+AssetLoader::Asset* AssetLoader::importAsset(
+    const std::string& name,
+    bool toggle,
+    bool toggleOn)
+{
     ghoul::filesystem::Directory directory = currentDirectory();
     std::unique_ptr<Asset> newAsset = std::make_unique<Asset>(this, directory, name);
     const std::string id = newAsset->id();
@@ -102,9 +113,18 @@ AssetLoader::Asset* AssetLoader::importAsset(const std::string& name) {
         asset = it->second.get();
     }
 
-    // Create dependency link
-    Asset* dependant = _assetStack.back();
-    dependant->addDependency(asset);
+    if (toggle) {
+        // This is a togglable dependency
+        Asset* toggler = _assetStack.back();
+        toggler->addToggle(asset);
+        if (toggleOn) {
+            toggler->setToggle(asset, true);
+        }
+    } else {
+        // This is a hard dependency
+        Asset* dependant = _assetStack.back();
+        dependant->addDependency(asset);
+    }
 
     if (loaded) {
         return asset;
@@ -121,7 +141,7 @@ AssetLoader::Asset* AssetLoader::importAsset(const std::string& name) {
     } catch (const ghoul::lua::LuaRuntimeException& e) {
         LERROR(e.message << ": " << e.component);
     }
-    LDEBUG("Loaded asset " << asset->id());
+    LDEBUG("Imported asset " << asset->id());
     
     _importedAssets.emplace(id, std::move(newAsset));
     return _importedAssets[id].get();
@@ -191,6 +211,11 @@ void AssetLoader::pushAsset(Asset* asset) {
     lua_pushcclosure(*_luaState, &assetloader::importAsset, 1);
     lua_setfield(*_luaState, assetTableIndex, ImportFunctionName);
 
+    // Register importToggle function
+    lua_pushlightuserdata(*_luaState, asset->loader());
+    lua_pushcclosure(*_luaState, &assetloader::importAssetToggle, 1);
+    lua_setfield(*_luaState, assetTableIndex, ImportToggleFunctionName);
+
     // Register default onDeinitialize function
     lua_pushcfunction(*_luaState, &assetloader::noOperation);
[…]
