### 2017-07-10 OpenSpace/OpenSpace#352 @MatthewTurk
I've given it a clean try, and I was able to compile with only a few changes to the compilation instructions.

I'm on Ubuntu 16.04.  I was able to get GCC 7 and the toolchain installed by:

```
sudo add-apt-repository ppa:ubuntu-toolchain-r/test
sudo apt-get update
sudo apt-get install gcc-7 g++-7
```

Setting the CMake variables:

```
CMAKE_CXX_COMPILER:FILEPATH=/usr/bin/g++-7
CMAKE_CXX_FLAGS:STRING=-std=gnu++17
CMAKE_C_COMPILER:FILEPATH=/usr/bin/gcc-7
```

I was able to get much further, but I'm still running into issues; I've created a gist of the [compilation log](https://gist.github.com/MatthewTurk/d07479804a499065b204cdd4c3c3e634).  Any help would be greatly appreciated -- this looks to me like it might be a `std` namespacing issue, but I'm not completely certain the correct way to address it.
