### 2017-07-10 2d16616
Replace ${SGCT} token with ${CONFIG} in Launcher and OpenSpaceEngine.
