### 2017-07-10 a82c873
Adds ${LOGS} path token and LogDir attribute
