+++ openspace.cfg
@@ -48,7 +48,8 @@ return {
         CONFIG = "${BASE_PATH}/config",
         CACHE = "${BASE_PATH}/cache",
         FONTS = "${OPENSPACE_DATA}/fonts",
-        DOCUMENTATION = "${BASE_PATH}/documentation"
+        DOCUMENTATION = "${BASE_PATH}/documentation",
+        LOGS = "${BASE_PATH}/logs"
     },
     Fonts = {
         Mono = "${FONTS}/Droid_Sans_Mono/DroidSansMono.ttf",
@@ -56,6 +57,7 @@ return {
         Console = "${FONTS}/Inconsolata/Inconsolata-Regular.ttf"
     },
     Logging = {
+        LogDir = "${LOGS}",
         -- LogLevel = "Trace",
         LogLevel = "Debug",
         ImmediateFlush = true,
