+++ openspace.cfg
@@ -10,7 +10,7 @@ return {
     SGCTConfig = sgct.config.single{},
 
     -- A regular 1920x1080 window
-    -- SGCTConfig = sgct.config.single{1920, 1080},   
+    -- SGCTConfig = sgct.config.single{1920, 1080},
 
     -- A 1k fisheye rendering
     -- SGCTConfig = sgct.config.fisheye{1024, 1024},
@@ -19,7 +19,7 @@ return {
     -- SGCTConfig = sgct.config.fisheye{1024, 1024, res={4096, 4096}, quality="2k", tilt=27},
 
     -- Streaming OpenSpace via Spout to OBS
-    -- SGCTConfig = sgct.config.single{2560, 1440, shared=true, name="WV_OBS_SPOUT1"},   
+    -- SGCTConfig = sgct.config.single{2560, 1440, shared=true, name="WV_OBS_SPOUT1"},
 
 
     --SGCTConfig = "${CONFIG}/openvr_oculusRiftCv1.xml",
