### 2017-07-10 4f903ac
Feature/globebrowsing (#334)
Layer support for globe browsing:

Add layers using the function openspace.globebrowsing.addLayer
Delete layers using openspace.globebrowsing.deleteLayer
Layer type does not necessarily have to be of tile type. For example solidcolor does not use tiles
Blend modes for layers are Normal, Add, Subtract, Multiply, Color
Layer adjustments to affect layers. The current only active one is chroma key to cut out a color from the layer. Transfer functions or clipping masks are examples of layer adjustments for the future.
Support for adding layer specifications for quickly 
[…]
