+++ modules/globebrowsing/other/prioritizingconcurrentjobmanager.h
@@ -43,7 +43,7 @@ namespace globebrowsing {
 template<typename P, typename KeyType>
 class PrioritizingConcurrentJobManager {
 public:
-    PrioritizingConcurrentJobManager(std::shared_ptr<LRUThreadPool<KeyType>> pool);
+    PrioritizingConcurrentJobManager(LRUThreadPool<KeyType> pool);
 
     /**
      * Enqueues a job which is identified using a given key
@@ -85,9 +85,10 @@ public:
 
 private:
     ConcurrentQueue<std::shared_ptr<Job<P>>> _finishedJobs;
-    /// An LRU thread pool is used since the jobs can be bumped and hence prioritized.
-    std::shared_ptr<LRUThreadPool<KeyType>> _threadPool;
     std::mutex _finishedJobsMutex;
+    /// An LRU thread pool is used since the jobs can be bumped and hence prioritized.
+    LRUThreadPool<KeyType> _threadPool;
+    
 };
 
 } // namespace globebrowsing
