### 2017-07-10 OpenSpace/OpenSpace#352 @alexanderbock
I updated the instructions for compilation ingeneral and Linux in particular.

The compilation errors look to me as if a pre C++17 standard library is used.   `std::enable_if_t`, `std::remove_cv_t`, ... are all defined in C++14.

On our Jenkins system, I have performed the same toolchain install, but then made `g++-7` the default through: `sudo rm /usr/bin/g++ && ln -s /usr/bin/g++-7 /usr/bin/g++` and `sudo rm /usr/bin/gcc && ln -s /usr/bin/gcc-7`.  That worked for me on three Ubuntu systems.  Not sure how Linux handles the different standard libraries though, but I think that is the most likely culprit
