### 2017-07-10 18d8612
Initial implementation of multiple ATM's (still missing frustum culling and other small details).
