+++ src/rendering/framebufferrenderer.cpp
@@ -127,6 +127,7 @@ void FramebufferRenderer::initialize() {
     updateRendererData();
     updateRaycastData();
     updateDeferredcastData();
+    updateHDRData();
 
     glBindFramebuffer(GL_FRAMEBUFFER, _mainFramebuffer);
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D_MULTISAMPLE, _mainColorTexture, 0);
@@ -244,10 +245,19 @@ void FramebufferRenderer::update() {
 
     // If the resolve dictionary changed (or a file changed on disk)
     // then rebuild the resolve program.
+    if (_hdrBackGroundProgram && _hdrBackGroundProgram->isDirty()) {
+        try {
+            _hdrBackGroundProgram->rebuildFromFile();
+        } catch (const ghoul::RuntimeError& error) {
+            LERRORC(error.component, error.message);
+        }
+    }
+
     if (_resolveProgram->isDirty()) {
         try {
             _resolveProgram->rebuildFromFile();
-        } catch (const ghoul::RuntimeError& error) {
+        }
+        catch (const ghoul::RuntimeError& error) {
             LERRORC(error.component, error.message);
         }
     }
@@ -333,8 +343,8 @@ void FramebufferRenderer::updateResolution() {
         GL_FLOAT,
         nullptr);
 
-    //glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, 0);
-    //glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, 0);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);    
 #endif
 
     glBindTexture(GL_TEXTURE_2D_MULTISAMPLE, _mainOtherDataTexture);
@@ -533,6 +543,19 @@ void FramebufferRenderer::updateDeferredcastData() {
     _dirtyDeferredcastData = false;
 }
 
+void FramebufferRenderer::updateHDRData() {
+    try {
+        _hdrBackGroundProgram = ghoul::opengl::ProgramObject::Build(
+            "HDR Background Control",
+            "${SHADERS}/framebuffer/hdrBackground.vert",
+            "${SHADERS}/framebuffer/hdrBackground.frag"
+        );
+    }
+    catch (const ghoul::RuntimeError& e) {
+        LERRORC(e.component, e.message);
+    }
+}
+
 void FramebufferRenderer::render(float blackoutFactor, bool doPerformanceMeasurements) {
     std::unique_ptr<performance::PerformanceMeasurement> perf;
     if (doPerformanceMeasurements) {
@@ -560,7 +583,7 @@ void FramebufferRenderer::render(float blackoutFactor, bool doPerformanceMeasure
     glGetIntegerv(GL_FRAMEBUFFER_BINDING, &defaultFbo);
 
     glBindFramebuffer(GL_FRAMEBUFFER, _mainFramebuffer);
-    // DEBUG: deferred g-buffer
+    // deferred g-buffer
     GLenum textureBuffers[4] = { GL_COLOR_ATTACHMENT0, GL_COLOR_ATTACHMENT1, GL_COLOR_ATTACHMENT2, GL_COLOR_ATTACHMENT3 };
     glDrawBuffers(4, textureBuffers);
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
[…]
