+++ include/openspace/rendering/abufferrenderer.h
@@ -69,7 +69,8 @@ public:
     void setCamera(Camera* camera) override;
     void setScene(Scene* scene) override;
     void setResolution(glm::ivec2 res) override;
-    void setNAaSamples(int nAaSamples) override;
+    void setNAaSamples(const int nAaSamples) override;
+    void setHDRExposure(const float hdrExposure) override;
 
     void preRaycast(const RaycasterTask& raycasterTask);
     void postRaycast(const RaycasterTask& raycasterTask);
@@ -130,6 +131,7 @@ private:
     GLuint _vertexPositionBuffer;
     int _nAaSamples;
 
+    float _hdrExposure;
     float _blackoutFactor;
 
     ghoul::Dictionary _rendererData;
