### 2017-07-10 e238719
[WebGui] Begin work on handling lost connection
