### 2017-07-10 8a3c904
Reverse config changes for easier iteration on this branch
