+++ openspace.cfg
@@ -27,7 +27,8 @@ return {
 
     -- Sets the scene that is to be loaded by OpenSpace. A scene file is a description
     -- of all entities that will be visible during an instance of OpenSpace
-    Scene = "${SCENE}/default.scene",
+    Scene = "${SCENE}/enlilnh.scene",
+    -- Scene = "${SCENE}/default.scene",
     -- Scene = "${SCENE}/globebrowsing.scene",
     -- Scene = "${SCENE}/rosetta.scene",
     -- Scene = "${SCENE}/dawn.scene",
@@ -74,7 +75,7 @@ return {
     Documentation = "${DOCUMENTATION}/Documentation.html",
     FactoryDocumentation = "${DOCUMENTATION}/FactoryDocumentation.html",
 
-    ShutdownCountdown = 3,
+    ShutdownCountdown = 0.5,
     -- OnScreenTextScaling = "framebuffer",
     -- PerSceneCache = true,
     -- DisableRenderingOnMaster = true,
@@ -90,6 +91,6 @@ return {
        },
 --      FilterSeverity = { }
 
-    }
-    --RenderingMethod = "ABuffer" -- alternative: "Framebuffer"
+    },
+    RenderingMethod = "ABuffer" -- alternative: "Framebuffer"
 }
