### 2017-07-10 c38e8c5
Rename WebGui -> WebBrowser
