+++ modules/webgui/include/web_render_handler.h
@@ -1,56 +1,0 @@
-/*****************************************************************************************
- *                                                                                       *
- * OpenSpace                                                                             *
- *                                                                                       *
- * Copyright (c) 2014-2017                                                               *
- *                                                                                       *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
- * software and associated documentation files (the "Software"), to deal in the Software *
- * without restriction, including without limitation the rights to use, copy, modify,    *
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
- * permit persons to whom the Software is furnished to do so, subject to the following   *
- * conditions:                                                                           *
- *                                                                                       *
- * The above copyright notice and this permission notice shall be included in all copies *
- * or substantial portions of the Software.                                              *
- *                                                                                       *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
- * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
- ****************************************************************************************/
-
-#ifndef OPENSPACE_MODULES_WEBGUI__WEB_RENDER_HANDLER_H
-#define OPENSPACE_MODULES_WEBGUI__WEB_RENDER_HANDLER_H
-
-#include <ghoul/opengl/opengl>
-#include <ghoul/logging/logmanager.h>
-#include <fmt/format.h>
-#include <include/cef_render_handler.h>
-#include <include/cef_app.h>
-
-namespace openspace {
-
-class WebRenderHandler : public CefRenderHandler {
-public:
-    virtual void draw(void) = 0;
-    virtual void render() = 0;
-
-    void reshape(int, int);
-
-    bool GetViewRect(CefRefPtr<CefBrowser> browser, CefRect &rect) override;
-    void OnPaint(CefRefPtr<CefBrowser> browser, PaintElementType type, const RectList &dirtyRects, const void *buffer,
-                 int width, int height) override;
-
-protected:
-    int width = 0, height = 0;
-    GLuint texture;
-
-    IMPLEMENT_REFCOUNTING(WebRenderHandler);
-};
-
-} // namespace openspace
-
-#endif //OPENSPACE_MODULES_WEBGUI__WEB_RENDER_HANDLER_H
