### 2017-07-10 37055c6
Update README.md
Add Slack channel
