### 2017-07-10 5a2e217
limit finger input to LMA to 3 (makes it easier to converge on 3+ cases)
