### 2017-07-10 672d024
Get subprocess path and gui url from openspace.cfg
