+++ openspace.cfg
@@ -91,6 +91,10 @@ return {
        },
 --      FilterSeverity = { }
 
-    }
-    --RenderingMethod = "ABuffer" -- alternative: "Framebuffer"
+    },
+    --RenderingMethod = "ABuffer" -- alternative: "Framebuffer",
+
+    WebHelperLocation = "./openspace_web_helper-figure_this_out",
+    -- WebGuiUrl = "file://${BASE_PATH}/gui/index.html"
+    WebGuiUrl = "http://localhost:8080"
 }
