### 2017-07-11 99a450f
std on all algorithms
