### 2017-07-11 56be681
Perform Oren Nayar diffuse shading on globes.
