### 2017-07-11 4309419
Remove unused propertyowner
