### 2017-07-11 f091444
Make it possible to toggle asset dependencies in the gui
