+++ src/scene/assetloader.cpp
@@ -48,7 +48,7 @@ namespace {
     const char* OnDeinitializeFunctionName = "onDeinitialize";
 
     const char* AssetsTableName = "_assets";
-    const char* TogglesTableName = "_toggles";
+    const char* DependantsTableName = "_dependants";
 
     const char* _loggerCat = "AssetLoader";
 }
@@ -59,12 +59,28 @@ namespace assetloader {
 
 int importAsset(lua_State* state) {
     AssetLoader *assetLoader = (AssetLoader*)lua_touserdata(state, lua_upvalueindex(1));
-    return assetLoader->importAssetLua();
+
+    int nArguments = lua_gettop(state);
+    if (nArguments != 1) {
+        return luaL_error(state, "Expected 1 argument, got %i", nArguments);
+    }
+
+    std::string assetName = luaL_checkstring(state, 1);
+    return assetLoader->importAssetLua(assetName, false, false);
 }
 
 int importAssetToggle(lua_State* state) {
     AssetLoader *assetLoader = (AssetLoader*)lua_touserdata(state, lua_upvalueindex(1));
-    return assetLoader->importAssetToggleLua();
+
+    int nArguments = lua_gettop(state);
+    if (nArguments != 2) {
+        return luaL_error(state, "Expected 2 arguments, got %i", nArguments);
+    }
+        
+    bool toggleEnabled = lua_toboolean(state, 2);
+
+    std::string assetName = luaL_checkstring(state, 1);
+    return assetLoader->importAssetLua(assetName, true, toggleEnabled);
 }
 
 int resolveLocalResource(lua_State* state) {
@@ -97,7 +113,7 @@ AssetLoader::AssetLoader(ghoul::lua::LuaState* luaState, std::string assetRoot, 
 
 AssetLoader::Asset* AssetLoader::importAsset(
     const std::string& name,
-    bool toggle,
+    bool togglableInitializationRequirement,
     bool toggleOn)
 {
     ghoul::filesystem::Directory directory = currentDirectory();
@@ -113,18 +129,13 @@ AssetLoader::Asset* AssetLoader::importAsset(
         asset = it->second.get();
     }
 
-    if (toggle) {
-        // This is a togglable dependency
-        Asset* toggler = _assetStack.back();
-        toggler->addToggle(asset);
-        if (toggleOn) {
-            toggler->setToggle(asset, true);
-        }
-    } else {
-        // This is a hard dependency
-        Asset* dependant = _assetStack.back();
-        dependant->addDependency(asset);
+    InitializationRequirement reqInit = InitializationRequirement::Yes;
+    if (togglableInitializationRequirement && !toggleOn) {
+        reqInit = InitializationRequirement::No;
     }
+
+    Asset* dependant = _assetStack.back();
+    dependant->addDependency(asset, togglableInitializationRequirement, reqInit);
 
     if (loaded) {
         return asset;
@@ -144,6 +155,7 @@ AssetLoader::Asset* AssetLoader::importAsset(
     LDEBUG("Imported asset " << asset->id());
     
     _importedAssets.emplace(id, std::move(newAsset));
+
     return _importedAssets[id].get();
 }
 
@@ -224,10 +236,10 @@ void AssetLoader::pushAsset(Asset* asset) {
     lua_pushcfunction(*_luaState, &assetloader::noOperation);
     lua_setfield(*_luaState, assetTableIndex, OnDeinitializeFunctionName);
 
-    // Register empty toggle table on imported asset.
-    // (importer => toggle object)
+    // Register empty dependant table on imported asset.
+    // (importer => dependant object)
     lua_newtable(*_luaState);
-    lua_setfield(*_luaState, assetTableIndex, TogglesTableName);
[…]
