### 2017-07-11 39b9086
Fixed HDR controls for multiple atm's.
