+++ include/openspace/rendering/abufferrenderer.h
@@ -71,6 +71,10 @@ public:
     void setResolution(glm::ivec2 res) override;
     void setNAaSamples(const int nAaSamples) override;
     void setHDRExposure(const float hdrExposure) override;
+    void setHDRBackground(const float hdrBackground) override;
+    void setGamma(const float gamma) override;
+
+    float hdrBackground() const override;
 
     void preRaycast(const RaycasterTask& raycasterTask);
     void postRaycast(const RaycasterTask& raycasterTask);
@@ -132,6 +136,8 @@ private:
     int _nAaSamples;
 
     float _hdrExposure;
+    float _hdrBackground;
+    float _gamma;
     float _blackoutFactor;
 
     ghoul::Dictionary _rendererData;
