### 2017-07-11 0b760c2
Move WebGui related things to WebGuiModule
