+++ modules/webgui/include.cmake
@@ -1,0 +1,6 @@
+set(DEFAULT_MODULE OFF)
+
+set(OPEMSPACE_DEPENCENCIES
+    webbrowser
+    server
+    )

[…]
