+++ modules/webbrowser/src/screen_space_browser.cpp
@@ -78,7 +78,7 @@ bool ScreenSpaceBrowser::initialize() {
     createPlane();
     // Load a special version of the regular ScreenRenderable shaders. This mirrors the
     // image along the Y axis since the image produced by CEF was flipped.
-    createShaders("${MODULE_WEB}/shaders/");
+    createShaders("${MODULE_WEBBROWSER}/shaders/");
     browserInstance->load(url);
 
     return isReady();
