### 2017-07-11 ae5b111
Adding back logging to config
