### 2017-07-11 59e0669
Make sure web EventHandler has browser before firing events
Also do some cleanup all over in WebBrowser.
