### 2017-07-11 5f50fc4
[WebBrowser] Merge CefHost and WebBrowserModule
With the new browser/gui structure, CefHost was redundant and could just
as well be merged into the module.
