### 2017-07-11 ff48f24
Adding satellites with ATM.
