+++ modules/base/rendering/renderableplane.cpp
@@ -202,6 +202,9 @@ void RenderablePlane::render(const RenderData& data) {
 
     _shader->setUniform("modelViewProjectionTransform",
         data.camera.projectionMatrix() * glm::mat4(modelViewTransform));
+
+    _shader->setUniform("modelViewTransform",
+        glm::mat4(data.camera.combinedViewMatrix() * glm::dmat4(modelViewTransform)));
     
     ghoul::opengl::TextureUnit unit;
     unit.activate();
