+++ data/scene/satellites/tle/gps-ops.txt
@@ -1,93 +1,93 @@
 GPS BIIR-2  (PRN 13)    
-1 24876U 97035A   17157.59786847  .00000019  00000-0  00000-0 0  9996
-2 24876  55.5513 227.5274 0034589 106.0906 254.3641  2.00561883145810
+1 24876U 97035A   17192.00017457  .00000062  00000-0  00000-0 0  9993
+2 24876  55.5457 226.1733 0033457 105.0535 255.3932  2.00561546146509
 GPS BIIR-3  (PRN 11)    
-1 25933U 99055A   17157.81823278 -.00000081  00000-0  00000-0 0  9990
-2 25933  51.6254  78.0928 0166095  94.4525 354.0300  2.00558378129440
+1 25933U 99055A   17192.30253387 -.00000043  00000-0  00000-0 0  9999
+2 25933  51.6477  76.5822 0167280  95.2661  52.2837  2.00557811130131
 GPS BIIR-4  (PRN 20)    
-1 26360U 00025A   17157.82515987  .00000060  00000-0  00000-0 0  9996
-2 26360  53.1170 155.3222 0040993  91.9775 104.3666  2.00561532125131
+1 26360U 00025A   17192.10302545  .00000028  00000-0  00000-0 0  9994
+2 26360  53.1229 153.8937 0041333  93.0207  13.5306  2.00559340125826
 GPS BIIR-5  (PRN 28)    
-1 26407U 00040A   17157.93926722 -.00000076  00000-0  00000-0 0  9998
-2 26407  56.6372 345.5470 0199158 270.1220 289.7792  2.00567008123812
+1 26407U 00040A   17192.08018765 -.00000064  00000-0  00000-0 0  9999
+2 26407  56.6269 344.2179 0199321 270.1095 101.4752  2.00566674124502
 GPS BIIR-6  (PRN 14)    
-1 26605U 00071A   17157.87543271  .00000026  00000-0  00000-0 0  9993
-2 26605  55.1227 225.2520 0093903 248.3083 174.0281  2.00565489121414
+1 26605U 00071A   17192.23740583  .00000062  00000-0  00000-0 0  9997
+2 26605  55.1175 223.8862 0095078 248.5040 145.1152  2.00564000122100
 GPS BIIR-7  (PRN 18)    
-1 26690U 01004A   17157.89507387  .00000063  00000-0  00000-0 0  9995
-2 26690  53.0389 158.2602 0183394 256.9116 288.9071  2.00570451119814
+1 26690U 01004A   17192.27125375  .00000033  00000-0  00000-0 0  9996
+2 26690  53.0447 156.8246 0183776 257.3765 270.8420  2.00569677120500
 GPS BIIR-8  (PRN 16)    
-1 27663U 03005A   17157.94951778 -.00000076  00000-0  00000-0 0  9997
-2 27663  56.7061 345.2850 0092504  24.7076 300.8478  2.00560319105174
+1 27663U 03005A   17192.24901613 -.00000063  00000-0  00000-0 0  9998
+2 27663  56.6954 343.9516 0092168  25.1635 225.6285  2.00558114105866
 GPS BIIR-9  (PRN 21)    
-1 27704U 03010A   17157.88465515 -.00000050  00000-0  00000-0 0  9993
-2 27704  53.8528  98.6489 0242249 264.5394 334.3334  2.00563816103977
+1 27704U 03010A   17191.86308833 -.00000027  00000-0  00000-0 0  9992
+2 27704  53.8708  97.2387 0241873 265.0719  28.2704  2.00566260104650
 GPS BIIR-10 (PRN 22)    
[…]
