+++ data/scene/satellites/tle/geo.txt
@@ -1,1332 +1,1332 @@
 TDRS 3                  
-1 19548U 88091B   17157.86807803 -.00000309  00000-0  00000-0 0  9997
-2 19548  14.4416   8.0727 0035139 315.3143 182.2791  1.00277792 92347
+1 19548U 88091B   17191.97410105 -.00000301  00000-0  00000-0 0  9994
+2 19548  14.4434   7.7197 0033314 313.6559 255.8357  1.00264680 92649
 SKYNET 4C               
-1 20776U 90079A   17157.24407692  .00000117  00000-0  00000-0 0  9991
-2 20776  13.8245  16.2209 0003137  48.2774 311.7180  1.00276680 97853
+1 20776U 90079A   17190.15345603  .00000123  00000-0  00000-0 0  9998
+2 20776  13.8368  15.8879 0002895  93.7434 266.3230  1.00273320 98181
 TDRS 5                  
-1 21639U 91054B   17157.76236800  .00000078  00000-0  00000-0 0  9998
-2 21639  14.3705  21.5279 0024207 350.4285 350.3256  1.00280423 94648
+1 21639U 91054B   17192.23704788  .00000083  00000-0  00000-0 0  9990
+2 21639  14.3904  21.1592 0023281 352.0698 193.7294  1.00272910 94991
 TDRS 6                  
-1 22314U 93003B   17157.86825340 -.00000301  00000-0  00000-0 0  9999
-2 22314  13.8980  24.4494 0003763 305.2309 192.4426  1.00268641 89385
+1 22314U 93003B   17192.11251377 -.00000296  00000-0  00000-0 0  9992
+2 22314  13.9216  24.1020 0001631 289.9184 329.6334  1.00265529 89655
 ASTRA 1D                
-1 23331U 94070A   17157.42011801 -.00000298  00000-0  00000-0 0  9990
-2 23331   7.5473  49.3635 0000659 296.5264   4.6759  1.00104533 83216
+1 23331U 94070A   17192.15320394 -.00000270  00000-0  00000-0 0  9999
+2 23331   7.5978  49.1136 0003274  55.0235 167.7487  1.00271751 83566
 BRASILSAT B2            
-1 23536U 95016A   17157.69941272 -.00000285  00000-0  00000-0 0  9997
-2 23536   7.1712  51.2584 0002332  20.9281   6.8855  1.00271709 81251
+1 23536U 95016A   17191.52514480 -.00000282  00000-0  00000-0 0  9991
+2 23536   7.2181  51.0048 0002128  65.3635 293.2964  1.00270899 81661
 AMSC 1                  
-1 23553U 95019A   17156.69144898 -.00000124  00000-0  00000-0 0  9999
-2 23553  10.0698  39.9829 0003984  11.9967 347.9227  1.00271961 81090
+1 23553U 95019A   17191.59542779 -.00000118  00000-0  00000-0 0  9990
+2 23553  10.1113  39.6742 0002836  60.4528 299.6126  1.00272553 81450
 TDRS 7                  
-1 23613U 95035B   17157.60700767 -.00000209  00000-0  00000-0 0  9993
-2 23613  14.9710  15.2291 0030718   2.0194 181.2233  1.00269402 80185
+1 23613U 95035B   17192.06674826 -.00000203  00000-0  00000-0 0  9992
+2 23613  14.9826  14.8618 0029077   4.8190  18.3183  1.00273130 80438
 ECHOSTAR 1              
-1 23754U 95073A   17157.49163003 -.00000246  00000-0  00000-0 0  9995
-2 23754   0.7924  90.2435 0002974 350.0571 274.5284  1.00272032 78375
[…]

[…]
