### 2017-07-11 362cb63
[WebBrowser] Find helper exec dynamically
This uses Ghoul's Directory::readFiles to find a file matching a filename
provided by the config (might not be necessary).
