+++ openspace.cfg
@@ -94,7 +94,7 @@ return {
     },
     --RenderingMethod = "ABuffer" -- alternative: "Framebuffer",
 
-    WebHelperLocation = "./openspace_web_helper-figure_this_out",
+    WebHelperLocation = "openspace_web_helper",
     -- WebGuiUrl = "file://${BASE_PATH}/gui/index.html"
     WebGuiUrl = "http://localhost:8080"
 }
