### 2017-07-12 9ec90eb
Creating asset modules for each type of tree
