### 2017-07-12 09d562f
Enable accurate normals for globes
