+++ modules/globebrowsing/rendering/layershadermanager.cpp
@@ -91,17 +91,22 @@ LayerShaderManager::LayerShaderPreprocessingData
         preprocessingData.layeredTextureInfo[i] = layeredTextureInfo;
     }
         
-    const auto& generalProps = globe.generalProperties();
-    const auto& debugProps = globe.debugProperties();
+    const RenderableGlobe::GeneralProperties& generalProps = globe.generalProperties();
+    const RenderableGlobe::DebugProperties& debugProps = globe.debugProperties();
     auto& pairs = preprocessingData.keyValuePairs;
         
+    pairs.emplace_back("useAccurateNormals",
+        std::to_string(generalProps.useAccurateNormals)
+    );
     pairs.emplace_back("useAtmosphere", std::to_string(generalProps.atmosphereEnabled));
     pairs.emplace_back("performShading", std::to_string(generalProps.performShading));
     pairs.emplace_back("showChunkEdges", std::to_string(debugProps.showChunkEdges));
     pairs.emplace_back("showHeightResolution",
-        std::to_string(debugProps.showHeightResolution));
+        std::to_string(debugProps.showHeightResolution)
+    );
     pairs.emplace_back("showHeightIntensities",
-        std::to_string(debugProps.showHeightIntensities));
+        std::to_string(debugProps.showHeightIntensities)
+    );
     pairs.emplace_back("defaultHeight", std::to_string(Chunk::DEFAULT_HEIGHT));
 
     pairs.emplace_back("tilePaddingStart",
