### 2017-07-12 5022eca
Use property for accurate normals.
