+++ modules/globebrowsing/CMakeLists.txt
@@ -218,7 +218,9 @@ set(SHADER_FILES
     ${CMAKE_CURRENT_SOURCE_DIR}/shaders/texturetilemapping.hglsl
     ${CMAKE_CURRENT_SOURCE_DIR}/shaders/tile.hglsl
     ${CMAKE_CURRENT_SOURCE_DIR}/shaders/tilefragcolor.hglsl
-    ${CMAKE_CURRENT_SOURCE_DIR}/shaders/tilevertexheight.hglsl
+    ${CMAKE_CURRENT_SOURCE_DIR}/shaders/tileheight.hglsl
+    ${CMAKE_CURRENT_SOURCE_DIR}/shaders/tilevertexskirt.hglsl
+    ${CMAKE_CURRENT_SOURCE_DIR}/shaders/globeshading.hglsl
 )
 
 source_group("Shader Files" FILES ${SHADER_FILES})
