### 2017-07-12 30e1b2b
make touchmarkers render on render callback over postdraw
