### 2017-07-12 fb0ead7
Initialize new module ResearchKit
