### 2017-07-12 472ab03
Add Module::Name constant for Server, Webgui and WebBrowser
