### 2017-07-12 8f5d5c0
[WebBrowser] Keep track of browsers in module
