### 2017-07-12 bc7a529
[WebBrowser] Use unique ptr for ScreenSpaceBrowser's instance
