### 2017-07-12 db9db49
Provide a static member variable Name for all OpenSpaceModules
