+++ openspace.cfg
@@ -10,8 +10,11 @@ return {
     -- SGCTConfig = sgct.config.single{},
 
     -- A regular 1920x1080 window
-    SGCTConfig = sgct.config.single{1920, 1080, border=false, windowPos={0,0}},   
-
+    SGCTConfig = sgct.config.single{1920, 1080, border=false, windowPos={0,0}, shared=true, name="WV_OBS_SPOUT1"},   
+	
+	-- Touch Powerwall
+	-- SGCTConfig = "${CONFIG}/sci_powerwall.xml",
+	
     -- A 1k fisheye rendering
     -- SGCTConfig = sgct.config.fisheye{1024, 1024},
 
