### 2017-07-12 9776205
Create Timer class and application timer
