### 2017-07-12 bbc8edf
[WebBrowser] Fix some compiler warnings
