### 2017-07-12 51d722c
First rough implementation. Need to change configurations to properties
