### 2017-07-13 ec0a90d
Removed ConfigurationManager changes and replaced with Properties. Not perfect, but can acess from "Global Properties.ResearchKit.Timer._timeout"
