### 2017-07-13 ab7aafa
Create script file and scene to run shutdown timer
