+++ openspace.cfg
@@ -27,7 +27,8 @@ return {
 
     -- Sets the scene that is to be loaded by OpenSpace. A scene file is a description
     -- of all entities that will be visible during an instance of OpenSpace
-    Scene = "${SCENE}/default.scene",
+    -- Scene = "${SCENE}/default.scene",
+    Scene = "${SCENE}/shutdown.scene",
     -- Scene = "${SCENE}/globebrowsing.scene",
     -- Scene = "${SCENE}/rosetta.scene",
     -- Scene = "${SCENE}/dawn.scene",
