### 2017-07-13 895f0ac
Move logs into ${LOGS} directory
