+++ openspace.cfg
@@ -62,7 +62,7 @@ return {
         LogLevel = "Debug",
         ImmediateFlush = true,
         Logs = {
-            { Type = "html", File = "${BASE_PATH}/log.html", Append = false }
+            { Type = "html", File = "${LOGS}/log.html", Append = false }
         },
         CapabilitiesVerbosity = "Full"
     },
@@ -71,7 +71,7 @@ return {
     },
     LuaDocumentation = "${DOCUMENTATION}/LuaScripting.html",
     PropertyDocumentation = "${DOCUMENTATION}/Properties.html",
-    ScriptLog = "${BASE_PATH}/ScriptLog.txt",
+    ScriptLog = "${LOGS}/ScriptLog.txt",
     KeyboardShortcuts = "${DOCUMENTATION}/KeyboardMapping.html",
     Documentation = "${DOCUMENTATION}/Documentation.html",
     FactoryDocumentation = "${DOCUMENTATION}/FactoryDocumentation.html",
