### 2017-07-13 5e47623
Correctly calculate height when scale is changed.
