### 2017-07-14 OpenSpace/OpenSpace#352 @MatthewTurk
Thanks!  I was able to get it to work using `update-alternatives` to point to the other GCC.
