### 2017-07-15 #281 @arfon
Hi @alexanderbock, before I can accept this submission could you please reformat your `paper.md` file. A couple of issues:

1. The `affiliations` need pulling out into a separate block in the YAML header (see [this example](https://raw.githubusercontent.com/arfon/fidgit/master/paper/paper.md))
2. None of the references in your bibtex are mentioned directly in the paper so by default they won't be included by pandoc when compiling the paper. Could you please make sure you're actually citing them (see how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
