### 2017-07-25 #281 @alexanderbock
Sorry for the delay;  there was another deadline ongoing.

@arfon   I changed the `paper.md` accordingly.  I thought that I had seen in one of the instructions that inline affiliations were okay as well.

I have two affiliation tags next to my name in there, is specifying the attribute multiple times allowed or is there a different syntax for this?
