### 2017-07-26 #281 @arfon
@alexanderbock - you nearly had it right. I fixed things up in https://github.com/OpenSpace/OpenSpace/pull/373

Once you've merged https://github.com/OpenSpace/OpenSpace/pull/373 could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
