### 2017-07-26 #281 @alexanderbock
@arfon  The pullrequest is merged, thanks for the fix.  If have uploaded the software at Zenodo: http://doi.org/10.5281/zenodo.835206   (https://zenodo.org/record/835206)
