### 2017-07-27 #281 @arfon
@MatthewTurk many thanks for your review here and @danielskatz for editing this submission ✨

@alexanderbock - your submission is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00281 :zap: 🚀 💥
